% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RouteDataFunction.R
\name{add_stay_id}
\alias{add_stay_id}
\title{Add Stay IDs}
\usage{
add_stay_id(df)
}
\arguments{
\item{df}{A data frame with spatial indices.}
}
\value{
A data frame with \code{stay_id} and \code{stay_len} columns added.
}
\description{
Adds stay IDs to a data frame based on changes in spatial indices.
}
\examples{
df_with_stay_ids <- add_stay_id(routes)
}
