% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_movement_vectors.R
\name{animate_movement_vectors}
\alias{animate_movement_vectors}
\title{Animate movement vectors}
\usage{
animate_movement_vectors(bf, ...)
}
\arguments{
\item{bf}{A BirdFlow object}

\item{...}{
  Arguments passed on to \code{\link[=lookup_timestep_sequence]{lookup_timestep_sequence}}
  \describe{
    \item{\code{season}}{a season name, season alias, or "all".  See
\code{\link[=lookup_season_timesteps]{lookup_season_timesteps()}} for options.}
    \item{\code{start}}{The starting point in time specified as a
timestep, character date, or date object.}
    \item{\code{end}}{The ending point in time as a date or timestep.}
    \item{\code{direction}}{Either "forward" or "backward" defaults to
\code{"forward"} if not processing dates.  If using date input \code{direction} is
optional and is only used to verify the direction implicit in the dates.}
    \item{\code{season_buffer}}{Only used with \code{season} input. \code{season_buffer} is
passed to \code{\link[=lookup_season_timesteps]{lookup_season_timesteps()}} and defaults to 1; it is the number
of timesteps to extend the season by at each end.}
    \item{\code{n_steps}}{Alternative to \code{end}  The end will be \code{n_steps} away from
\code{start} in \code{direction}; and the resulting sequence will have \code{n_step}
transitions and \code{n_steps + 1}  timesteps.}
  }}
}
\value{
A \code{gganim} object. \code{print()} will plot it with default
options, or use \code{\link[gganimate:animate]{gganimate::animate()}} to set the options. See the example for
recommended settings.
}
\description{
\code{animate_movement_vectors()} produces a \code{gganim} object in which each frame
is a map of vectors showing the average modeled movement for all birds from
each cell in the landscape at a given timestep. It is analogous to a series
of images created with \code{\link[=plot_movement_vectors]{plot_movement_vectors()}}.
}
\details{
Each arrow represents the average of all the transitions from a single cell.
The tail of the arrow is the center of that cell, the head is the average
location at the following timestep for birds that start at that cell.

The timestep and/or date label is the starting timestep for the transition
that is displayed and the format depends on
\link[=birdflow_options]{birdflow_options("time_format")}

Thicker lines and less transparency (darker shading) indicate higher density
in the eBird S&T distribution for the beginning timestep of the displayed
transition.

Use the "ragg_png" device when rendering animations as in the
example code.
}
\examples{

\donttest{
bf <- BirdFlowModels::amewoo
a <- animate_movement_vectors(bf)
}

\dontrun{

# Animate, display, and save
#   Note: "ragg_png" is considerably faster and produces cleaner output than
#         the default device.
gif <- gganimate::animate(a, fps = 1, device = "ragg_png",
                          width = 6, height = 5,
                          res = 150, units = "in")
print(gif)

# Save
gif_file <- tempfile("animation", fileext = ".gif")
gganimate::save_animation(gif, gif_file)
file.remove(gif_file) # cleanup
}

}
