% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_BirdFlowIntervals.R
\name{as_BirdFlowIntervals}
\alias{as_BirdFlowIntervals}
\title{Deprecated function to convert BirdFlowRoutes to BirdFlowIntervals}
\usage{
as_BirdFlowIntervals(
  birdflow_routes,
  max_n = 1000,
  min_day_interval = 7,
  max_day_interval = 180,
  min_km_interval = 200,
  max_km_interval = 8000
)
}
\arguments{
\item{birdflow_routes}{A \code{BirdFlowRoutes} object.}

\item{max_n}{The maximum number of intervals to sample.
Defaults to 1000. Only used by the \code{BirdFlowRoutes} method.}

\item{min_day_interval}{The minimum days required in an interval.
Defaults to 7. Only used by the \code{BirdFlowRoutes} method.}

\item{max_day_interval}{The maximum days required in an interval.
Defaults to 180. Only used by the \code{BirdFlowRoutes} method.}

\item{min_km_interval}{The minimum distance required for an interval.
Defaults to 200. Only used by the \code{BirdFlowRoutes} method.}

\item{max_km_interval}{The maximum distance required for an interval.
Defaults to 2000. Only used by the \code{BirdFlowRoutes} method.}
}
\value{
A \code{BirdFlowIntervals} object.
}
\description{
This function is deprecated. Please use \code{\link[=as_birdflow_intervals]{as_birdflow_intervals()}} instead.
}
\seealso{
\code{\link[=as_birdflow_intervals]{as_birdflow_intervals()}} should be used instead of this function.
}
\keyword{internal}
