% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_BirdFlowRoutes.R
\name{as_BirdFlowRoutes}
\alias{as_BirdFlowRoutes}
\title{Deprecated function to convert Routes to BirdFlowRoutes}
\usage{
as_BirdFlowRoutes(
  routes,
  bf,
  aggregate = "random",
  valid_only = TRUE,
  sort_id_and_dates = TRUE,
  reset_index = FALSE
)
}
\arguments{
\item{routes}{A \code{Routes} object.}

\item{bf}{A \code{BirdFlow} object for spatial and temporal reference.}

\item{aggregate}{The aggregation method if more than one timestep is
presented in a route.
Options include \code{mean}, \code{median}, \code{midweek}, \code{random}. Default to \code{random}.
See \link{snap_to_birdflow} for a description of each aggregation method.}

\item{valid_only}{Logical. Should only valid points be included?
Defaults to \code{TRUE}.}

\item{sort_id_and_dates}{Logical. Should data be sorted by route ID and date?
Defaults to \code{TRUE}.}

\item{reset_index}{Logical. Should indices be reset after sorting?
Defaults to \code{FALSE}.}
}
\value{
A \code{BirdFlowRoutes} object.
}
\description{
This function is deprecated. Please use \code{\link[=as_birdflow_routes]{as_birdflow_routes()}} instead.
}
\seealso{
\code{\link[=as_birdflow_routes]{as_birdflow_routes()}} should be used instead of this function.
}
\keyword{internal}
