% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_distr.R
\name{as_distr}
\alias{as_distr}
\alias{as_distr.data.frame}
\alias{as_distr.SpatRaster}
\alias{as_distr.sf}
\title{Convert to a BirdFlow distribution}
\usage{
as_distr(x, bf, ...)

\method{as_distr}{data.frame}(x, bf, crs = NULL, ...)

\method{as_distr}{SpatRaster}(x, bf, normalize = TRUE, zero_na = TRUE, ...)

\method{as_distr}{sf}(x, bf, ...)
}
\arguments{
\item{x}{An object to be converted either a data.frame with x and y columns
indicating point locations, a \emph{sf} object containing points, or a raster
object containing values to be treated as a distribution.}

\item{bf}{A reference BirdFlow object.}

\item{...}{Arguments used by other methods.}

\item{crs}{The coordinate reference system that the x and y coordinates are
in.  If NULL the function will assume coordinates are in the same CRS as
\code{bf}.}

\item{normalize}{if \code{TRUE} normalize each distribution to sum to 1}

\item{zero_na}{if \code{TRUE} replace \code{NA} values with \code{0}.}
}
\value{
An object containing distribution data to be projected with x.
Either a vector with \href{n_active()}{n_active(bf)} values or a matrix with
that many rows and a column for each distribution.
}
\description{
Methods to convert raster or point data to distributions.  Each layer of a
raster will be treated as a distribution.  With point data each point will
be treated as a "one hot" distribution containing mostly 0 and a single
1 at the location corresponding to the point.
}
\details{
If \code{x} is a data.frame it should have columns \code{x} and \code{y} containing the
coordinates of points.   \code{crs} should indicate the coordinate reference
system the points are in - only necessary if it differs from \code{crs(bf)}.

If \code{x} is either an \emph{sf} object or a data frame it should represent points.
Each point will be converted into a separate distribution
with the cell corresponding to the point having a value of 1 with the
remaining values set to 0.

If \code{x} is a \code{\link[terra:SpatRaster-class]{terra::SpatRaster()}} it will be projected to align with the
cells in \code{bf} with \code{\link[terra:project]{terra::project()}} using \code{method = "average"} and then
cropped and/or extended to match the extent of \code{bf}.  Warnings will be throw
if some of the value in \code{x} is lost due to cropping or due to masking out the
inactive cells.
}
\examples{
\donttest{
bf <- BirdFlowModels::amewoo
xy <- i_to_xy(c(1, 2), bf)

# data.frame method: one "one hot" distribution per point
pts <- data.frame(x = xy$x, y = xy$y)
as_distr(pts, bf)

# sf method
sf_pts <- sf::st_as_sf(pts, coords = c("x", "y"), crs = crs(bf))
as_distr(sf_pts, bf)

# SpatRaster method
r <- rast(bf, which = 1)
as_distr(r, bf)
}
}
