% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_msg.R
\name{bf_msg}
\alias{bf_msg}
\title{Internal function to send a message from within BirdFlow functions}
\usage{
bf_msg(..., sep = "")
}
\arguments{
\item{...}{Text that will be pasted together to make a message.}

\item{sep}{(optional) separator between text elements in \code{...}
defaults to no separation.}
}
\value{
Nothing is returned if verbose is TRUE the message is printed.
}
\description{
It checks to see if BirdFlowR is in verbose mode
(\verb{birdflow_options("verbose"}) and if so pastes it's arguments together
and  prints the result with  \code{cat()}.
In the future it might be updated to use \code{message()}
}
\seealso{
\code{\link[=birdflow_options]{birdflow_options()}} for changing verbosity.
}
\keyword{internal}
