% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_msg.R
\name{bf_suppress_msg}
\alias{bf_suppress_msg}
\title{Conditionally suppress messages from expressions in BirdFlowR code}
\usage{
bf_suppress_msg(exp)
}
\arguments{
\item{exp}{R code that might throw a message (originating outside of
\pkg{BirdFlowR}.}
}
\description{
This internal functions is used to suppress messages thrown
by functions called in BirdFlowR code if \code{birdflow_options("verbose")}
is \code{FALSE}.
}
\examples{
\dontrun{
# bf_suppress_msg isn't exported so can't be run in examples
# in internal code or after  devtools::load_all() example will work
   ob <- birdflow_options("verbose")
   birdflow_options(verbose = FALSE)
   bf_suppress_msg( message("hi" ))
   birdflow_options(verbose = TRUE)
   bf_suppress_msg( message("hi" ))
   birdflow_options(ob)
 }

}
\seealso{
\code{\link[=preprocess_species]{preprocess_species()}} uses this when calling \pkg{ebirdst}
functions that display messages.
When BirdFlowR functions generate messages they should use \code{\link[=bf_msg]{bf_msg()}} so that
\code{birdflow_options("verbose")} is honored.
}
\keyword{internal}
