% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_BirdFlowIntervals.R
\name{BirdFlowIntervals}
\alias{BirdFlowIntervals}
\title{Deprecated function to make a BirdFlowIntervals object from a data frame}
\usage{
BirdFlowIntervals(data, species, metadata, geom, dates, source = NULL)
}
\arguments{
\item{species}{A list with species information. Only used by the
\code{data.frame} method.}

\item{metadata}{A list with additional metadata. Only used by the
\code{data.frame} method.}

\item{geom}{A list describing spatial geometry. Only used by the
\code{data.frame} method.}

\item{dates}{A data frame with date-related information. Only used by the
\code{data.frame} method.}

\item{source}{A character string indicating the source of the data. Only
used by the \code{data.frame} method.}
}
\value{
A \code{BirdFlowIntervals} object.
}
\description{
This function is deprecated. Please use \code{\link[=as_birdflow_intervals]{as_birdflow_intervals()}} instead
(its \code{data.frame} method).
}
\seealso{
\code{\link[=as_birdflow_intervals]{as_birdflow_intervals()}} should be used instead of this function.
}
\keyword{internal}
