% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_path.R
\name{cache_path}
\alias{cache_path}
\title{Internal function to get the local path to the cache for the current
model collection.}
\usage{
cache_path(collection_url)
}
\arguments{
\item{collection_url}{The url of a collection. Should be the path to
the base directory (not an index.html file).}
}
\value{
Path to the local cache for the current collection including a
trailing delimiter.
}
\description{
Use \href{birdflow_options()}{birdflow_options("cache")} to get the main cache
directory (parent of the collection directory) and
\href{birdflow_options()}{birdflow_options("collection_url")} for the current model
collection url.
}
\keyword{internal}
