% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_abundance_stats.R
\name{calc_abundance_stats}
\alias{calc_abundance_stats}
\title{Calculate stats from a spatRaster with abundances for each timestep}
\usage{
calc_abundance_stats(x, circular = TRUE)
}
\arguments{
\item{x}{a multilayer spatRaster with relative abundances for each timestep}

\item{circular}{if TRUE (the default) use the first timestep abundance as the
ending abundance.}
}
\value{
A list with
\item{n_params}{The number of parameters there would be in the model if fit
on \code{x}}
\item{count}{ a vector of the number of cells occupied at each timestep}
\item{area}{ a vector of area (sq m) that is in included cells for each
timestep} \item{res}{ the resolution of the raster in km}
}
\description{
Calculate the total number of parameters and the occupied area for each
timestep from an abundance raster.
}
\details{
This is a helper to \code{\link[=preprocess_species]{preprocess_species()}}
}
\keyword{internal}
