% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_year_number.R
\name{calc_year_number}
\alias{calc_year_number}
\title{Calculate year number from a sequence of timesteps}
\usage{
calc_year_number(x)
}
\arguments{
\item{x}{a sequence of integers representing timesteps.}
}
\value{
sequence of year numbers all either 1 or 2 depending on whether the
corresponding value in x is in the first or second year of x.
}
\description{
This internal function assigns a sequential year to each
observation based on whether the sequence has passed over a presumed year
boundary and started over. If the sequence is backwards and passes over the
boundary the result will start at 2 (later year) and end at 1 (earlier year).
}
\details{
It is used by plot_routes() when calculating half proportional years.
}
\keyword{internal}
