% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_year_offset.R
\name{calc_year_offset}
\alias{calc_year_offset}
\title{Calculate year offset from a sequence of timesteps}
\usage{
calc_year_offset(x)
}
\arguments{
\item{x}{a sequence of integers representing timesteps.}
}
\value{
sequence of year offsets these will be 0 or 1 for forward sequences
and 0 or -1 for backwards.
}
\description{
This internal function assigns a year offset for each
observation based on whether the sequence has passed over a presumed year
boundary and started over. The sequence always starts with 0.
Forward sequences increment with each new year, while backwards sequences
decrements (0, -1, etc.). Elsewhere there are limitations on \code{route()} that
prevent routes longer than 1 full year (back to start), but this function
does not have that limit.
}
\details{
It is used by \code{route()} to convert from circular to linear time by
making dates monotonic.
}
\keyword{internal}
