% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_population.R
\docType{data}
\name{callaghan_abundance}
\alias{callaghan_abundance}
\title{Abundance Estimates from Callaghan et. al. 2021}
\format{
A data frame with 9,700 rows and 11 columns:
\describe{
\item{\code{common_name}}{Common name}
\item{\code{scientific_name}}{Scientific name}
\item{\code{order}, \code{family}}{Taxonomic order and family}
\item{\code{lower_ci}}{The lower 95\% confidence interval}
\item{\code{abundance_estimate}}{Estimated population size}
\item{\code{upper_ci}}{The upper 95\% confidence interval}
\item{\code{range_adjusted}}{\code{TRUE} indicates that the area of the species
range from a source other than the eBird pixels was used to adjust the
abundance estimate.  If an independent range area was not available than
the area of the eBird cells was used.  See Callaghan et. al. 2021"}
\item{\code{training_species}}{\code{TRUE} if the species was used to train the
abundance model (Callaghan et. al. 2021)}
\item{\code{species_code}}{The eBird species code derived by joining
\code{scientific_name} to the \code{ebird_taxonomy} in the \strong{auk} package.
\code{NA} indicates that the scientific name from Callaghan et. al. did not
match any scientific names in the eBird taxonomy table.}
\item{\code{in_ebird_2022}}{Is the \code{species_code} in \code{ebirds::ebirdst_runs} from
the 2022 version of the eBird package. See: \code{\link[ebirdst:ebirdst_version]{ebirdst::ebirdst_version()}}
and \link[ebirdst:ebirdst_runs]{ebirdst::ebirdst_runs}. Starting with the 2022 version year only
a subset of the species were fit by eBird in each year.}
}
}
\source{
Callaghan, Corey T., Shinichi Nakagawa, and William K. Cornwell. "Global
abundance estimates for 9,700 bird species." Proceedings of the National
Academy of Sciences 118.21 (2021): e2023170118.
\url{https://doi.org/10.1073/pnas.2023170118}
}
\usage{
callaghan_abundance
}
\description{
The Callaghan et. al. (2021) population estimates and 95\% confidence
interval for 9700 species.
}
\keyword{datasets}
