% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_distr.R
\name{collapse_distr}
\alias{collapse_distr}
\title{Function to convert a raster bird distribution into its collapsed, vector
equivalent}
\usage{
collapse_distr(x, obj)
}
\arguments{
\item{x}{Either a matrix representing a single bird distribution or an array
representing multiple distributions.}

\item{obj}{A \code{BirdFlow} model or the \code{geom} component of one.}
}
\value{
Either a vector representing a single distribution in its collapsed
form or, if x represents multiple distributions,  a matrix with one
distribution per column.
}
\description{
This function is primarily for internal use. It converts between the raster
representation of data in which rows and columns indicate position in space
and a vector representation that contains all the non-masked cells in row
major order - starting at the top left and proceeding left to right along
each row.  The collapsed form is used for projecting the flow model while the
expanded form is used to export, import, and visualize the data.
}
\seealso{
\itemize{
\item \link{expand_distr} does the opposite of \code{collapse_distr}.
\item \link{index_conversions} converts among indexes of the data in expanded and
collapsed formats, and Cartesian space.
}
}
