% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_transitions.R
\name{combine_transitions}
\alias{combine_transitions}
\title{Function to combine a sequence of transition matrices into one}
\usage{
combine_transitions(bf, ...)
}
\arguments{
\item{bf}{A BirdFlow object}

\item{...}{
  Arguments passed on to \code{\link[=lookup_transitions]{lookup_transitions}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
This returns the transition probabilities associated with a
sequence of timesteps. It will have a column for every unmasked cell
at the starting timestep and a row for every unmasked cell in the
last timestep, with cell values being the probably of transitioning from
that row to that column between the start and end of the time sequence
described by \code{...}
}
\description{
Function to combine a sequence of transition matrices into one
}
\examples{
\donttest{
bf <- BirdFlowModels::amewoo
m <- combine_transitions(bf, start = 1, end = 3)
dim(m)
}
}
