% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_list_item_names.R
\name{compare_list_item_names}
\alias{compare_list_item_names}
\title{compare names of two nested lists}
\usage{
compare_list_item_names(x, y, map = "x", differences)
}
\arguments{
\item{x}{list}

\item{y}{reference list}

\item{map}{used during recursion to tracking where in x differences are found}

\item{differences}{used during recursion to collecting differences}
}
\value{
data frame with where and difference columns will have 0 rows if no
differences found.
}
\description{
Look for difference in the list item names at any nested level.
Used internally to verify BirdFlow object structure
}
\keyword{internal}
