% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst_version.R
\name{date_to_week}
\alias{date_to_week}
\title{lookup ebirst weeks from dates}
\usage{
date_to_week(dates, version = 2022)
}
\arguments{
\item{dates}{a vector of dates that can be processed by as.POSIXlt()}

\item{version}{What version of the ebirdst date scheme to use. 2021 for
the older, and 2022 for the newer version first used in
ebirdst v 3.2022.0  All other values will be reset to whichever version
is closer.}
}
\value{
A vector of weeks (1 to 52) corresponding with \code{dates}
}
\description{
This is a slightly modified copy of ebirdst::date_to_st_week() that
allows calculating weeks from dates without depending on ebirdst.
}
\keyword{internal}
