% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_resolution.R
\name{determine_resolution}
\alias{determine_resolution}
\title{Determine BirdFlow model resolution                                     ####}
\usage{
determine_resolution(
  sp_path,
  res,
  max_params,
  gpu_ram,
  clip,
  crs,
  download_species,
  project_method
)
}
\arguments{
\item{sp_path}{The species path used with \pkg{ebirdst} to download and load
data}

\item{res}{The target resolution of the BirdFlow model in kilometers. If
\code{res} is NULL (default) then a resolution that results in less than \code{max_params}
parameters will be used, while also minimizing the resolution and limiting
the number of significant digits.}

\item{max_params}{The maximum number of fitted parameters that the BirdFlow
model should contain. Ignored if \code{res} is not NULL.  Otherwise a resolution
will be chosen that yields this many fitted parameters. See \code{gpu_ram} for
the default way of setting \code{max_params} and \code{res}. Note: the reduction in
parameters resulting from truncation (see \code{...}) is not factored into the
calculation.}

\item{gpu_ram}{Gigabytes of ram on GPU machine that will fit the models.
If \code{res} is NULL and \code{max_params} is NULL this is used to estimate
\code{max_params}which is, in turn, used to determine the resolution. Ignored
if either\code{ res} or \code{max_params} is set.}

\item{clip}{A polygon or the path to a file containing a polygon. It must
have a CRS and should either be a \link[terra:SpatVector-class]{SpatVector()} object
or produce one when called with \link[terra:vect]{vect(clip)}}

\item{crs}{Coordinate reference system (CRS) to use.  Defaults to the custom
projection eBird has assigned to this species - see
\code{\link[ebirdst:load_fac_map_parameters]{ebirdst::load_fac_map_parameters()}}). It will be interpreted by
\code{\link[terra:crs]{terra::crs()}} to generate a well known text representation of the CRS.}

\item{download_species}{The species code used with \pkg{ebirdst} this might
be "example_data" or "yebsap-example" but otherwise will be a real
species code.}

\item{project_method}{This is the method used to reproject it is a local
variable set within \code{preprocess_species}.}
}
\value{
The resolution in km either as set directly by the user or as
derived from \code{max_params} or \code{gpu_ram}.
}
\description{
Internal function to determine the resolution to use when creating a
BirdFlow model.  It is called by \code{preprocess_species()} and nowhere else,
but is complicated enough to justify being it's own function.
}
\details{
When the user specifies a resolution that is the resolution used when
creating the model.

If the \code{res} argument is \code{NULL} the heuristic here attempts to set a
from the \code{gpu_ram} parameter which specifies the the GB of ram available on
the machine used to fit the models.

It turned out to be really hard to anticipate how many cells would
contain data after a resolution change. The code estimates by
calculating the area of the non-zero cells in the current resolution
and then figures out the resolution where the number of cells required
to cover that area matches our target number of parameters.

However, it's a poor estimate because it ignores the fact that coarse cells
along the edges overlap fine cells that contain a mix of no data and data.
reduce the bias.

Next it resamples to the estimated resolution, evaluates the new number of
non-zero cells (and thus parameters), and makes a new estimate, repeating
the process until the realized number of parameters from the resolution
converges on 90 to 100 \% of the target. This results in a fairly precise
maximum resolution that can be fit given the number of parameters.

The last step is to round up (reducing parameters) to a cleaner number.
}
\keyword{internal}
