% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R, R/n_parameters.R
\name{BirdFlow dimensions}
\alias{BirdFlow dimensions}
\alias{nrow.BirdFlow}
\alias{nrow}
\alias{ncol}
\alias{dim}
\alias{ext}
\alias{ncol.BirdFlow}
\alias{dim.BirdFlow}
\alias{n_timesteps}
\alias{n_distr}
\alias{n_transitions}
\alias{n_active}
\alias{crs,BirdFlow-method}
\alias{ext,BirdFlow-method}
\alias{res,BirdFlow-method}
\alias{xres,BirdFlow-method}
\alias{yres,BirdFlow-method}
\alias{xmin,BirdFlow-method}
\alias{ymin,BirdFlow-method}
\alias{xmax,BirdFlow-method}
\alias{ymax,BirdFlow-method}
\alias{is_cyclical}
\alias{n_parameters}
\title{Dimensions of a BirdFlow object}
\usage{
\method{nrow}{BirdFlow}(x)

\method{ncol}{BirdFlow}(x)

\method{dim}{BirdFlow}(x)

n_timesteps(x)

n_distr(x)

n_transitions(x)

n_active(x)

\S4method{crs}{BirdFlow}(x, proj = FALSE, describe = FALSE, parse = FALSE)

\S4method{ext}{BirdFlow}(x)

\S4method{res}{BirdFlow}(x)

\S4method{xres}{BirdFlow}(x)

\S4method{yres}{BirdFlow}(x)

\S4method{xmin}{BirdFlow}(x)

\S4method{ymin}{BirdFlow}(x)

\S4method{xmax}{BirdFlow}(x)

\S4method{ymax}{BirdFlow}(x)

is_cyclical(x)

n_parameters(x)
}
\arguments{
\item{x}{A BirdFlow object}

\item{proj}{logical. If \code{TRUE} the crs is returned in PROJ-string notation}

\item{describe}{logical. If \code{TRUE} the name, EPSG code, and the name and extent of the area of use are returned if known}

\item{parse}{logical. If \code{TRUE}, wkt parts are parsed into a vector (each line becomes an element)}
}
\value{
\code{nrow()} returns the number of rows in the raster extent of the
BirdFlow model

\code{ncol()} number of columns in the raster extent

\code{dim()} number of rows and columns in the raster extent.

\code{n_timesteps()} number of timesteps
distributions.

\code{n_distr()} number of distributions

\code{n_transitions()} number of transitions, if the model is circular in
time this will equal \code{n_timesteps()}.

\code{n_active()} number of active cells (locations).

\code{crs()} the coordinate reference system, a character with well known
text (WKT) by default, but see arguments.

\code{ext()} a \link[terra:SpatExtent-class]{SpatExtent} object,
which contains the xmin, xmax, ymin, and ymax of the extent.

\code{res()} two numbers, the cell width and height (x and y resolution).

\code{xres()} the width of the cells (x resolution).

\code{yres()} the height of the cells (y resolution).

\code{xmin()} minimum x coordinate of extent.

\code{ymin()} minimum y coordinate of extent.

\code{xmax()} maximum x coordinate of extent.

\code{ymax()} maximum y coordinate of extent.

\code{is_cyclical()} returns \code{TRUE} if the BirdFlow model has a transition
from the last timestep to the first and \code{FALSE} otherwise.

\code{n_parameters()} the number of of parameters that the BirdFlow model
contains or will contain. This is the number of cells in the marginal
matrices + the sum of the dynamic mask for the first timestep (number of
unmasked cells at the first timestep).  If the model isn't dynamically
masked this is equivalent to \code{n_active(x)^2 * n_transitons(x) + n_active(x)}
}
\description{
Functions to return BirdFlow model dimensions and other basic information
}
\examples{
\donttest{
bf <- BirdFlowModels::amewoo
nrow(bf)
ncol(bf)
dim(bf)
n_timesteps(bf)
n_distr(bf)
n_transitions(bf)
n_active(bf)
is_cyclical(bf)
crs(bf)
ext(bf)
res(bf)
xres(bf)
yres(bf)
xmin(bf)
xmax(bf)
ymin(bf)
ymax(bf)
}
}
\seealso{
\pkg{terra} defines the S4 generics for \link[terra:crs]{crs()},
\link[terra:ext]{ext()}, \link[terra:dimensions]{res()}, \link[terra:dimensions]{xres()}, and
\link[terra:dimensions]{yres()}

\code{\link[=get_distr]{get_distr()}} returns distributions from a \code{BirdFlow} object.
}
