% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmi_to_i.R
\name{dmi_to_i}
\alias{dmi_to_i}
\title{Convert between dynamic mask index and static location index}
\usage{
dmi_to_i(dmi, timestep, bf)
}
\arguments{
\item{dmi}{Dynamic mask index values.  These will always be integers between
\code{1} and the sum of the dynamic mask for the given timestep.}

\item{timestep}{Either a single timestep to be used for all conversions
or a vector of the timesteps associated with each input value: \code{dmi} for
\code{dmi_to_i()} or \code{i} for \code{i_to_dmi()}.}

\item{bf}{A BirdFlowR model. Note the conversion is specific to this
model and not valid for any others.}

\item{i}{Location index. This indicates a location based on an index of the
cells included by the static mask. These start in the top left location and
fill in by row.}
}
\value{
The location index,\code{i}, of each location indicated by \code{dmi} or
vis-versa.
}
\description{
\code{dmi_to_i()} and \code{i_to_dmi()} are for internal and advanced use;
they are not likely to be helpful to most users.
See \href{i_to_x}{index conversions} for, likely, more useful conversions.
These two functions convert between indices along the cells that are
included in the dynamic mask (\code{dmi}) and standard location
indices (\code{i}) along the cells that are included by the static mask.
This conversion requires knowing the \code{timestep}
associated with each \code{dmi} or \code{i} value
as the mapping between the two is different for each timestep.
}
\examples{
 bf <- BirdFlowModels::amewoo
 dmi <- c(11:20)
 timesteps <- c(1, 1, 1, 1, 3, 3, 5,7:9 )

 i <- dmi_to_i(dmi, timesteps, bf)
 dmi2 <- i_to_dmi(i, timesteps, bf)

 isTRUE(all.equal(dmi, dmi2))
}
