% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst_version.R
\name{ebirdst_pkg_ver}
\alias{ebirdst_pkg_ver}
\title{look up ebirdst version}
\usage{
ebirdst_pkg_ver()
}
\value{
ebirdst version or NA if not installed
}
\description{
look up ebirdst version
}
\keyword{internal}
