% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_performance.R
\name{evaluate_performance}
\alias{evaluate_performance}
\title{Evaluate BirdFlow model performance}
\usage{
evaluate_performance(x, distr_only = FALSE)
}
\arguments{
\item{x}{A BirdFlow object}

\item{distr_only}{set to TRUE to calculate only \code{mean_distr_cor} and
\code{min_distr_cor} metrics.}
}
\value{
\describe{
\item{mean_step_cor}{Indicates on average how well the model project a
single timestep. The mean correlation, across all timesteps, between the
training (ebirdst) distribution projected forward one timestep, and the
training distribution for the projected timestep.
}
\item{min_step_cor}{Indicates the quality of the worst single step
projection. The minimum correlation (across all timesteps) between
a single step projection each training distribution and the training
distribution for the projected timestep.}
\item{traverse_cor}{Indicates how well does the model project through all
timesteps. The correlation between last distribution projected iteratively
from the fist training distribution; and the last training distribution. }
\item{mean_distr_cor}{Indicates on average how well the marginal preserves
the training distributions. The mean correlation between the training
distributions and distributions calculated from the marginals.}
\item{min_distr_cor}{Indicates how well the poorest marginal preserves the
training distribution. The minimum observed correlation between a marginal and
training distribution.}
}
}
\description{
Calculate several the correlation between projected distributions and
the ebirdst distributions used to train the BirdFlow model.
}
\details{
"Training distribution" is used to describe the ebirdst
distributions used to train the BirdFlow models.  "Marginal distribution"
describes a distribution calculated from row or column sums of a marginal,
joint probability matrix. "Projected distribution" is used to describe
a training distribution that is multiplied with a transition matrix
(derived from the marginal distribution) to project forward one timestep;
or multiplied repeatedly with transition matrices to project forward
multiple timesteps.

Correlations are calculated for only the non-dynamically masked cells.
}
\examples{
bf <- BirdFlowModels::amewoo
evaluate_performance(bf)
}
\keyword{internal}
