% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_state.R
\name{expand_state}
\alias{expand_state}
\title{Convert a collapsed model state into an expanded state}
\usage{
expand_state(x, geom)
}
\arguments{
\item{x}{Either a vector with one value per active cell in the model or a
matrix with a row for each active cell and a column for each state, or
possible a higher dimension object.  In all cases the first dimension is
for states in the model.}

\item{geom}{The geometry component of a BirdFlow model. For convenience a
full BirdFlow model can be passed as well. The \code{mask}, \code{nrow}, and
\code{ncol} elements are used.}
}
\value{
an expanded version of \code{x} with one additional dimension in which
the first two dimensions represent rows and columns in space and
replace the first dimension in the input.
}
\description{
This function converts a collapsed state in which the first dimension
represents all the active cells in the model into an expanded state in which
the first two dimensions represent the locations of the cell in physical
space (row, col) and are thus suitable for plotting or conversion into
spatial objects.
}
\details{
In its collapsed form a state vector represents the state of a model at one
point in time as a probability distribution. Each value can be interpreted as
the relative abundance or probability of finding a bird at the corresponding
location. A special case is when the vector has mostly 0's and a single 1 in
which case it represents a single position of a bird or group of birds.
Additional dimensions can represent multiple timesteps and/or multiple
individual birds, model runs etc.

The location information in the collapsed state is not easily accessible as
position in the vector is ordered based on row dominant unmasked cells in
landscape. Since R uses column dominant order in matrices, to fill in the
right order the first two dimensions must be reversed.  They are switched
back to proper order after filling.
}
