% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_birdflow.R, R/import_birdflow.R
\name{export_birdflow}
\alias{export_birdflow}
\alias{import_birdflow}
\title{Export and import BirdFlow models in HDF5 format}
\usage{
export_birdflow(bf, file = NULL, format = "hdf5", overwrite = TRUE)

import_birdflow(hdf5, ..., version)
}
\arguments{
\item{bf}{A BirdFlow model}

\item{file}{The file name to write. If not supplied or if the path is to
a directory the file name will default to
\verb{"<species code>_<ebirdst year>_<resolution in (km)>}.}

\item{format}{The format to export either \code{"hdf5"} or \code{"rds"}.}

\item{overwrite}{The default \code{TRUE} will delete a preexisting file and then
write a new one in it's place.  if \code{FALSE}, \code{export_birdflow()} will throw
an error if  \code{file} already exists.}

\item{hdf5}{Path to an HDF5 file containing a fitted BirdFlow model.}

\item{...}{Deprecated, arguments to be passed to a version specific
internal functions.}

\item{version}{(optional) force reading of BirdFlow models as a particular
version. Normally, this will be determined from metadata in the HDF5
file.}
}
\value{
\code{export_birdflow()} writes a file and invisibly returns \code{TRUE}
if successful.

\code{import_birdflow()} returns a BirdFlow object.
}
\description{
\code{export_birdflow()} saves a BirdFlow object to an
HDF5 (Hierarchical Data Format version 5)

\code{import_birdflow()} imports a BirdFlow model from an HDF5
(Hierarchical Data Format version 5) file.
}
\details{
The standard workflow for generating a fitted BirdFlow model is:
\enumerate{
\item Use \code{\link[=preprocess_species]{preprocess_species()}} to download and format eBird data into an
HDF5 file (this calls \code{export_birdflow()} internally).
\item Fit the model and add marginals and hyperparameters to the HDF5 file
with \href{https://github.com/birdflow-science/BirdFlowPy}{BirdFlowPy}
\item Import the model with \code{import_birdflow()}.
\item Save the model to a new file either with:
\itemize{
\item \code{\link[=saveRDS]{saveRDS()}} to a .rds file, or
\item \code{export_birdflow()} to a new .hdf5 file.
}
}

During the first import after fitting with
\href{https://github.com/birdflow-science/BirdFlowPy}{BirdFlowPy}
\code{import_birdflow()} does a fair amount of renaming and reformatting of the
data - for example it renames the marginals and adds a marginal index.
If the resulting BirdFlow model is exported again with \code{export_birdflow()}
these updates are retained, so the internal structure of the HDF5 file
will be different.  \code{import_birdflow()} can handle either format.
}
\section{HDF5 file version}{


The HDF5 BirdFlow model files have an internal version number that is
incremented on major changes to the HDF5 file structure. The file
version is retained in the imported object.
\itemize{
\item Version 1 predates \link[=preprocess_species]{BirdFlowR preprocessing} and only
contains marginals and a few other python objects, importing it requires
having an associated TIFF file that has the extent and distribution data.
The version number (1) is kept in \code{"/version"} within the HDF5 file.
\item Version 2 is the first version that includes preprocessing in BirdFlowR.
It does not include dynamic masking.  The version number (2) is stored
in  \code{"/metadata/birdflow_version"}.  Some version 2 files include
hyper parameters in the HDF5 but these are not read into R.
\item Version 3 marks the transition to dynamic masking, the HDF5 created during
preprocessing gained \code{"/geom/dynamic_mask"} and \code{"/distances"}  (great
circle distance matrix).  As in version 2 the version is
stored in \code{"/metadata/birdflow_version"} but unlike previous versions the
R package version that did the preprocessing is saved in character
format to \code{"/metdata/birdflowr_version"}. When version 3 files are imported
into R with model fits the hyperparameters are saved to
\verb{$metadata$hyperparameters} (a list).
}
}

\examples{
\donttest{
bf <- BirdFlowModels::amewoo
hdf5_path <- tempfile(fileext = ".hdf5")
export_birdflow(bf, hdf5_path)
bf2 <- import_birdflow(hdf5_path)
identical(class(bf), class(bf2))
}
}
\seealso{
\itemize{
\item \code{\link[=sparsify]{sparsify()}} to reduce the object size after importing.
\item \code{\link[=build_transitions]{build_transitions()}} to add transition matrices for quicker processing
at the cost of larger file sizes.
\item \code{\link[=export_rasters]{export_rasters()}} to save the distributions and mask from a
BirdFlow model to raster files.
}
}
