% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tifs.R
\name{export_tifs}
\alias{export_tifs}
\title{export distributions or masks from a BirdFlow object to TIFF files.}
\usage{
export_tifs(
  bf,
  dir,
  crs = NULL,
  multiband = TRUE,
  singleband = !multiband,
  what = c("distr", "mask")
)
}
\arguments{
\item{bf}{A BirdFlow object}

\item{dir}{the directory where output should be stored}

\item{crs}{The coordinate reference system to use in the output files,
defaults to  \code{crs(bf)}}

\item{multiband}{If TRUE export a multiband file}

\item{singleband}{If TRUE export files for each week in the model.}

\item{what}{(character) either "distr", "mask" or, the default, both
(as a two element vector). \code{what} controls what components of \code{bf} are
exported.}
}
\value{
Nothing is returned, but files are created on disk.
}
\description{
This function exports the distributions and/or dynamic masks from a
BirdFlow object to one or more TIFF files.
}
\details{
This replaces the old behavior of \code{\link[=preprocess_species]{preprocess_species()}} which
allowed exporting TIFF files of the distributions while preprocessing.
}
\examples{
\dontrun{
  bf <- load_model("amewoo_prebreeding")
  dir <- tempdir()
  crs <-"EPSG:4326"
  export_tifs(bf, dir = dir, singleband = TRUE, crs = crs)
}

}
\keyword{internal}
