% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_birdflow.R
\name{extend_birdflow}
\alias{extend_birdflow}
\title{Extend BirdFlow extent}
\usage{
extend_birdflow(x, y)
}
\arguments{
\item{x}{A single BirdFlow object, or one or more paths to BirdFlow objects
stored as either hdf5 or rds files.}

\item{y}{An extent or an object that yields an extent when passed
to \code{\link[terra:ext]{terra::ext()}}.}
}
\value{
If \code{x} is a BirdFlow model object \code{extend_birdflow()} returns an
extended version of the same model.  If \code{x} is the path to one or more
BirdFlow models than those files are modified and a logical vector of the
same length is returned with TRUE for success.
}
\description{
Extend BirdFlow extent
}
\details{
Models can only be extended beyond their original extent, however if they
have already been extended it is possible to "extend" them again to
a smaller extent as long as it is also larger than the original extent.
}
\examples{
\donttest{
bf <- BirdFlowModels::amewoo

# Define extended extent for example
e <-  ext(bf)
buffer <- 3 * res(bf)
e[1] <- e[1] - buffer[1]
e[2] <- e[2] + buffer[1]
e[3] <- e[3] - buffer[2]
e[4] <- e[4] + buffer[2]

bf2 <- extend_birdflow(bf, e)

bf3 <- shrink_birdflow(bf2)

# Compare extents
data.frame(item = names(as.vector(ext(bf))),
            initial = as.vector(ext(bf)),
            extended = as.vector(ext(bf2)),
            shrunk = as.vector(ext(bf3)))
}

\dontrun{
# Plot
library(terra)
plot_distr(get_distr(bf, 1), bf)
plot_distr(get_distr(bf2, 1), bf2)
plot_distr(get_distr(bf3, 1), bf3)
  }

}
\seealso{
\code{\link[=shrink_birdflow]{shrink_birdflow()}} returns a model to it's original extent.
}
