% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_birdflow.R
\name{extend_geom}
\alias{extend_geom}
\title{Extend geometry component of a BirdFlow object}
\usage{
extend_geom(geom, y)
}
\arguments{
\item{geom}{The geometry component of a BirdFlow object}

\item{y}{An object that returns an extent when passed to \code{\link[terra:ext]{terra::ext()}},
this can be an extent, a SpatRaster, or a BirdFlow model.}
}
\value{
extended geometry (covering larger area)
}
\description{
This is an internal helper function called twice by \code{\link[=extend_birdflow]{extend_birdflow()}}
it adjust the nrow, ncol, ext, and mask elements of the
geom component of a BirdFlow model to expand the extent while preserving
the same number, location, and alignment of the unmasked cells -
thus nothing else in the object needs to change.
}
\keyword{internal}
