% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_threshold.R
\name{find_threshold}
\alias{find_threshold}
\title{function to determine the threshold in values that preserves a given
proportion of the value.}
\usage{
find_threshold(x, p)
}
\arguments{
\item{x}{A vector of values}

\item{p}{A proportion of x that we should retain}
}
\value{
A threshold in values of x such that dropping everything smaller than
the threshold retains at least \code{p} proportion of the total of \code{x}
}
\description{
function to determine the threshold in values that preserves a given
proportion of the value.
}
