% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_dead_ends.R
\name{fix_current_dead_ends}
\alias{fix_current_dead_ends}
\title{Fix the dead ends currently present in the model}
\usage{
fix_current_dead_ends(bf, de)
}
\arguments{
\item{bf}{BirdFlow model}

\item{de}{(optional) output from \code{\link[=find_dead_ends]{find_dead_ends()}}}
}
\value{
a BirdFlow model with selected marginal rows and columns zeroed out.
#' @keywords internal
}
\description{
THe internal function \code{fix_current_dead_ends()} zeros
out the row or column of the marginal that leads into current dead ends,
eliminating all transitions into the existing dead end states.  It tends to
create new dead ends in the process so is called iteratively by
\code{\link[=fix_dead_ends]{fix_dead_ends()}}.
}
