% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_raster.R
\name{flatten_raster}
\alias{flatten_raster}
\title{Convert a raster bird distribution into its flattened, vector equivalent}
\usage{
flatten_raster(x, bf)
}
\arguments{
\item{x}{Either a matrix representing a single bird distribution or an
array representing multiple distributions with dimensions: row, col, and
distribution.}

\item{bf}{A \code{BirdFlow} model.}
}
\value{
Either a vector representing a single distribution in its collapsed
form or, if \code{x} represents multiple distributions,  a matrix with one
distribution per column.
}
\description{
This function converts between a raster representation of data in which rows
and columns indicate position in space and a vector representation that
contains only the active (not masked) cells in row major order - starting at
the top left and proceeding left to right along each row.  The collapsed form
is used for projecting the flow model while the expanded form is used to
export, import, and visualize the data.
}
\examples{
\donttest{
bf <- BirdFlowModels::amewoo
d <- get_distr(bf, 1)
m <- rasterize_distr(d, bf, format = "numeric")
f <- flatten_raster(m, bf)
all.equal(as.numeric(d), as.numeric(f))
}
}
\seealso{
\itemize{
\item \code{\link[=expand_distr]{expand_distr()}} does the opposite of \code{flatten_raster()}.
\item \code{\link[=rasterize_distr]{rasterize_distr()}} has replaced this function in the public API.  With
\code{format = "numeric"} it returns an identical object to \code{expand_distr()} but
by default (`format = "SpatRaster" it goes one step further and adds spatial
metadata to make a \link[terra:SpatRaster-class]{terra::SpatRaster}.
\item \link{index_conversions} for ways to convert among indexes of the data in
raster row and column, index along the flattened vector, and Cartesian space.
}
}
