% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\title{Forecast bird distributions}
\usage{
forecast(x, distr, start, end, direction)
}
\arguments{
\item{x}{A BirdFlow model object}

\item{distr}{a starting distribution}

\item{start, end}{These define the time period to forecast over. They can be
\link[base:date]{Date} objects, integer timesteps, or a string with
"year-month-day" e.g. "2022-11-28".}

\item{direction}{either "forward" or "backward", only used if \code{start} and
\code{end} represent timesteps.}
}
\value{
a matrix with rows for each location and columns for each timestep
}
\description{
forecast() projects bird distributions into the future or past
}
\details{
given an initial distribution, a start time and end time generate a
probability distributions for each timestep between the start and end.

If using dates the order of the dates are used to determine if the
projection is forward or backward in time and then the dates are resolved to
timesteps. Projections will therefore be at most one year in length and dates
further than a year apart may yield nonsensical results.

It is possible to supply multiple starting distributions as a matrix with
a row for each location and a column for each distribution. In this case
the result will be an array with dimensions: location, distribution,
and timestep.
}
