% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_predictive_distance_metric.R
\name{get_distance_metric}
\alias{get_distance_metric}
\title{To be done}
\usage{
get_distance_metric(intervals, observations, bf)
}
\arguments{
\item{intervals}{To be done}

\item{observations}{To be done}

\item{bf}{BF model}
}
\value{
mean distance metric
}
\description{
To be done
}
