% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_based_validation.R
\name{get_interval_based_metrics}
\alias{get_interval_based_metrics}
\title{Get interval based metrics, including distance metrics
for all transition pairs}
\usage{
get_interval_based_metrics(birdflow_intervals, bf)
}
\arguments{
\item{birdflow_intervals}{A BirdFlowIntervals object}

\item{bf}{BF model}
}
\value{
mean metrics across transition pairs
}
\description{
Get interval based metrics, including distance metrics
for all transition pairs
}
