% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marginal.R
\name{get_marginal}
\alias{get_marginal}
\title{Return a marginal matrix from a BirdFlowR model}
\usage{
get_marginal(x, marginal = NULL, from = NULL)
}
\arguments{
\item{x}{A BirdFlow object}

\item{marginal}{A marginal code, e.g. "M_01-02"}

\item{from}{The first timestep associated with the marginal. Note marginals
are always forward so the second marginal will  be \code{from + 1} or \code{1} (when
\code{from} is the last timestep).}
}
\value{
A  marginal matrix
}
\description{
Marginals in BirdFlow models are stored such that the cell [i, j]
represents the probability of the bird being in state i in the prior
timestep and state j in the next.  Thus the number of rows in the marginal
equals the number of cells within the dynamic mask for the prior timestep
and the columns count is equal to the included cells for the following
timestep.
}
\examples{
\donttest{
bf <- BirdFlowModels::amewoo
get_marginal(bf, from = 1)
get_marginal(bf, marginal = "M_01-02")
}
}
\seealso{
\code{\link[=lookup_transitions]{lookup_transitions()}} will generate a list of the transitions
needed to predict or route between two points in time.  \code{\link[=get_transition]{get_transition()}}
will return a transition matrix - often calculated on the fly from a
marginal.
}
