% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Retrieve BirdFlow metadata}
\usage{
get_metadata(x, what)
}
\arguments{
\item{x}{BirdFlow model}

\item{what}{Indicates what metadata to return. The default \code{"all"} returns
the complete list. Options are described in details.}
}
\value{
Metadata from the BirdFlow model. Potentially a list.
}
\description{
\code{get_metadata} retrieves information about the BirdFlow model: when it was
created, what version of eBird Status & Trends data was used, whether it has
been sparsified,  etc.
}
\details{
\code{what} may be used with these options:\tabular{ll}{
   \code{all} \tab Return the complete list, this is the default \cr
   \code{has_marginals} \tab Logical, indicates whether the model has marginals.  Marginals encode forward and backward transitions, and distributions in an n by n matrix for each pair of adjacent timesteps. \cr
   \code{has_transitions} \tab Logical, indicates whether the model has stored transitions (separate from marginal) \cr
   \code{has_distr} \tab Logical, indicates whether the model has stored distributions (separate from marginals); if it does they are from \pkg{ebirdst} and are the distributions the model was trained on \cr
   \code{n_transitions} \tab Number of transitions in the model. \cr
   \code{n_active} \tab The number of active states in the model. This defines the dimensions of the marginals and transitions matrices and the number of columns in flattened distributions. In a full (not sparse) model all active states could be visited. \cr
   \code{n_timesteps} \tab The number of timesteps in the model. \cr
   \code{ebird_version_year} \tab The version year for the eBird Status and Trends Data the model was trained on see \link[ebirdst:ebirdst_version]{ebirdst_version()} \cr
   \code{ebird_release_year} \tab The year the eBird S & T data was released \link[ebirdst:ebirdst_version]{ebirdst_version()} \cr
   \code{ebird_access_end_date} \tab The date at which the training eBirds S & T data will no longer be available. This does not prevent continued use of the BirdFlow model. \cr
   \code{birdflow_preprocess_date} \tab The date the data was downloaded with \pkg{ebirdst} and formatted for BirdFlow model fitting (see \code{\link[=preprocess_species]{preprocess_species()}} \cr
   \code{birdflow_model_date} \tab The date the model was fitted and/or exported to hdf5 from Python. \cr
   \code{is_sparse} \tab Logical, if \code{TRUE} the model is sparse. See \code{\link[=sparsify]{sparsify()}} \cr
   \code{sparse_stats} \tab If the model is sparse this contains sparsification statistics (a list). \cr
}
}
\examples{
\donttest{
library(BirdFlowModels)
get_metadata(amewoo)
get_metadata(amewoo, "is_sparse")
}
}
\seealso{
\code{\link[=has_transitions]{has_transitions()}}, \code{\link[=has_marginals]{has_marginals()}}, \code{\link[=has_distr]{has_distr()}}, \code{\link[=n_active]{n_active()}},
\code{\link[=n_transitions]{n_transitions()}}, \code{\link[=n_timesteps]{n_timesteps()}} all return metadata components directly.

\code{\link[=species_info]{species_info()}} is similar but for information about the species
represented in the model.

\code{\link[ebirdst:ebirdst_version]{ebirdst::ebirdst_version()}} is called while assembling the model to set
\code{ebird_version_year}, \code{ebird_release_year} and \code{ebird_access_end_date}.

\link[=nrow]{Dimensions} documents getting various attributes of a BirdFlow
model, some of which overlap \code{get_metadata()}.
}
