% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_population.R
\name{get_population}
\alias{get_population}
\title{Return the population estimate for a Bird Flow model or species}
\usage{
get_population(x, what = "abundance")
}
\arguments{
\item{x}{Either a single BirdFlow model, or
one or more species eBird codes, scientific names, or common names.}

\item{what}{Indicates what information will be returned. Options are:
\describe{
\item{\code{"df"}}{A data frame with relevant rows from \link{callaghan_abundance}}
\item{\code{"abundance"}}{The abundance estimate.}
\item{\code{"lower_ci"}}{The lower 95 percent confidence interval}
\item{\code{"upper_ci"}}{The upper 95 percent confidence interval}
}}
}
\value{
See \code{what}. In all cases the values correspond to \code{x} and will be
\code{NA} when \code{x} cannot be cross walked to \code{\link{callaghan_abundance}}.
}
\description{
Return the population estimate for a Bird Flow model or species
}
\examples{
bf <- BirdFlowModels::amewoo
get_population(bf)
get_population("American Black Duck")
get_population("amewoo")
}
\references{
Callaghan, Corey T., Shinichi Nakagawa, and William K. Cornwell. "Global
abundance estimates for 9,700 bird species." Proceedings of the National
Academy of Sciences 118.21 (2021): e2023170118.
\url{https://doi.org/10.1073/pnas.2023170118}
}
