% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transition.R
\name{get_transition}
\alias{get_transition}
\title{Return a transition matrix from a BirdFlowR model}
\usage{
get_transition(x, transition)
}
\arguments{
\item{x}{A BirdFlow object}

\item{transition}{A transition code, e.g. "T_01-02"}
}
\value{
A transition matrix
}
\description{
This function retrieves a transition matrix from a BirdFlow object,
possibly calculating it from a marginal.
}
\details{
\code{get_transition()} will construct a transition matrix from the
marginals if \code{x} doesn't have transitions, or return the relevant stored
transition matrix if it does.

The format for a transition code is "T_[from]-[to]" where [from] and
[to] are timesteps padded with zeros. Direction is important; "T_03-04"
represents a transition backward in time.

The transition matrices are structured such that you multiply the matrix by
the distribution to project either forward or backwards. If more than one
distribution is projected at once they should be stored in columns of a
matrix with dimensions(\code{\link[=n_active]{n_active()}}, \code{\link[=n_distr]{n_distr()}}).

Given a marginal in which the cell [i, j] represents the probability of
the bird being in state i in the prior timestep and state j in the next, to
generate the forward transition matrix we divide each row of the marginal
by its sum and then transpose. Backwards transitions matrices are generated
by dividing each column by its sum, without transposing.
}
\examples{
\donttest{
bf <- BirdFlowModels::amewoo
get_transition(bf, "T_01-02")
}
}
\seealso{
\code{\link[=lookup_transitions]{lookup_transitions()}} will generate a list of the transitions
needed to predict or route between two points in time that can then be
passed to this function. The internal function \code{\link[=transition_from_marginal]{transition_from_marginal()}}
does the calculations.
}
