% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance_matrix.R
\name{great_circle_distance_lonlat_input}
\alias{great_circle_distance_lonlat_input}
\title{Calculate the great circle distance based on longitude and latitude}
\usage{
great_circle_distance_lonlat_input(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1}{latitude of point 1}

\item{lon1}{longitude of point 1}

\item{lat2}{latitude of point 2}

\item{lon2}{longitude of point 2}
}
\value{
the great circle distance
}
\description{
Calculate the great circle distance
}
\keyword{internal}
