% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_birdflow_v3.R
\name{import_birdflow_v3}
\alias{import_birdflow_v3}
\title{import BirdFlow model from an hdf5 file}
\usage{
import_birdflow_v3(hdf5)
}
\arguments{
\item{hdf5}{Path to an hdf5 file}
}
\value{
a BirdFlow object
}
\description{
This function imports a BirdFlow model data from an HDF5 file written by
python. It works for version 2 hdf5.
}
\keyword{internal}
