% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_prototype.R
\name{import_prototype}
\alias{import_prototype}
\title{Function to import BirdFlow model from the Shiny App}
\usage{
import_prototype(base_dir, species)
}
\arguments{
\item{base_dir}{The base directory of the shiny app}

\item{species}{The ebird species code for the model}
}
\value{
a BirdFlow object
}
\description{
This function imports the BirdfFlow model data from the prototype Shiny
application. It is a stopgap until a cleaner workflow is developed to export
models from python in a standard format that can be read by R.
}
