% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_conversions.R
\name{index conversions}
\alias{index conversions}
\alias{index_conversions}
\alias{x_to_col}
\alias{y_to_row}
\alias{row_to_y}
\alias{col_to_x}
\alias{i_to_rc}
\alias{i_to_row}
\alias{i_to_col}
\alias{i_to_x}
\alias{i_to_y}
\alias{i_to_xy}
\alias{rc_to_i}
\alias{xy_to_i}
\alias{latlon_to_xy}
\alias{xy_to_latlon}
\title{Functions to convert among spatial indices}
\usage{
x_to_col(x, bf)

y_to_row(y, bf)

row_to_y(row, bf)

col_to_x(col, bf)

i_to_rc(i, bf)

i_to_row(i, bf)

i_to_col(i, bf)

i_to_x(i, bf)

i_to_y(i, bf)

i_to_xy(i, bf)

rc_to_i(row, col, bf)

xy_to_i(x, y, bf)

latlon_to_xy(lat, lon, bf)

xy_to_latlon(x, y, bf)
}
\arguments{
\item{x, y}{x and y coordinates in the BirdFlow model's CRS.
This typically represents an easting and northing in meters. For functions
that require both \code{x} and \code{y} a two column matrix or data.frame
containing \code{x} and \code{y} columns in that order can be passed to \code{x} in
which case \code{y} should be omitted.}

\item{bf}{A BirdFlow model.}

\item{row, col}{The row and column index of a cell in the BirdFlow model and
associated raster data. Alternatively, a two column matrix or data.frame
containing row and column indices in columns 1 and 2 respectively can be
passed to \code{row} in which case the \code{col} argument should be omitted.}

\item{i}{The index along a state vector that contains the data for
unmasked cells.}

\item{lat, lon}{The latitude and longitude in WGS 1984 (EPSG:4326). A two
column matrix or data frame can also be passed to \code{lat}.}
}
\value{
\describe{

\item{\code{x_to_col(x, bf)} and \code{y_to_row(y, bf)}}{Return the column or row
index that each \code{x} or \code{y} coordinate falls within.}

\item{\code{row_to_y(row, bf)} and \code{col_to_x(col, bf)}}{Return the y or x
coordinate of the center each row or column. }

\item{\code{i_to_row(i, bf)} and \code{i_to_col(i, bf)}}{Return the row or column
index corresponding to the vector state index, \code{i}.}

\item{\code{i_to_rc(i, bf)}}{Returns a two column matrix of the row and column
index in the raster corresponding to the index, \code{i} of the vector state.}

\item{\code{i_to_x(i, bf)} and \code{i_to_y(i, bf)}}{Return the x or y coordinate
from the vector state index, \code{i}.}

\item{\code{i_to_xy(i, bf)}}{Returns a two column matrix of the x and y
coordinates corresponding to the index, \code{i} of the vector state space.}

\item{\code{rc_to_i(row, col, bf)} and \code{xy_to_i(x, y, bf)}}{Return the state
space index corresponding to x and y coordinates or row and column indices.}
}

\item{\code{latlon_to_xy(lat, lon, bf)}}{Returns a two column matrix of the x
and y coordinates corresponding to the supplied latitude and longitude. The
output is in the CRS of \code{bf} (\code{crs(bf)}).}

\item{\code{xy_to_latlon(x, y, bf)}}{Returns a two column matrix of the
latitude and longitude of points in WGS84 given their coordinates in the
BirdFlow object's CRS.}
}
\description{
These functions convert among different ways of referencing
locations in BirdFlow models and output. They convert among the
spatial coordinates (x and y), raster row and col indices, and the
index \code{i} along a location vector. \code{latlon_to_xy()} and \code{xy_to_latlon()}
convert between WGS 1984 latitude and longitude and x and y coordinates
in a BirdFlow object's coordinate reference system (CRS).
}
\details{
In general, these functions mirror NA input in output; and
functions that return rows, columns, or the index (\code{i}) will return NA
if the input isn't within the extent; or for (\code{i}) within the active cells.

For most CRSs longitude translates to the x coordinate and
latitude to the y coordinate. The traditional order of coordinates
latitude and longitude; row and column, and x and y are not consistent, but
that is still what is used here, i.e. latitude, row, and y are generally
correlated, and are first, first, and second in argument order.
}
\examples{
\donttest{
bf <- BirdFlowModels::amewoo
i <- 1 # first active cell

# From state vector index (i) to row, column, x, y, and back
rc <- i_to_rc(i, bf)
rc
i_to_row(i, bf)
i_to_col(i, bf)
xy <- i_to_xy(i, bf)
xy
i_to_x(i, bf)
i_to_y(i, bf)
rc_to_i(rc$row, rc$col, bf)
xy_to_i(xy$x, xy$y, bf)

# Row/column to x/y coordinates and back
x <- col_to_x(rc$col, bf)
y <- row_to_y(rc$row, bf)
x_to_col(x, bf)
y_to_row(y, bf)

# Latitude/longitude (WGS84) to and from the model's CRS
latlon <- xy_to_latlon(xy$x, xy$y, bf)
latlon
latlon_to_xy(latlon$lat, latlon$lon, bf)
}
}
\seealso{
\itemize{
\item \code{\link[=expand_distr]{expand_distr()}} converts a vector distribution into it's raster (matrix)
equivalent or a matrix (representing multiple distributions) into an array
equivalent.
\item \code{\link[=rasterize_distr]{rasterize_distr()}} converts a vector distribution into a
\link[terra:SpatRaster-class]{terra::SpatRaster} - similar to those created by \code{\link[terra:rast]{terra::rast()}}.
\item \code{\link[=dmi_to_i]{dmi_to_i()}} and \code{\link[=i_to_dmi]{i_to_dmi()}} convert between location indices (\code{i}) along
cells included by the static mask and dynamic mask indices (\code{dmi})
along the cells included in the dynamic mask for a given timestep.
}
}
