% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{load_model}
\alias{load_model}
\title{Load BirdFlow models from a collection}
\usage{
load_model(model, update = TRUE)
}
\arguments{
\item{model}{The model name to load}

\item{update}{If TRUE (the default) then both the index and cached model file
are verified against the server's version to make sure they are up-to-date.
If FALSE then neither the index or the model will be checked. Set to FALSE
after downloading the model(s) you need if you want to make sure the model
does not change during your analysis (even if updated on the server); or
if working offline.}
}
\value{
The designated BirdFlow model is returned.
}
\description{
Load a named model from the current (likely the default) model collection.
If the model is not cached locally it will be downloaded. If the model
doesn't exist in the local cache or isn't up-to-date the the cache will be
updated prior to loading.
}
