% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_transitions.R
\name{lookup_transitions}
\alias{lookup_transitions}
\title{Lookup a series of transitions connecting two dates or timesteps}
\usage{
lookup_transitions(x, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=lookup_timestep_sequence]{lookup_timestep_sequence}}
  \describe{
    \item{\code{x}}{A BirdFlow object}
    \item{\code{start,end}}{The starting and ending points in time specified as
\code{timesteps}, dates as characters or date objects, or if  \code{end} is omitted
start may be set to "all" or a season name to be interpreted by
\code{\link[=lookup_season_timesteps]{lookup_season_timesteps()}}.}
    \item{\code{direction}}{Either "forward" or "backward".  Determines direction if
\code{start} and \code{end} are timesteps (numeric), or \code{start} is "all" or a season.
If \code{start} and \code{end} represent dates and \code{direction} is used an error will
be thrown if \code{direction} isn't consistent with direction implied by the
dates.  If \code{start} and \code{end} do not represent dates then \code{direction} will
default to "forward".}
    \item{\code{buffer}}{Only used if \code{start} is a season. A number of
timesteps to extend the season by. It is passed to
\code{\link[=lookup_season_timesteps]{lookup_season_timesteps()}} and defaults to 1.}
  }}
}
\value{
A character vector with the named transitions required to get between
\code{start} and  \code{end}
}
\description{
\code{lookup_transitions()} returns an ordered vector of transition names that
connect start to end. If \code{start} and \code{end} are dates than their order
determines whether the transitions flow forward or backward in time.  If they
are timesteps than the \code{direction} argument should be used to indicate
whether to project "forward" or "backward" in time possibly passing the year
boundary.
}
\details{
Transitions are named "T_[from]-[to]" where [from] and [to]
are timesteps padded with zeros. Direction is important; "T_03-04"
represents a transition backward in time.
}
\keyword{internal}
