% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cache_readme.R
\name{make_cache_readme}
\alias{make_cache_readme}
\title{Internal function to create or update readme files within the local
BirdFlow model cache.}
\usage{
make_cache_readme(collection_url)
}
\value{
Nothing is returned
}
\description{
This is called anytime new files are downloaded to the cache by
\code{\link[=load_collection_index]{load_collection_index()}} or \code{\link[=load_model]{load_model()}}.
}
\keyword{internal}
