% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dates.R
\name{make_dates}
\alias{make_dates}
\title{Internal function to make the dates component of a BirdFlow model}
\usage{
make_dates()
}
\value{
Dates table appropriate for the current version of eBirdst
}
\description{
Called from \link{preprocess_species} and not intended for other use.
}
\keyword{internal}
