% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_marginal_index.R
\name{make_marginal_index}
\alias{make_marginal_index}
\title{Internal function to make a marginal index for a BirdFlow object.}
\usage{
make_marginal_index(bf)
}
\arguments{
\item{bf}{A BirdFlow object}
}
\value{
A data.frame that facilitates looking up marginals from transition
names.  It has columns:
\item{from}{starting timestep for transition (direction matters)}
\item{to}{ending timestep for transition}
\item{direction}{transition direction, either "forward", or "backward"}
\item{transition}{transition name e.g. "T_02-02"}
\item{marginal}{marginal name e.g. "M_01-02", order matches
forward transition order, so smaller number is generally first
except with the last marginal in a circular model eg "M_52-01"}
}
\description{
\code{make_marginal_index()} is called by \verb{[import_birdflow()]}
and \code{truncate_birdflow()} to add a marginal index to a BirdFlow object.
}
\details{
\code{\link[=n_transitions]{n_transitions()}}, \code{\link[=n_timesteps]{n_timesteps()}}, \code{\link[=is_cyclical]{is_cyclical()}}, and
\code{\link[=get_timestep_padding]{get_timestep_padding()}} need to work on the object, which means that
\code{metadata$n_timesteps}, \code{metatdata$n_transitions}, and
\code{metadata$timestep_padding} should all  be properly set.
The marginals don't have to exist yet.
}
\keyword{internal}
