% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pyear_breaks.R
\name{make_pyear_breaks}
\alias{make_pyear_breaks}
\title{Generate breakpoints in proportion of year values that correspond to nice
dates}
\usage{
make_pyear_breaks(range, bf, target_n = 8, hpy = TRUE)
}
\arguments{
\item{range}{A range in proportion of year (PY) or half proportion of year
(HPY) values over which breaks will be calculated.}

\item{bf}{A BirdFlow object - only used with \code{get_dates(bf)}.}

\item{target_n}{The target number of breaks}

\item{hpy}{if \code{TRUE} than treat range as HPY. If FALSE as PY.}
}
\value{
a sequence of break points in PY or HPY units.
}
\description{
This internal function is used by \code{\link[=plot_routes]{plot_routes()}} to determine where to
set the breaks (labels) in the color scale. The color corresponds to dates
represented as a proportion of the year.  This functions picks breaks that
map to quarters of the year, the first of the month, the first and fifteenth
of the month, or eBird S&T nominal weeks. Choosing from the above based on
which method is closest to the target number of breaks (\code{target_n}).
}
\keyword{internal}
