% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_stats.R
\name{marginal_stats}
\alias{marginal_stats}
\alias{sum_marginals}
\alias{calc_pct_zero}
\title{marginal statistics}
\usage{
marginal_stats(bf)

sum_marginals(bf)

calc_pct_zero(bf)
}
\arguments{
\item{bf}{BirdFlowModel}
}
\value{
\code{marginal_stats()} returns a list with
\item{sum}{the sum of all the marginals}
\item{pct_zero}{the percent of the values across all marginals that are
zero.}

\code{sum_marginals()} returns the sum of all marginals.

\code{calc_pct_zero()} returns the percent of marginal values that are zero
}
\description{
Internal functions to calculate the sum of all marginals and the
percent of marginal values that are zero.
}
\details{
These are used by sparsify and fix_dead_ends.  marginal_stats is slightly
more efficient than calling the other two functions independently.
}
