% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mollweide_wkt.R
\docType{data}
\name{mollweide_wkt}
\alias{mollweide_wkt}
\title{mollweide_wkt}
\format{
This is a string defining the mollweide projection in well known text 2
}
\usage{
mollweide_wkt
}
\description{
This defines the projection used within BirdFlow models
}
\details{
https://spatialreference.org/ref/sr-org/7399/
}
\references{
https://epsg.io/54009
}
\keyword{datasets}
