% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_birdflow.R
\name{new_BirdFlow}
\alias{new_BirdFlow}
\title{Private function to create the empty shell of a BirdFlow object}
\usage{
new_BirdFlow()
}
\value{
An empty BirdFlow object
}
\description{
Private function to create the empty shell of a BirdFlow object
}
\keyword{internal}
