% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_expand_distr.R
\name{old_expand_distr}
\alias{old_expand_distr}
\title{Convert a collapsed model state into an expanded state}
\usage{
old_expand_distr(x, obj)
}
\arguments{
\item{x}{Either a vector with one value per active cell in the model or a
matrix with a row for each distribution and a column for each active cell.}

\item{obj}{A BirdFlow model or the \code{geom} component of one.}
}
\value{
An expanded version of \code{x} with one additional dimension in which
the first two dimensions represent rows and columns in space and
replace the first dimension in the input.
}
\description{
This function converts a vector representation of a distribution into a raster
matrix in which the rows and columns are ordered in space.
}
\details{
Or, it converts a matrix representation of n distributions (1 per row)
into an array with dimensions: rows, columns, n.

In its collapsed form a distribution vector has one element for each active cell
in the landscape.  Each value can be interpreted as
the relative abundance or probability of finding a bird at the corresponding
location.

A special case is when the vector has mostly 0's and a single 1 in
which case it represents a single position of a bird or group of birds.

If  \code{x} has two dimensions it can represent multiple time steps and/or multiple
individual birds, model runs etc.

The location information in the collapsed state is not easily accessible as
position in the vector is ordered based on row-major ordered unmasked cells
in the extent.
}
