% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bmtr.R
\name{plot_flux}
\alias{plot_flux}
\title{Plot Bird Flow Migration Traffic Rate (BMTR)}
\usage{
plot_flux(flux, ...)
}
\arguments{
\item{flux}{the output from \code{\link[=calc_bmtr]{calc_bmtr()}} or, deprecated, \code{\link[=calc_flux]{calc_flux()}}}

\item{...}{
  Arguments passed on to \code{\link[=plot_bmtr]{plot_bmtr}}
  \describe{
    \item{\code{bf}}{A BirdFlow object}
    \item{\code{subset}}{A subset of the transitions in \code{bmrt} to plot, can be
a logical vector of the same length as the number of transitions in \code{bmtr};
a numeric index of transitions in \code{bmtr}, or a subset of the transition names
in \code{bmtr}.}
    \item{\code{limits}}{Two numbers representing the range in bmtr values to
display. Values outside of this range will be truncated to the range. With
the default of \code{NULL} the entire range is plotted.}
    \item{\code{dynamic_scale}}{If \code{TRUE} then the range of the data in each
transition is mapped to the color palette. This makes it easier to see
the variation within a single transition but results in an inconsistent
scale among transitions.}
    \item{\code{coast_linewidth}}{The line width used to plot the coast. Set to \code{NULL}
to skip plotting the coastline.}
    \item{\code{coast_color}}{The color used to plot the coastline, or \code{NULL} to skip
plotting the coastline.}
    \item{\code{gradient_colors}}{The colors palette used to represent the BMTR
intensity.}
    \item{\code{title}}{The plot title}
    \item{\code{value_label}}{The label for the BMTR values.}
    \item{\code{transform}}{A transformation to apply to the color scaling.
\code{"identity"}, and \code{"sqrt"} are recommended.
If \code{"log"} is used zeros will be replaced with
1/2 the smallest non-zero value prior to transforming.
Legend will still reflect the original values.
Passed to \code{\link[ggplot2:scale_gradient]{ggplot2::scale_color_gradient()}}.}
    \item{\code{crop_bbox}}{An optional named numeric vector with elements \code{xmin},
\code{xmax}, \code{ymin}, \code{ymax} in the model CRS to zoom the plot to a specific
region.  See also \code{sf::st_bbox()} which generates a suitable object.
When \code{NULL} (default) the full model extent is shown.}
  }}
}
\value{
\code{plot_bmtr} returns a \strong{ggplot2} object.  It can be displayed with
\code{print()}.
}
\description{
DEPRECATED FUNCTION.  Please use \code{\link[=plot_bmtr]{plot_bmtr()}} instead.
}
