% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_movement_vectors.R
\name{plot_movement_vectors}
\alias{plot_movement_vectors}
\title{Plot bird movement vectors}
\usage{
plot_movement_vectors(bf, start = 1, direction = "forward", mv)
}
\arguments{
\item{bf}{a BirdFlow object}

\item{start}{The starting timestep for the transition to be modeled}

\item{direction}{"Forward" by default. Set to "backward" to calculate
vectors for a transition backwards in time.}

\item{mv}{This is optional and primarily for internal use by
\code{\link[=animate_movement_vectors]{animate_movement_vectors()}} it allows explicitly providing movement vectors
as a data frame, in which case \code{start} and \code{direction} are ignored.}
}
\value{
an object that inherits classes \code{gg} and \code{ggplot} it can be plotted
with \code{print()}.
}
\description{
This plots the mean movement out of every cell in a BirdFlow model for a
given starting time, based on the transition probabilities for the following
transition.  Each arrow starts in a cell center and ends at the average
location projected for Birds from that cell. The line width and alpha
(transparency) are set such that arrows have more visual weight if there
is a higher probability of a bird being at the starting location and time.
}
\details{
Importantly, this is a visual representation of the transitions encoded in
the model, not of the average movement of a bird through a given cell at a
given point in time, which would involve both the birds starting at that
cell and birds passing through it from other starting locations.
}
\examples{
\donttest{
bf <- BirdFlowModels::amewoo
plot_movement_vectors(bf, start = 7)
}


}
\seealso{
\itemize{
\item \code{\link[=calc_movement_vectors]{calc_movement_vectors()}} for the vector calculations.
\item \code{\link[=animate_movement_vectors]{animate_movement_vectors()}} to produce animations of the vectors over
time.
}
}
