% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single_route.R
\name{plot_single_route}
\alias{plot_single_route}
\title{plot_single_route()}
\usage{
plot_single_route(points, bf, route, max_stay_len)
}
\arguments{
\item{points}{The points component of a list returned by \code{\link[=route]{route()}}
or \code{\link[=route_migration]{route_migration()}}}

\item{bf}{A BirdFlow object}

\item{route}{The route number to plot}

\item{max_stay_len}{This is used to determine the scaling of the circles
that indicate the stay length. To plot many routes consistently set this
to the maximum observed value across all routes.}
}
\value{
a ggplot object.  Use \code{\link[=print]{print()}} to display it.
}
\description{
Function to plot a single route while using color to indicate when the bird
moved.
}
\details{
This function is provisional and might change or be dropped.
}
