% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_predictive_distance_metric.R
\name{predict_encounter_distribution}
\alias{predict_encounter_distribution}
\title{To be done}
\usage{
predict_encounter_distribution(b_e, bf, gcd, st_dists)
}
\arguments{
\item{b_e}{To be done}

\item{bf}{To be done}

\item{gcd}{To be done}

\item{st_dists}{To be done}
}
\value{
A dataframe
}
\description{
To be done
}
