% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.BirdFlow}
\alias{print.BirdFlow}
\title{Print BirdFlow models}
\usage{
\method{print}{BirdFlow}(x, ...)
}
\arguments{
\item{x}{A BirdFlow object.}

\item{...}{arguments passed from other methods}
}
\value{
\code{x} returned invisibly and unchanged.
}
\description{
Print a summary of a \code{BirdFlow} object: species, raster
dimensions, resolution, number of active cells, and in-memory object
size.
}
\examples{
\donttest{
bf <- BirdFlowModels::amewoo
print(bf)
}
}
