% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RouteDataFunction.R
\name{print.BirdFlowRoutes}
\alias{print.BirdFlowRoutes}
\title{Print a BirdFlowRoutes Object}
\usage{
print.BirdFlowRoutes(birdflow_routes)
}
\arguments{
\item{birdflow_routes}{A \code{BirdFlowRoutes} object to print.}
}
\value{
Invisibly returns the input \code{birdflow_routes} object.
}
\description{
Custom print method for \code{BirdFlowRoutes} objects, summarizing their contents, metadata,
and BirdFlow-specific attributes.
}
\examples{
# Create a BirdFlowRoutes object
birdflow_routes <- BirdFlowRoutes(route_df, species = NULL, geom = NULL, dates = NULL)
print(birdflow_routes)
}
