% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Routes.R
\name{print.Routes}
\alias{print.Routes}
\alias{print.BirdFlowRoutes}
\title{Print Routes and BirdFlowRoutes objects}
\usage{
\method{print}{Routes}(x, ...)
}
\arguments{
\item{x}{A \code{Routes} or \code{BirdFlowRoutes} object to print.}

\item{...}{other arguments not used by this method.}
}
\value{
Invisibly returns the input object.
}
\description{
Print a summary of a \code{Routes} and \code{BirdFlowRoutes} objects
}
\examples{
# Create a Routes object
route_df <- data.frame(
  route_id = c("001", "001", "001", "001", "001",
  "003", "003", "003", "004"),
  date = as.Date(c("2025-01-01", "2025-01-08", "2025-01-15", "2025-01-21",
  "2025-02-10", "2025-03-01", "2025-05-01", "2025-06-01", "2025-05-01")),
  lon = c(-75.0060, -75.0060, -74.0060, -87.6298, -87.6298, -87.6298,
  -89.6298, -85.6298, -95.3698),
  lat = c(39.7128, 39.7128, 40.7128, 41.8781, 41.8781, 41.8781,
  42.8781, 40.8781, 29.7604),
  route_type = c("tracking", "tracking", "tracking", "tracking",
  "tracking", "motus", "motus", "motus", "motus")
)
routes <- as_routes(route_df,
  species = list(common_name = "American Woodcock"))

print(routes)

# BirdFlowRoutes
\donttest{
bf <- BirdFlowModels::amewoo
bf_routes <- as_birdflow_routes(routes, bf)
}

}
