% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RouteDataFunction.R
\name{print_type_breakdown}
\alias{print_type_breakdown}
\title{Print summary of route types, their count, and number of points}
\usage{
print_type_breakdown(x, crossline)
}
\arguments{
\item{x}{A \code{Routes}, \code{BirdFlowROutes} or \code{BirdFlowIntervals} object.}

\item{crossline}{Characters to be printed after printing the summary}
}
\value{
invisibly returns the table
}
\description{
This internal helper function is for use in the print methods for the
route and interval data classes.
}
\details{
It only prints the breakdown of routes by type if there is more than one type
of route in \code{x}.  When there's only one type the output would be redundant.
}
\keyword{internal}
