% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_distr.R
\name{rasterize}
\alias{rasterize}
\alias{rasterize_distr}
\alias{rast,BirdFlow-method}
\title{Convert a BirdFlow distribution into a raster}
\usage{
rasterize_distr(distr, bf, format = "SpatRaster")

\S4method{rast}{BirdFlow}(x, which = "all")
}
\arguments{
\item{distr}{A distribution in its vector form or a
matrix in which each column represents a different distribution.}

\item{bf}{A BirdFlow object.}

\item{format}{One of \code{'SpatRaster'} for a \link[terra:SpatRaster-class]{terra::SpatRaster} object,
\code{'numeric'} for a matrix or array, or\code{'dataframe'} for raster data
suitable for plotting with \code{\link[ggplot2:geom_tile]{ggplot2::geom_raster()}}}

\item{x}{A BirdFlow object.}

\item{which}{Indicates which timesteps to return. Can be one or more
integers indicating timesteps; character dates in the format
year-month-day e.g. \code{"2019-02-25"}; \code{\link[base:Dates]{Date}} objects;
or \code{"all"} which will return distributions for all timesteps.}
}
\value{
For \code{rasterize_distr()} the return type depends on the \code{format} argument:
\itemize{
\item \code{"SpatRaster"} (the default) returns a \link[terra:SpatRaster-class]{terra::SpatRaster} object.
\item \code{"numeric"} returns a matrix (one distribution) or array (multiple
distributions).  In either case the first two dimensions will be y (rows),
and x (columns).
\item \code{"dataframe"} will return the raster information in a data frame with a
row for every value (long format) with columns:
\itemize{
\item \code{x}, \code{y} the x and y coordinates of the cell center.
\item \code{i} the location index (in \code{bf}) of the cell.
\item \code{label} The label associated with the distribution, taken from column
names of \code{distr} - typically it indicates time. It is an ordered factor
with the level order matching the order of the distribution in \code{distr}.
The ordered factor is helpful when animating.
\item \code{value}  The cell value, typically density
\item \code{order}  The column index of the distribution in \code{distr} or if only one
distribution \code{1}.
The object is suitable for plotting with [\link[ggplot2:geom_tile]{ggplot2::geom_raster}].
}
}

\code{rast()} returns a \link[terra:SpatRaster-class]{terra::SpatRaster}
}
\description{
\code{rast()} converts a BirdFlow object directly
to a \link[terra:SpatRaster-class]{SpatRaster}.
\code{rasterize_distr()} converts a \href{as_distr()}{distribution} into a
\link[terra:SpatRaster-class]{SpatRaster}, numeric matrix or array, or a raster data
frame.
}
