% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_distr.R
\name{rasterize_distr}
\alias{rasterize_distr}
\title{Function to convert a state vector into a SpatRaster object}
\usage{
rasterize_distr(x, obj)
}
\arguments{
\item{x}{A distribution in its vector form or a matrix in which each column
represents a different distribution.}

\item{obj}{A BirdFlow model.}
}
\value{
A \link[terra:SpatRaster-class]{terra::SpatRaster} object.
}
\description{
This function creates a \link[terra:SpatRaster-class]{SpatRaster} similar to those
created by \code{\link[terra:rast]{terra::rast()}} from one or more distributions in their compact
vector form.
}
