% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_state.R
\name{rasterize_state}
\alias{rasterize_state}
\title{Function to convert a state vector into a SpatRaster object}
\usage{
rasterize_state(x, obj)
}
\arguments{
\item{x}{A vector of a model state, typically representing the
distribution of a species.}

\item{obj}{A BirdFlow model.}
}
\value{
A \link[terra:SpatRaster-class]{terra::SpatRaster} object.
}
\description{
This function creates a \link[terra:SpatRaster-class]{SpatRaster} similar to those
created by \code{\link[terra:rast]{terra::rast()}} from a model state.
}
\details{
It may eventually be replaced with one or more rast() methods.
}
