% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_geom.R
\name{read_geom}
\alias{read_geom}
\title{Read geom component of a BirdFlow hdf5}
\usage{
read_geom(hdf5)
}
\arguments{
\item{hdf5}{the path to an hdf5 file}
}
\value{
The geom component of a birdflow model
}
\description{
This internal function is called by \code{\link[=import_birdflow]{import_birdflow()}} and
\code{\link[=extend_birdflow]{extend_birdflow()}} to read and format the geom component of the model
}
\keyword{internal}
