% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_timestep.R
\name{reformat_timestep}
\alias{reformat_timestep}
\title{reformat timestep labels}
\usage{
reformat_timestep(x, bf)
}
\arguments{
\item{x}{one (vector) or more (matrix) distributions, with column lablels
consisting of a "t" and the timestep.}

\item{bf}{A BirdFlow object}
}
\value{
x with update column labels, as dictated by
\link{birdflow_options("time_format")}
}
\description{
given a vector of timestep labels provide a vector of formated labels based
on the value of \code{birdflow_options("time_format")} (see that function for
options.)
}
\details{
Internally distributions are labled with "t" and the timestep integer.
When returning them to the user \code{\link[=update_distr_labels]{update_distr_labels()}} is called to change
the format which in turn calls this function.
}
\keyword{internal}
