% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst_version.R
\name{res_label}
\alias{res_label}
\title{Lookup up resolution label for installed ebirdst package version}
\usage{
res_label(res)
}
\arguments{
\item{res}{A resolution label. One of "lr", "mr", "hr", "27km", "9km", or
"3km}
}
\value{
The resolution label in the format used by the currently installed
version of ebirdst.  With the 3.2002.0 they switched from the the
abbreviation versions  ("lr") to the numeric resolution versions ("3km")
}
\description{
Lookup up resolution label for installed ebirdst package version
}
\examples{
  res_label("hr")
  res_label(c("lr", "mr", "hr"))
}
