% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_migration.R
\name{route_migration}
\alias{route_migration}
\title{generate migration routes from a BirdFlow model}
\usage{
route_migration(bf, n, migration = "prebreeding", season_buffer = 1)
}
\arguments{
\item{bf}{\code{BirdFlow} model}

\item{n}{the number of routes to generate}

\item{migration}{"prebreeding", "pre", or "spring" for the prebreeding
migration; or "postbreeding", "post", or "fall" for the postbreeding
migration.}

\item{season_buffer}{a buffer in timesteps (likely weeks) to add to the beginning
and end of the season. The default of 1 means we start a week before the
metadata suggests the season starts and continue one week past the end.}
}
\value{
This will likely change. Currently returns a list with:
\item{points}{A data.frame with coordinates, date, and route id}
\item{lines}{a \link[sf:sf]{sf} object containing one line per route.}
}
\description{
create stochastic migration routes for a species by sampling appropriate
starting locations and then \link[=route]{routing} for the
duration of the migration window.
}
\details{
The two migration periods are defined based on the species information in the
BirdFlow model:
\describe{
\item{\code{prebreeding}}{ migration starts at \code{species_info("nonbreeding_end")} and ends at \code{species_info("breeding_start")}}
\item{\code{postbreeding}}{ migration starts at \code{species_info("breeding_end")} and ends at \code{species_info("nonbreeding_start")}}
}
}
\examples{
  bf <-  BirdFlowModels::amewoo
  rts <- route_migration(bf, 2)
  plot(rts$lines)
  head(rts$points)

}
\seealso{
\code{route_migration()} is a convenience wrapper for \code{\link[=route]{route()}}.
\link[=predict.BirdFlow]{predict()} projects future or past distributions based
on a starting location or distribution.
\code{\link[=lookup_season_timesteps]{lookup_season_timesteps()}} does what its name suggests.
}
