% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Routes-internal.R
\name{Routes-internal}
\alias{Routes-internal}
\alias{new_routes}
\alias{new_birdflow_routes}
\alias{new_birdflow_intervals}
\title{Internal (private) routes and intervals class creation functions}
\usage{
new_routes(data, species, source)

new_birdflow_routes(
  data,
  species,
  metadata,
  geom,
  dates,
  source = NULL,
  sort_id_and_dates = TRUE,
  reset_index = FALSE,
  stay_calculate_col = "date"
)

new_birdflow_intervals(data, species, metadata, geom, dates, source = NULL)
}
\arguments{
\item{data}{A data frame containing route/interval data for \code{Routes},
\code{BirdFlowRoutes} or \code{BirdFlowIntervals}.}

\item{species}{Either a single character that will be passed to
\code{\link[ebirdst:get_species]{ebirdst::get_species()}} to lookup species information or a list with
species metadata which must include \code{common_name} and can optionally
also include \code{scientific_name} and  \code{species_code} or any other standard
BirdFlow species metadata. See \code{\link[=species_info]{species_info()}} for a description
of the full list.}

\item{source}{A character string indicating the source of the data.}

\item{metadata}{A list with additional metadata.}

\item{geom}{A list describing spatial geometry,
such as \code{nrow}, \code{ncol}, \code{crs}, and \code{mask}.}

\item{dates}{A data frame with date-related information,
including \code{date}, \code{start}, \code{end}, and \code{timestep}.}

\item{sort_id_and_dates}{Logical. Should the data be sorted by \code{route_id}
and \code{dates}?}

\item{reset_index}{Logical. Should the index of the data frame be reset
after sorting?}

\item{stay_calculate_col}{The column name for calculating the \code{stay_id} and
\code{stay_len} in \code{BirdFlowRoutes} object. Defaults to \code{date}.}
}
\value{
Each function returns an S3 object of the corresponding class
(\code{Routes}, \code{BirdFlowRoutes}, or \code{BirdFlowIntervals}).
}
\description{
Internal (private) functions to create and validate
\code{Routes}, \code{BirdFlowRoutes}, and \code{BirdFlowIntervals} objects.

These functions ensure input data meets the required structure and standard
for use within BirdFlow models.
}
\details{
\itemize{
\item \strong{\code{new_routes()}}: Creates a \code{Routes} object from a data frame.
\item \strong{\code{new_birdflow_routes()}}: Creates a \code{BirdFlowRoutes} object,
extending \code{Routes} with additional BirdFlow-specific spatial and
temporal information.
\item \strong{\code{new_birdflow_intervals()}}: Creates a \code{BirdFlowIntervals} object,
representing intervals between timesteps in BirdFlow data.
}

All objects are internally validated during creation, ensuring required
columns, valid data types, and proper formats.
}
\seealso{
\itemize{
\item \code{\link[=as_routes]{as_routes()}} Create a \code{Routes} object
\item \code{\link[=as_birdflow_routes]{as_birdflow_routes()}} Convert \code{Routes} to \code{BirdFlowRoutes}
\item \code{\link[=as_birdflow_intervals]{as_birdflow_intervals()}} Extract movement between pairs of locations
from \code{BirdFlowRoutes} for use with model evaluation.
\item \href{?object_validators}{Object Validators} Private functions for validating
routes and intervals.
}
}
\keyword{internal}
