% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_state.R
\name{sample_state}
\alias{sample_state}
\title{Sample locations from a model state}
\usage{
sample_state(x, n = 1)
}
\arguments{
\item{x}{a vector representing a single model state; or a matrix or array
representing multiple states. If more than 1 dimension the first should be
used to represent the locations. The values in x are treated as relative
probability of the species being in each position.}

\item{n}{Only used if \code{x} is a vector representing a single model state, in
which case that model state will be sampled \code{n} times to generate a
matrix representing \code{n} sampled locations from that distribution.}
}
\value{
an object with the same dimensions as \code{x} in which the distribution
for each model state in \code{x} all the probability is assigned to a single
location containing a 1 while the rest contain 0s.
}
\description{
This function treats \code{x} as one or more probability distributions and converts
each to a single location by assigning a 1 to one value and
converting the rest to 0.
}
\examples{
# Generate a hypothetical state space vector
# representing the probability of finding a bird in each location.
# This is based on a normal distribution with higher probabilities in the
# center of the vector
s <- dnorm(seq(-2, 2, by = .25))
s <- s / sum(s)


# sample it once
loc <- sample_state(s)

# sample it 200 times
locations <- sample_state(s, 200)

\dontrun{
# Plot the probability of each location
plot(1:length(s), s, xlab = "location", ylab = "prob")
# Plot the number of times each location was selected
plot(1:length(s), apply(locations, 1, sum),
     xlab = "location", ylab = "times sampled")
}
}
