% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorten_distance_matrix.R
\name{shorten_distance_matrix}
\alias{shorten_distance_matrix}
\alias{expand_distance_matrix}
\title{shorten and expand distance matrices}
\usage{
shorten_distance_matrix(dm)

expand_distance_matrix(vals)
}
\arguments{
\item{dm}{a distance matrix, a symmetrical n x n, matrix which contains all
the distances among n locations.}

\item{vals}{The values from a lower triangle of a distance matrix in row
major order.}
}
\value{
\code{shorten_distance_matrix()} returns a vector of values.
\code{expand_distance_matrix()} returns the full matrix given that vector.
}
\description{
These functions are to facilitate storing just the non-duplicated elements
of a distance matrix in a vector to and hdf5 file.
Since the distance matrix is symmetrical and has zero's on the diagonal all
the distances can be stored in a vector that contains fewer than half the
values in the full matrix.
}
\details{
\code{shorten_distance_matrix()} extracts the lower triangle in column major
order from a distance matrix.

\code{expand_distance_matrix()} will reassemble the full distance matrix

The purpose is to halve the number of values being stored in the hdf5 file.

For a 5 x 5 distance matrix it would be values corresponding with the
numbered cells below.\tabular{lrrrrr}{
    \tab r1 \tab r2 \tab r3 \tab r4 \tab r5 \cr
   c1 \tab NA \tab NA \tab NA \tab NA \tab NA \cr
   c2 \tab 1 \tab NA \tab NA \tab NA \tab NA \cr
   c3 \tab 2 \tab 5 \tab NA \tab NA \tab NA \cr
   c4 \tab 3 \tab 6 \tab 8 \tab NA \tab NA \cr
   c5 \tab 4 \tab 7 \tab 9 \tab 10 \tab NA \cr
}


Due to the symmetry of a distance matrix that's equivalent to the upper
triangle in row major order - which is probably how it will be treated
in python.\tabular{lrrrrr}{
    \tab c1 \tab c2 \tab c3 \tab c4 \tab c5 \cr
   r1 \tab NA \tab 1 \tab 2 \tab 3 \tab 4 \cr
   r2 \tab NA \tab NA \tab 5 \tab 6 \tab 7 \cr
   r3 \tab NA \tab NA \tab NA \tab 8 \tab 9 \cr
   r4 \tab NA \tab NA \tab NA \tab NA \tab 10 \cr
   r5 \tab NA \tab NA \tab NA \tab NA \tab NA \cr
}
}
\examples{
\dontrun{
x <- runif(5, 1, 100)
y <- runif(5, 1, 100)
dm <- as.matrix(dist(cbind(x, y)))
a <- shorten_distance_matrix(dm)
dm2 <- expand_distance_matrix(a)
all.equal(dm, dm2, check.attributes = FALSE)
}
}
\keyword{internal}
