% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink_birdflow.R
\name{shrink_geom}
\alias{shrink_geom}
\title{Shrink geometry component of a BirdFlow object}
\usage{
shrink_geom(geom)
}
\arguments{
\item{geom}{The geometry component of a BirdFlow object}
}
\value{
shrunk (or original) geometry
}
\description{
This is an internal helper function called twice by \code{\link[=shrink_birdflow]{shrink_birdflow()}}
it adjusts the \code{nrow}, \code{ncol}, \code{ext}, and \code{mask} elements of the
\code{geom} component of a BirdFlow model to the original extent,
while preserving the same number, location, and alignment of the
unmasked cells - thus nothing else in the object needs to change.
If the geometry has not be extended (\code{\link[=shrink_birdflow]{shrink_birdflow()}}, \code{\link[=extend_geom]{extend_geom()}})
then it is returned as is.
}
\keyword{internal}
