% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_info.R
\name{species}
\alias{species}
\title{lookup species information from a BirdFlow model}
\usage{
species(x, what)
}
\arguments{
\item{x}{A \code{BirdFlow} model}

\item{what}{(optional) 'all' for all information (in a list) or the name of
the information to be returned. See details for full option list.}
}
\value{
if \code{what} is \code{all} a list of all the species information available.
otherwise the element named by \code{what}.
}
\description{
BirdFlow model species metadata is from \link[ebirdst:ebirdst_runs]{ebirdst_runs}
}
\details{
\code{species()} returns data originally stored in the
\link[ebirdst:ebirst_runs]{ebirdst::ebirst_runs} and uses the column names from that object described
in Format below.
}
