% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_info.R
\name{species_info}
\alias{species_info}
\alias{species}
\title{Retrieve species information from a BirdFlow model}
\usage{
species_info(x, what)

species(x, what)
}
\arguments{
\item{x}{A \code{BirdFlow} model.}

\item{what}{\code{"all"} for all information (in a list) or the name (see details)
of the desired information. If \code{what} is omitted \code{species()} defaults
to \code{"common_name"} and \code{species_info()} defaults to \code{"all"}.}
}
\value{
The element named by \code{what}, unless \code{what} is \code{"all"} in which case a
list of all the species information.  If \code{what} is omitted \code{species()} will
return the common name of the species, and \code{species_info()} will return the
complete list.
}
\description{
\code{species_info()} and \code{species()} return species data stored in a BirdFlow
model. They differ only in that \code{species()} returns the
common name by default so provides a succinct way to get the species name.
}
\details{
The \code{what}
argument takes the column names used in \link[ebirdst:ebirdst_runs]{ebirdst::ebirdst_runs} as input
(descriptions from \pkg{ebirdst}):\tabular{ll}{
   \code{species_code} \tab Six letter eBird code in eBird Taxonomy v2018 \cr
   \code{scientific_name} \tab Scientific name from eBird Taxonomy v2018 \cr
   \code{common_name} \tab English common name from eBird Taxonomy v2018 \cr
   \code{breeding_quality} \tab Breeding season quality \cr
   \code{breeding_start} \tab Breeding season start date \cr
   \code{breeding_end} \tab Breeding season end date \cr
   \code{nonbreeding_quality} \tab Non-breeding season quality \cr
   \code{nonbreeding_start} \tab Non-breeding season start date \cr
   \code{nonbreeding_end} \tab Non-breeding season end date \cr
   \code{postbreeding_migration_quality} \tab Post-breeding season quality \cr
   \code{postbreeding_migration_start} \tab Post-breeding season start date \cr
   \code{postbreeding_migration_end} \tab Post-breeding season end date \cr
   \code{prebreeding_migration_quality} \tab Pre-breeding season quality \cr
   \code{prebreeding_migration_start} \tab Pre-breeding season start date \cr
   \code{prebreeding_migration_end} \tab Pre-breeding season end date \cr
    \tab  \cr
   \code{all} \tab returns the complete list \cr
}


For convenience the following short versions are also accepted:\tabular{ll}{
   \code{code} \tab (\code{species_code}) \cr
   \code{common} \tab (\code{common_name}) \cr
   \code{name} \tab (\code{common_name}) \cr
   \code{scientific} \tab (\code{scientific_name}) \cr
   \code{species} \tab (\code{common_name}) \cr
}
}
\section{Dropped items}{

The 8 variables below are in \link[ebirdst:ebirdst_runs]{ebirdst::ebirdst_runs} but are dropped
from the BirdFlow model and thus can not be retrieved by
\code{species_info()}.

Four variables that track whether the full range is covered by eBird that
must be TRUE for a BirdFlow model to be fit and are then dropped:\tabular{ll}{
   \code{postbreeding_migration_range_modeled} \tab Is the full range modeled? \cr
   \code{prebreeding_migration_range_modeled} \tab Is the full range modeled? \cr
   \code{nonbreeding_range_modeled} \tab Is the full range modeled? \cr
   \code{breeding_range_modeled} \tab Is the full range modeled? \cr
}


\code{resident} is verified to be \code{FALSE} before the model is fit. It and three
related variables are then dropped:\tabular{ll}{
   \code{resident} \tab Classifies this species a resident or a migrant \cr
   \code{resident_quality} \tab Resident quality \cr
   \code{resident_start} \tab For resident species, the year-round start date \cr
   \code{resident_end} \tab For resident species, the year-round end date \cr
}
}

