% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_methods.R
\name{st_bbox.BirdFlow}
\alias{st_bbox.BirdFlow}
\title{Get the bounding box of a BirdFlow model}
\usage{
\method{st_bbox}{BirdFlow}(obj, ...)
}
\arguments{
\item{obj}{A \code{BirdFlow} object.}

\item{...}{Not used.}
}
\value{
An \code{\link[sf:st_bbox]{sf::st_bbox()}} object (class \code{bbox}) with the
coordinate reference system of \code{obj} attached.
}
\description{
\code{st_bbox()} method for \code{BirdFlow} objects. Returns the spatial extent
of the model's raster geometry as an \code{sf} bounding box, with the
model's coordinate reference system attached.
}
\examples{
\donttest{
bf <- BirdFlowModels::amewoo
sf::st_bbox(bf)
}
}
