% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_methods.R
\name{st_crs.BirdFlow}
\alias{st_crs.BirdFlow}
\title{Get the coordinate reference system of a BirdFlow model}
\usage{
\method{st_crs}{BirdFlow}(x, ...)
}
\arguments{
\item{x}{A \code{BirdFlow} object.}

\item{...}{Not used.}
}
\value{
An \code{\link[sf:st_crs]{sf::st_crs()}} object (class \code{crs}) describing the
coordinate reference system of \code{x}.
}
\description{
\code{st_crs()} method for \code{BirdFlow} objects. Returns the coordinate
reference system stored in the model's geometry (\code{x$geom$crs}).
}
\examples{
\donttest{
bf <- BirdFlowModels::amewoo
sf::st_crs(bf)
}
}
