% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppress_specific_warnings.R
\name{suppress_specific_warnings}
\alias{suppress_specific_warnings}
\title{Suppress warnings that match one or more regular expressions}
\usage{
suppress_specific_warnings(x, patterns = NULL)
}
\arguments{
\item{x}{An expression.}

\item{patterns}{One or more patterns to check warning messages against.}
}
\value{
Possibly output from \code{x}
}
\description{
\code{suppress_specific_warnings()} will suppress warnings that match regular
expression patterns that are supplied via
the \code{patterns} argument, without suppressing warnings that don't match the
patterns.
}
\keyword{internal}
