% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackDataClass.R
\name{Track}
\alias{Track}
\alias{check_data_input_format.Track}
\alias{print.Track}
\alias{TrackCollection}
\alias{check_track.TrackCollection}
\alias{add_track.TrackCollection}
\alias{summary.TrackCollection}
\alias{print.TrackCollection}
\title{Create a Track object}
\usage{
Track(
  track_id,
  track_date,
  track_longitude,
  track_latitude,
  track_x,
  track_y,
  track_type = "Undefined",
  track_info = ""
)

check_data_input_format.Track(
  track_id,
  track_date,
  track_longitude,
  track_latitude,
  track_x,
  track_y,
  track_type,
  track_info
)

\method{print}{Track}(track)

TrackCollection(tracks = list())

check_track.TrackCollection(track)

add_track.TrackCollection(track_collection, track)

\method{summary}{TrackCollection}(track_collection)

\method{print}{TrackCollection}(track_collection)
}
\arguments{
\item{track_id}{A string or number representing the unique identifiers of the track.}

\item{track_date}{A vector representing the dates associated with the track.}

\item{track_longitude}{A vector of longitudes for the track points.}

\item{track_latitude}{A vector of latitudes for the track points.}

\item{track_x}{A vector of x-coordinates in a projected coordinate system (projected into BirdFlow CRS).}

\item{track_y}{A vector of y-coordinates in a projected coordinate system (projected into BirdFlow CRS).}

\item{track_type}{A character string indicating the type of track (default: "Undefined").}

\item{track_info}{Additional information or metadata about the track (default: "").}

\item{track}{An object of class "Track" to add to the collection.}

\item{tracks}{a list of objects of class "Track". Default to empty list.}

\item{track_collection}{An object of class "TrackCollection".}
}
\value{
An object of class "Track" and "BirdFlowRoute".

No return value. Prints a summary to the console.

An object of class "TrackCollection" containing a list of tracks (default: empty list).

No return value. Raises an error if the input is not of class "Track".

An updated \code{TrackCollection} object with the new \code{Track} added.

A list containing:
\describe{
\item{\code{track_count_summary}}{The number of tracks in the collection.}
\item{\code{track_type_summary}}{A table summarizing the types of tracks in the collection.}
}

No return value. Prints a summary to the console.
}
\description{
The \code{Track} function creates an object of class "Track" and "BirdFlowRoute".

This function print the track as a summary

The \code{TrackCollection} function creates an object of class "TrackCollection".

This function checks whether the input is a valid object of class "Track".

This function adds a \code{Track} object to a \code{TrackCollection}.

This function summarizes the content of a \code{TrackCollection}, including the
number of tracks and the types of tracks.

This function provides a human-readable summary of a \code{TrackCollection}.
}
