% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transition.R
\name{transition_from_marginal}
\alias{transition_from_marginal}
\title{transition_from_marginal}
\usage{
transition_from_marginal(m, direction)
}
\arguments{
\item{m}{a marginal}

\item{direction}{the desired transition direction, either "forward" or
"backward"}
}
\value{
a transition matrix formulated such that you multiply the matrix by
a distribution to project the distribution.  See \link{get_transition} for more
details.
}
\description{
internal function to generate a transition matrix from a marginal
}
\details{
this is called from get_transition. If at some point we decide
to store transitions rather than marginals it will also be called from
import_birdflow.
}
