% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_info.R
\name{ts_info}
\alias{ts_info}
\title{Internal function to determine timestep sequence direction and whether
it crosses year boundary}
\usage{
ts_info(ts)
}
\arguments{
\item{ts}{A sequence of timesteps}
}
\value{
A list with:
\item{direction}{Either "forward" or "backward"}
\item{loops}{TRUE if the sequence crosses the year boudary, FALSE if it
does not.}
}
\description{
It assumes that all lags between adjacent steps that don't cross the year
boundary are either -1 or 1.
}
\keyword{internal}
