% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_BirdFlow.R
\name{validate_geom}
\alias{validate_geom}
\title{Validate geom component of a BirdFlow or related object}
\usage{
validate_geom(geom, n_active, throw_error = TRUE)
}
\arguments{
\item{geom}{A geom list}

\item{n_active}{The number of active cells in the model
see \code{n_active()} also stored in \code{bf$metadat$n_active}}

\item{throw_error}{If \code{TRUE} throw errors that are
found. IF false return a problem data frame.
throw_error = FALSE is for compatibility with
\code{validate_BirdFlow}}
}
\value{
a problem data frame
}
\description{
This is called from \code{validate_BirdFlow}, as well
as some of the route and interval validation functions
}
\keyword{internal}
