% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_between.R
\name{weight_between}
\alias{weight_between}
\title{Assign weights to points based on how close they are to the great circle
line associated with each transition}
\usage{
weight_between(
  bf,
  weight_fun = NULL,
  points = NULL,
  radius = NULL,
  n_directions = 1,
  skip_unconnected = TRUE,
  batch_size = 1e+05,
  ...
)
}
\arguments{
\item{bf}{A BirdFlow model}

\item{points}{The points to evaluate betweenness on. If NULL the cell
centers of all the raster cells within the BirdFlow model that are between
active cells in the model will be used. This is calculated by comparing the
cell centers to a buffered convex hull around the active cell centers.}

\item{radius}{The probability density along a transect perpendicular to
the line and intersecting the point is summed over the inteval within
\code{radius} meters of the point. \code{radius} defaults to half the cell
size (\code{mean(res(bf))/2}).}

\item{skip_unconnected}{If \code{TRUE} then only connections that exist in \code{bf}
will be evaluated, and between matrix will erroneously indicate that
the weights associated with transitions that aren't used is always 0.
The resulting array can still be used with the model it was built for because
those missing connections would always have zero probability.}

\item{batch_size}{controls the number of movement lines that are processed
at a time. A smaller
\code{batch_size} will conserve memory at a slight performance cost.  The number
of batches will be less than or equal to \code{n_active(bf)^2 / batch_size}.}

\item{n_direction}{The number of (equally spaced) directional bins to
classify bearings into.  Currently only \code{1} is supported.}
}
\value{
A list with:
\item{between}{An array with dimensions representing the
"from" location, the "to" location, and the \code{points}. Cells are weights and
will be non-zero if the radius around the point intersects 1.96 standard
deviations of the normal distribution of probabilities around the great
circle line.}
\item{points}{A data,frame of points that define the third dimension
in \code{between}.  It is identical to the input \code{points} if they are not \code{NULL}.
Otherwise it will be a data frame with columns \code{x}, \code{y}, and \code{i}
corresponding to the third dimension in \code{between}.
\code{i} will be \code{NA} for points that are not within the mask but
fall between active cells.}
\item{radius}{The radius of the circle in meters.}
@seealso \code{\link[=is_between]{is_between()}} and \code{\link[=calc_flux]{calc_flux()}}
}
\description{
This internal function is used to create a betweenness weights array
with dimensions \href{n_active()}{n_active(bf)},
\href{n_active()}{n_active(bf)}, and \code{length(points)}.
The first two dimensions represent from and to cells of possible connections
between  pairs of locations within the BirdFlow model and both have
\code{n_active()} elements.
The third dimension represents reference points that might be between each
connection.
Cell values are the weight to use when adding the transition probabilities
to the reference point flux.
}
\details{
If \code{points}  are \code{NULL} they default to the center of all cells within
the BirdFlow extent that fall between any active cells.
This includes all cell centers within a convex hull
(in spherical coordinates) around the active cells in \code{bf} and thus is
almost always more than just the active cells.

\code{weight_between()} and \code{is_between()} should only differ slightly in
their results when calculated on the same model.

1, The number of included reference points may differ.
2. The betweeness array will be real weights from 0 to 1 with
\code{weight_between()} and logical with \code{is_between()}
}
\keyword{internal}
