% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dynamic_mask.R
\name{add_dynamic_mask}
\alias{add_dynamic_mask}
\title{Convert a BirdFlow object without a dynamic mask into one with a dynamic mask}
\usage{
add_dynamic_mask(bf, dummy_mask = FALSE)
}
\arguments{
\item{bf}{A BirdFlow object}

\item{dummy_mask}{If TRUE a mask is addded to the object, but the mask is
TRUE for every cell. This yields a BirdFlow object that works with the
current mask dependent version of the package but mimics and old BirdFlow
model. Note if the old model included state based sparsification the
predictions should be identical even with \code{dummy_mask = FALSE} (the default).}
}
\value{
A BirdFlow object that has a dynamic_mask component and in which
the marginals only includes transitions between cells that are not
dynamically masked.
}
\description{
A dynamic mask is a logical matrix of the same dimensions as the distr matrix
for \code{bf} holding TRUE for cells where the corresponding location (row) and
time (column) is included in the model.
}
\examples{
bf <- add_dynamic_mask(BirdFlowModels::amewoo)

}
