% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RouteDataFunction.R
\name{add_stay_id_with_varied_intervals}
\alias{add_stay_id_with_varied_intervals}
\title{Add Stay IDs with Temporal Thresholds}
\usage{
add_stay_id_with_varied_intervals(
  df,
  timestep_col = "timestep",
  time_threshold = Inf
)
}
\arguments{
\item{df}{A data frame with spatial and temporal data.}

\item{timestep_col}{The name of the column containing time steps. Defaults to \code{"timestep"}.}

\item{time_threshold}{A numeric threshold for time differences. Defaults to \code{Inf}.}
}
\value{
A data frame with \code{stay_id} and \code{stay_len} columns added.
}
\description{
Adds stay IDs to a data frame, considering changes in spatial indices and temporal thresholds.
Using add_stay_id_with_varied_intervals, rather than add_stay_id: It takes 'timestep' as input so account for varying intervals, if the data is not sampled in the same frequency.
}
\examples{
routes <- data.frame(
  route_id = c(1, 1, 1, 2, 2, 3, 3, 3),
  i = as.integer(c(1, 1, 2, 2, 3, 4, 4, 5)),  # Spatial index
  timestep = as.integer(c(1, 2, 5, 6, 10, 15, 16, 20))  # Time steps with varying intervals
)
df_with_varied_stay_ids <- add_stay_id_with_varied_intervals(routes, "timestep", time_threshold = Inf)
}
