% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_bmtr.R
\name{animate_bmtr}
\alias{animate_bmtr}
\title{Animate BirdFlow Migration Traffic Rate (BMTR)}
\usage{
animate_bmtr(bmtr, bf, title = species(bf), ...)
}
\arguments{
\item{bmtr}{A data frame created by
\link[=calc_bmtr]{calc_bmtr(format = "dataframe")}}

\item{bf}{A BirdFlow object}

\item{title}{The plot title}

\item{...}{
  Arguments passed on to \code{\link[=plot_bmtr]{plot_bmtr}}
  \describe{
    \item{\code{subset}}{A subset of the transitions in \code{bmrt} to plot, can be
a logical vector of the same length as the number of transitions in \code{bmtr};
a numeric index of transitions in \code{bmtr}, or a subset of the transition names
in \code{bmtr}.}
    \item{\code{limits}}{Two numbers representing the range in bmtr values to
display. Values outside of this range will be truncated to the range. With
the default of \code{NULL} the entire range is plotted.}
    \item{\code{dynamic_scale}}{If \code{TRUE} then the range of the data in each
transition is mapped to the color palette. This makes it easier to see
the variation within a single transition but results in an inconsistent
scale among transitions.}
    \item{\code{coast_linewidth}}{The line width used to plot the coast. Set to \code{NULL}
to skip plotting the coastline.}
    \item{\code{coast_color}}{The color used to plot the coastline, or \code{NULL} to skip
plotting the coastline.}
    \item{\code{gradient_colors}}{The colors palette used to represent the BMTR
intensity.}
    \item{\code{value_label}}{The label for the BMTR values.}
    \item{\code{transform}}{A transformation to apply to the color scaling.
\code{"identity"}, and \code{"sqrt"} are recommended.
If \code{"log"} is used zeros will be replaced with
1/2 the smallest non-zero value prior to transforming.
Legend will still reflect the original values.
Passed to \code{\link[ggplot2:scale_gradient]{ggplot2::scale_color_gradient()}}.}
    \item{\code{crop_bbox}}{An optional named numeric vector with elements \code{xmin},
\code{xmax}, \code{ymin}, \code{ymax} in the model CRS to zoom the plot to a specific
region.  See also \code{sf::st_bbox()} which generates a suitable object.
When \code{NULL} (default) the full model extent is shown.}
  }}
}
\value{
A \code{\link[gganimate:gganimate-package]{gganim}} object
}
\description{
Animate migration traffic rates produced by \code{\link[=calc_bmtr]{calc_bmtr()}}.
}
\examples{

\dontrun{
bf <- BirdFlowModels::amewoo

# Binary detection along shortest path
bmtr <- calc_bmtr(bf)
plot_bmtr(bmtr, bf)

animate_bmtr(bmtr, bf)

# Continuous detection with a wider spread kernel
# (gamma defaults to 3e10)
bmtr2 <- calc_bmtr(bf, method = "continuous", gamma = 6e10)
animate_bmtr(bmtr2, bf)

# Visualize spread for the continuous detection with wider spread
visualize_distance_weights(line_lengths = c(2, 5, 10) * xres(bf),
                           gamma = 6e10, res_m = xres(bf))

# Visualize spread for the continuous detection with default values
visualize_distance_weights(line_lengths = c(2, 5, 10) * xres(bf),
                           res_m = xres(bf))

}




}
\seealso{
\code{\link[=calc_bmtr]{calc_bmtr()}},\code{\link[=plot_bmtr]{plot_bmtr()}}
}
