% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_distr.R
\name{animate_distr}
\alias{animate_distr}
\title{Animate distributions}
\usage{
animate_distr(distr, bf, title = species(bf), ...)
}
\arguments{
\item{distr}{A set of distributions it should be a matrix with \code{n_active(bf)} rows
and a column for each distribution. The animation will proceed in the order}

\item{bf}{The BirdFlow object that \code{distr} is associated with.}

\item{title}{The title to use for the animation. The defualt is
the common name of the species.}

\item{...}{
  Arguments passed on to \code{\link[=plot_distr]{plot_distr}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
A gganimate object that can be displayed with \code{print()} or
or \code{gganimate::animate()}.  See example for how to export to a file.
}
\description{
Animate distributions as produced by \code{get_distr()},
\code{as_distr()}, or \code{predict()}.  The distributions will be displayed
in the column order in \code{distr} and column labels will be used as
plot subtitles.
}
\examples{

bf <- BirdFlowModels::amewoo
distr <- get_distr(bf, lookup_timestep_sequence(bf, season = "prebreeding"))
anim <- animate_distr(distr, bf)

\dontrun{
  # example render
  gif <- gganimate::animate(anim,
                            device = "ragg_png", # ragg_png is fast and pretty
                            width = 7, height = 6,
                            res = 150, units = "in")
  # Display
  print(gif)

  # Save
  gif_file <- tempfile("animation", fileext = ".gif")
  gganimate::save_animation(gif, gif_file)
  file.remove(gif_file) # cleanup
  }
}
