% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_bmtr.R
\name{animate_flux}
\alias{animate_flux}
\title{Animate Bird Flow Migration Traffic Rate (BMTR)}
\usage{
animate_flux(flux, ...)
}
\arguments{
\item{flux}{the output from \code{\link[=calc_bmtr]{calc_bmtr()}} or, deprecated, \code{\link[=calc_flux]{calc_flux()}}}

\item{...}{
  Arguments passed on to \code{\link[=animate_bmtr]{animate_bmtr}}
  \describe{
    \item{\code{bf}}{A BirdFlow object}
    \item{\code{title}}{The plot title}
  }}
}
\value{
A \code{\link[gganimate:gganimate-package]{gganim}} object
}
\description{
DEPRECATED FUNCTION.  Please use \code{\link[=animate_bmtr]{animate_bmtr()}} instead.
}
