% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_routes.R
\name{animate_routes}
\alias{animate_routes}
\title{Animate routes}
\usage{
animate_routes(routes, bf, ...)
}
\arguments{
\item{routes}{The output of \code{\link[=route]{route()}} or a similarly structured data frame.}

\item{bf}{A BirdFlow object}

\item{...}{
  Arguments passed on to \code{\link[=plot_routes]{plot_routes}}
  \describe{
    \item{\code{facet}}{If \code{TRUE} then use \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} to show each route
out into a separate subplot.}
    \item{\code{stay_units}}{The unit to plot the stay length at each location. Default
to \code{days}. Other options include \code{sec}, \code{mins}, \code{hours} and \code{weeks}.}
    \item{\code{max_stay_len}}{Used to scale the stay length dots. If \code{NULL}
(the default) it will be set to the maximum \code{"stay_len"} value in \code{routes}.
Set it manually to keep the dot scaling consistent across multiple plots.}
    \item{\code{use_seasonal_colors}}{If \code{TRUE} a color scale that uses blues, greens,
yellows, reds, for winter, spring, summer, and fall will be used with a
consistent mapping of dates to colors regardless of the range of dates
plotted. If \code{FALSE} then the data will be plotted using the full color scale.}
    \item{\code{pal}}{The color palette to use for plotting when \code{use_seasonal_cols} is
\code{FALSE}. Defaults to \link[viridisLite:viridis]{viridisLite::viridis(n = 5)}.}
    \item{\code{barheight}}{The height of the color gradient legend bar. Passed to
\code{\link[ggplot2:guide_colourbar]{ggplot2::guide_colorbar()}} as \code{barheight} argument. Depending on the output
resolution and plot size this may need to be adjusted. Can take a number or
the output from \code{\link[ggplot2:reexports]{ggplot2::unit()}}.}
    \item{\code{route_linewidth}}{Line width used for routes.}
    \item{\code{coast_linewidth}}{Line width used for coastlines.}
    \item{\code{dot_sizes}}{Two numbers indicating the smallest and largest dot sizes
used to represent stay length.}
  }}
}
\value{
A \code{gganim} object. \code{print()} will plot it with default
options, or use \code{\link[=animate]{animate()}} to set the options. See the example for
recommended settings.
}
\description{
Animate synthetic routes produced by \code{\link[=route]{route()}} and \code{\link[=route_migration]{route_migration()}},
producing a dynamic version of \code{\link[=plot_routes]{plot_routes()}}.
}
\details{
Note when rendering early frames (at a minimum the first) there will only
one point per route, resulting in a message:
"geom_path(): Each group consists of only one observation.
ℹ Do you need to adjust the group aesthetic?"
This will possibly be repeated while individuals remain in one location.
It can be safely ignored. The error is thrown while rendering and not
from within \code{animate_routes()} where it could not be suppressed."
}
\examples{


bf <- BirdFlowModels::amewoo
rts <- route(bf, 10,  season = "prebreeding")
anim <- animate_routes(rts, bf)

\dontrun{
  # example render
  timesteps <- unique(rts$timestep)
  gif <- gganimate::animate(anim,
                            device = "ragg_png", # is fast and pretty
                            width = 7, height = 6,
                            res = 150, units = "in",
                            nframes = length(timesteps) * 4, fps = 8)

  # Display
  print(gif)

  # Save
  gif_file <- tempfile("animation", fileext = ".gif")
  gganimate::save_animation(gif, gif_file)
  file.remove(gif_file) # cleanup
}

}
