% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RouteDataFunction.R
\name{as_BirdFlowIntervals}
\alias{as_BirdFlowIntervals}
\title{Convert BirdFlowRoutes to BirdFlowIntervals}
\usage{
as_BirdFlowIntervals(birdflow_routes, max_n = 1000)
}
\arguments{
\item{birdflow_routes}{A \code{BirdFlowRoutes} object.}

\item{max_n}{The maximum number of intervals to sample. Defaults to 1000.}
}
\value{
A \code{BirdFlowIntervals} object.
}
\description{
Converts a \code{BirdFlowRoutes} object into a \code{BirdFlowIntervals} object,
sampling interval pairs between time points.
}
\examples{
route_df <- data.frame(
  route_id = 1:3,
  date = as.Date(c("2024-01-01", "2024-01-02", "2024-01-03")),
  lon = c(-90, -89, -88),
  lat = c(40, 41, 42),
  route_type = c("tracking", "banding", "unknown")
)
routes_obj <- Routes(route_df)
bf <- BirdFlowModels::amewoo
birdflow_routes <- routes_obj |> as_BirdFlowRoutes(bf=bf)
birdflow_intervals <- as_BirdFlowIntervals(birdflow_routes, max_n = 1000)
}
