% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RouteDataFunction.R
\name{as_BirdFlowIntervals}
\alias{as_BirdFlowIntervals}
\title{Convert BirdFlowRoutes to BirdFlowIntervals}
\usage{
as_BirdFlowIntervals(birdflow_routes, max_n = 1000)
}
\arguments{
\item{birdflow_routes}{A \code{BirdFlowRoutes} object.}

\item{max_n}{The maximum number of intervals to sample. Defaults to 1000.}
}
\value{
A \code{BirdFlowIntervals} object.
}
\description{
Converts a \code{BirdFlowRoutes} object into a \code{BirdFlowIntervals} object,
sampling interval pairs between time points.
}
\examples{
birdflow_intervals <- as_BirdFlowIntervals(birdflow_routes, n = 1000)
}
