% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_BirdFlowIntervals.R
\name{as_BirdFlowIntervals}
\alias{as_BirdFlowIntervals}
\title{Convert BirdFlowRoutes to BirdFlowIntervals}
\usage{
as_BirdFlowIntervals(
  birdflow_routes,
  max_n = 1000,
  min_day_interval = 7,
  max_day_interval = 180,
  min_km_interval = 200,
  max_km_interval = 8000
)
}
\arguments{
\item{birdflow_routes}{A \code{BirdFlowRoutes} object.}

\item{max_n}{The maximum number of intervals to sample.
Defaults to 1000.}

\item{min_day_interval}{The minimum days required in an interval.
Defaults to 7.}

\item{max_day_interval}{The maximum days required in an interval.
Defaults to 180.}

\item{min_km_interval}{The minimum distance required for an interval.
Defaults to 200.}

\item{max_km_interval}{The maximum distance required for an interval.
Defaults to 2000.}
}
\value{
A \code{BirdFlowIntervals} object.
}
\description{
Converts a \code{BirdFlowRoutes} object into a \code{BirdFlowIntervals}
object,
sampling interval pairs between time points.
}
\examples{
route_df <- data.frame(
 route_id = c("001", "001", "001", "001", "001", "003", "003", "003", "004"),
 date = as.Date(c("2025-01-01", "2025-01-08", "2025-01-15",
 "2025-01-21", "2025-02-10", "2025-03-01", "2025-05-01", "2025-06-01", 
 "2025-05-01")),
 lon = c(-75.0060, -75.0060, -74.0060, -87.6298, -87.6298, -87.6298, 
 -89.6298, -85.6298, -95.3698),
 lat = c(39.7128, 39.7128, 40.7128, 41.8781, 41.8781, 41.8781, 42.8781, 
 40.8781, 29.7604),
 route_type = c("tracking", 'tracking', "tracking", 'tracking', 'tracking',
 "motus", "motus", "motus", "motus")
)
routes_obj <- Routes(route_df, species = "amewoo")
bf <- BirdFlowModels::amewoo
birdflow_routes <- routes_obj |> as_BirdFlowRoutes(bf=bf)
birdflow_intervals <- as_BirdFlowIntervals(birdflow_routes, max_n = 1000)
}
