% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RouteDataFunction.R
\name{as_BirdFlowRoutes}
\alias{as_BirdFlowRoutes}
\title{Convert Routes to BirdFlowRoutes}
\usage{
as_BirdFlowRoutes(
  routes,
  bf,
  valid_only = TRUE,
  sort_id_and_dates = TRUE,
  reset_index = FALSE
)
}
\arguments{
\item{routes}{A \code{Routes} object.}

\item{bf}{A \code{BirdFlow} object for spatial and temporal reference.}

\item{valid_only}{Logical. Should only valid points be included? Defaults to \code{TRUE}.}

\item{sort_id_and_dates}{Logical. Should data be sorted by route ID and date? Defaults to \code{TRUE}.}

\item{reset_index}{Logical. Should indices be reset after sorting? Defaults to \code{FALSE}.}
}
\value{
A \code{BirdFlowRoutes} object.
}
\description{
Converts a \code{Routes} object to a \code{BirdFlowRoutes} object, adding BirdFlow-specific spatiotemporal coordinates.
}
\examples{
birdflow_routes <- as_BirdFlowRoutes(routes, bf)
}
