% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Routes.R
\name{as_routes}
\alias{as_routes}
\title{Convert a data frame into a \code{Routes} object}
\usage{
as_routes(data, species = NULL, source = NULL)
}
\arguments{
\item{data}{A data frame with data on bird movement. Likely
tracking, banding, or Motus data on real birds; or possibly
synthetic versions of the same.It must have the following columns:
\describe{
\item{\code{date}}{Date or Date Time object of class \code{Date}, \code{POSIXlt},
or \code{POSIXct}}
\item{\code{lat},\code{lon}}{The latitude and longitude of the location in
WGS84 (EPSG::4326)}
\item{\code{route_type}}{The type of route. Standard values are
\code{"tracking"}, \code{"banding"}, \code{"motus"}, \code{"unknown"}, and \code{"synthetic"},
but any character value is accepted; non-standard values produce a
warning. Types can be mixed in the column.
}
}
Other columns are permitted and will be retained in \code{Routes} object
but dropped if they are converted to \code{BirdFlowRoutes}.}

\item{species}{Either: a character scalar suitable for use with
\code{\link[ebirdst:get_species]{ebirdst::get_species()}}; or a list with
species metadata which must include \code{common_name} and can optionally
also include \code{scientific_name} and  \code{species_code} and any other standard
BirdFlow species metadata. See \code{\link[=species_info]{species_info()}} for a description
of the full list. Note list input is not checked against eBird species
codes and names. Scalar input is preferred unless the species does not
conform to eBird's taxonomy.}

\item{source}{Optional text describing the source of the data.
\code{source} must be of class \code{character} can have one or more elements.}
}
\value{
An object of class \code{Routes} which has the following components
\item{data}{A data frame with the input \code{data}}
\item{species}{A list with, at a minimum items
\code{common_name},  \code{scientific_name}, and \code{species_code} and depending on
the \code{species} argument potentially having all the items returned by
\code{\link[=lookup_species_metadata]{lookup_species_metadata()}}}
\item{source}{Same as the input \code{source}}
}
\description{
\code{as_routes()} converts data frames containing bird movement
data into a formal \code{Routes} object with the same data.
}
\examples{
\donttest{
df <- data.frame(
  route_id = c(1, 1, 2, 2),
  date = as.Date(c("2021-01-04", "2021-01-11",
                    "2021-01-04", "2021-01-11")),
  lon = c(-87.6, -87.2, -85.0, -84.6),
  lat = c(35.8, 36.0, 33.5, 33.8),
  route_type = "tracking"
)
rts <- as_routes(df, species = "amewoo")
rts
}
}
\seealso{
\code{\link[=as_birdflow_routes]{as_birdflow_routes()}} for converting \code{Routes} to
\code{BirdFlowRoutes}.
}
