% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_transitions.R
\name{as_transitions}
\alias{as_transitions}
\title{Convert a sequence of timesteps into a sequence of transition names}
\usage{
as_transitions(timesteps, bf)
}
\arguments{
\item{timesteps}{A valid sequence of timesteps. See
\code{lookup_timestep_sequence()}.}

\item{bf}{A BirdFlowR model, used to determine how much padding
is needed around the timesteps.}
}
\value{
A directional sequence of transitions that connect \code{timesteps}
}
\description{
Convert a sequence of timesteps into a sequence of transition names
}
