% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdflow_crs.R
\docType{data}
\name{birdflow_crs}
\alias{birdflow_crs}
\title{Define the coordinate reference system}
\format{
This is a string defining a custom Mollweide projection
centered on the western hemisphere with well known text.
}
\usage{
birdflow_crs
}
\description{
This defines the coordinate reference system (CRS, AKA projection)
used by default for BirdFlow models. It is a customized
Mollweide projection, with the longitude of origin set to -90 deg.
centering the western hemisphere.  St Louis, Missouri is very close to
a longitude of -90 deg. The Mollweide projection preserves area, shape
distortion increases with distance from the longitude of origin.
}
\details{
The projection is similar to:
\href{https://spatialreference.org/ref/sr-org/7399/}{SR-ORG:7399} and
\href{https://epsg.io/54009}{ESRI:54009} which
have a longitude of origin of 0, centered on Greenwich, England.
}
\examples{
cat(birdflow_crs)
crs(birdflow_crs, proj = TRUE)

}
\references{
https://epsg.io/54009
}
\keyword{datasets}
