% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_transitions.R
\name{build_transitions}
\alias{build_transitions}
\title{add transition matrices to a BirdFlow object}
\usage{
build_transitions(x)
}
\arguments{
\item{x}{BirdFlow object}
}
\value{
BirdFlow object with transition matrices
}
\description{
Given a BirdFlow object with marginals and without transitions return
a BirdFlow object with both marginals and transitions.
}
\examples{
\dontrun{
bf1 <- BirdFlowModels::amewoo
bf2 <- build_transitions(bf)
}
}
