% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_distance_weights.R
\name{calc_dist_weights}
\alias{calc_dist_weights}
\title{calculate the weights of transitions for flux points}
\usage{
calc_dist_weights(
  dist_to_line,
  dist_along_line,
  line_lengths,
  radius_m,
  res_m,
  method = "m3"
)
}
\arguments{
\item{dist_to_line}{How far is the point from the line (m)}

\item{dist_along_line}{How far along the line is the point, after
projecting it  onto the line (m)}

\item{line_lengths}{How long is the line (m)}

\item{radius_m}{The radius of the transect at the flux points - used to
determine the band of brobability desnity that will be added to form
the weight.}

\item{res_m}{The resolution of the associated bird flow model, used to
determine the nugget added to the variance to represent the uncertainty in
the starting and ending location of the transition.}

\item{method}{The method used for calculating the standard deviation
in the probability distribution.  Currently \code{"m3"}, Martern 3/2; and
\code{"bb"}, brownian bridge are supported.}
}
\value{
A vector of weights of the same length as the first three arguments.
}
\description{
\code{calc_dist_weights()} is an internal function that takes summary stats
on the relationship between points and a transition line and returns
the weight that should be used for that transition.
}
\details{
The first three arguments can all be vectors in which case the calculations
will be vectorized over the corresponding elements.

This is a preliminary version of the function and will likely change.
}
\keyword{internal}
