% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_flux.R
\name{calc_flux}
\alias{calc_flux}
\title{Estimate bird flux}
\usage{
calc_flux(
  bf,
  points = NULL,
  radius = NULL,
  n_directions = 1,
  format = NULL,
  batch_size = 5e+05,
  check_radius = TRUE,
  weighted = FALSE
)
}
\arguments{
\item{bf}{A BirdFlow model}

\item{points}{A set of points to calculate movement through. If \code{points} is
\code{NULL} they will default to the BirdFlow model cells that are either active
or fall between two active cells. Otherwise a data frame with \code{x} and  \code{y}
columns containing point coordinates in \link[=crs]{crs(bf)}.}

\item{radius}{The radius in meters around the points used to assess whether
a movement line passes by (or through) the point. If a point is farther than
\code{radius} from a great circle line between two cells centers then it is not
between them.}

\item{n_directions}{The number of directional bins to use for recording
movement direction. Must be either \code{1} indicating no direction information
or an even number. This is a placeholder, currently only \code{1} is supported.}

\item{format}{The format to return the results in one of:
\describe{
\item{\code{"points"}}{Returns a list with \code{flux} a matrix or array of
flux values, and \code{points} a data frame of either the input \code{points} or the
default cell center derived points.}
\item{\code{"dataframe"}}{Returns a "long" data frame with columns:
\itemize{
\item \code{x} and \code{y} coordinates of the points.
\item \code{transition} Transition code.
\item \code{flux} The flux at the point. See "Units" below  .
\item \code{date} The date associated with the transition, will be at the midpoint
between timesteps.
}

}
\item{\code{"SpatRaster"}}{Returns a \code{terra::SpatRaster} with layers for each
transition.}
}}

\item{batch_size}{controls the number of movement lines that are processed
at a time. A smaller
\code{batch_size} will conserve memory at a slight performance cost.  The number
of batches will be less than or equal to \code{n_active(bf)^2 / batch_size}.}

\item{check_radius}{If  \code{TRUE} an error will be thrown if the radius
is not between the resolution and 1/4 the resolution of \code{bf}. Outside of
that range the algorithm is likely to yield distorted results.
\code{0.5 * mean(res(bf))} is the default, and recommended radius.}

\item{weighted}{If \code{FALSE} use the original and quicker version of flux
that sums all the marginal probability for transitions that pass within a
fixed distance of the point.  If \code{TRUE} assign a weight to the point and
transition combo that then is multiplied by the marginal probability before
summing.  This argument is experimental but the default value is identical
to the old version. The argument name and behavior when set to \code{TRUE} may
change.}
}
\value{
See \code{format} argument.
}
\description{
\code{calc_flux()} estimates the proportion of the species that passes near
a set of points during each transition in a BirdFlow model.
}
\section{Units}{


The total relative abundance passing through the circle centered on the point
is divided by the diameter of the circle in kilometers.  The units of the
returned value is therefore roughly the proportion
(\eqn{P}) of the
species's population that is expected to pass through each km
of a line oriented perpendicular to the movement at each point:
\eqn{\frac{P}{km \cdot week}}

Multiplying the result by the total population would yield:
\eqn{\frac{birds}{km \cdot week}}
}

\section{Limitations}{


\code{calc_flux()} makes the incorrect simplifying assumption
that birds follow the shortest (great circle) path
between the center of the the source and destination raster cells.  Caution
should be used when interpreting the results especially around
major geographic features such as coasts, large lakes, mountain ranges, and
ecological system boundaries that might result in non-linear migration paths.

\code{calc_flux()} assumes that a line passes by a point if any part of the line
is within the radius of the point.  This assumption breaks down if the
radius is much larger than the movement lengths as points that are ahead of
the line may still be within a radius of the line.  In the extreme a large
enough radius on a point outside of the entire range will capture all the
movement.  This isn't a problem with
the default points and radius as the points ahead of the line will never be
within the radius.

The default points for \code{calc_flux()} are aligned with the cell centers as
are the movement lines. This alignment means that a very small radius will
result in an overestimate of flux. The default value of half the cell size
is sufficient for this not to be a problem, as we are capturing and
standardizing the units based on the entire cell area that that point
represents.
}

\examples{

\dontrun{
bf <- BirdFlowModels::amewoo
flux <- calc_flux(bf)

plot_flux(flux, bf)

animate_flux(flux, bf)
}

}
