% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_flux.R
\name{calc_flux}
\alias{calc_flux}
\title{Estimate net bird movement}
\usage{
calc_flux(bf, points = NULL, radius = NULL, n_directions = 1, format = NULL)
}
\arguments{
\item{bf}{A BirdFlow model}

\item{points}{A set of points to calculate movement through. If \code{points} is
\code{NULL} they will default to the BirdFlow model cells that are either active
or fall between two active cells. Should be a data frame with \code{x} and  \code{y}
columns containing point coordinates in \link{crs(bf)}.}

\item{radius}{The radius in meters around the points, used to determine if
a path is passing through the point.}

\item{n_directions}{The number of directional bins to use for recording
movement direction. Must be either \code{1} indicating no direction information
or an even number.}

\item{format}{The format to return the results in one of:
\describe{
\item{\code{"points"}}{Returns a list with \code{net_movement} a matrix or array of
movement, and \code{points} a data frame of either the input    \code{points} or the
point locations }
\item{\code{"dataframe"}}{Returns a "long" data frame with columns:
*\code{x} and \code{y} coordinates of the points.
\itemize{
\item \code{transition} Transition code.
\item \code{movement} The total abundance moving through the point in the transition.
\item \code{date} The date associated with the transition, will be at the midpoint
between timesteps.}
\item{\code{"SpatRaster"}}{Returns a \code{terra::SpatRaster} with layers for each
transition.
}
}
}}
}
\value{
See \code{format} argument.
}
\description{
\code{calc_flux()} estimates the proportion of the species that passes through
a set of points during each transition in a BirdFlow model.
}
\examples{

\dontrun{
bf <- BirdFlowModels::amewoo
flux <- calc_flux(bf)

plot_flux(flux, bf)

animate_flux(flux, bf)
}

}
