% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bmtr.R
\name{calc_flux}
\alias{calc_flux}
\title{Calculate Bird Flow Migration Traffic Rate}
\usage{
calc_flux(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=calc_bmtr]{calc_bmtr}}
  \describe{
    \item{\code{bf}}{A BirdFlow model}
    \item{\code{points}}{A set of points to calculate movement through. If \code{points} is
\code{NULL} they will default to the BirdFlow model cells that are either active
or fall between two active cells. Otherwise a data frame with \code{x} and  \code{y}
columns containing point coordinates in \link[terra:crs]{crs(bf)}.}
    \item{\code{radius}}{The radius in meters around the points used to assess whether
a movement line passes by (or through) the point. If a point is farther than
\code{radius} from a great circle line between two cells centers then it is not
between them.}
    \item{\code{n_directions}}{The number of directional bins to use for recording
movement direction. Must be either \code{1} indicating no direction information
or an even number. This is a placeholder, currently only \code{1} is supported.}
    \item{\code{format}}{The format to return the results in one of:
\describe{
\item{\code{"points"}}{Returns a list with \code{bmtr} a matrix or array of
bmtr values, and \code{points} a data frame of either the input \code{points} or the
default cell center derived points.}
\item{\code{"dataframe"}}{Returns a "long" data frame with columns:
\itemize{
\item \code{x} and \code{y} coordinates of the points.
\item \code{transition} Transition code.
\item \code{bmtr} The bmtr at the point. See "Units" below  .
\item \code{date} The date associated with the transition, will be at the midpoint
between timesteps.
}

}
\item{\code{"SpatRaster"}}{Returns a \code{terra::SpatRaster} with layers for each
transition.}
}}
    \item{\code{weighted}}{If \code{FALSE} use the original and quicker version of bmtr
that sums all the marginal probability for transitions that pass within a
fixed distance of the point.  If \code{TRUE} assign a weight to the point and
transition combo that then is multiplied by the marginal probability before
summing.  This argument is experimental but the default value is identical
to the old version. The argument name and behavior when set to \code{TRUE} may
change.}
    \item{\code{batch_size}}{controls the number of movement lines that are processed
at a time. A smaller
\code{batch_size} will conserve memory at a slight performance cost.  The number
of batches will be less than or equal to \code{n_active(bf)^2 / batch_size}.}
    \item{\code{check_radius}}{If  \code{TRUE} an error will be thrown if the radius
is not between the resolution and 1/4 the resolution of \code{bf}. Outside of
that range the algorithm is likely to yield distorted results.
\code{0.5 * mean(res(bf))} is the default, and recommended radius.}
  }}
}
\value{
See \code{format} argument.
}
\description{
DEPRECATED FUNCTION.  Please use \code{\link[=calc_bmtr]{calc_bmtr()}} instead.
}
