% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_movement_vectors.R
\name{calc_movement_vectors}
\alias{calc_movement_vectors}
\title{Calculate the average movement for cells in a BirdFlow model}
\usage{
calc_movement_vectors(bf, start, direction = "forward")
}
\arguments{
\item{bf}{A BirdFlow model}

\item{start}{The starting timestep for the transition to be modeled}

\item{direction}{"Forward" by default. Set to "backward" to calculate
vectors for a transition backwards in time.}
}
\value{
A data frame with columns:
\item{i}{The location index of the starting location.  This is the row in
the distribution matrix that corresponds to the location.}
\item{start}{The starting timestep of the transition.}
\item{end}{The ending timestep of the transition.}
\item{start_x}{The x coordinate of the starting cell of the transition.}
\item{start_y}{The y coordinate of the starting cell of the transition.}
\item{end_x}{The weighted average destination x coordinate of all transitions
from the starting cell, with weights set to the transition probability for
each destination.}
\item{end_y}{The weighted average destination y coordinate. }
\item{weight}{This is the proportion of the population at the starting cell
in the eBird S&T distribution for the starting timestep.}
\item{width}{This is a range rescaling of weight and is used by both
\code{\link[=plot_movement_vectors]{plot_movement_vectors()}}, and \code{\link[=animate_movement_vectors]{animate_movement_vectors()}} to set the line
width of the arrows.}
}
\description{
\code{calc_movement_vectors()} calculates the average modeled movement of birds
from each location over a single transition in a BirdFlow model.
}
\details{
In practice each row of the transition matrix represents a single starting
location (\code{start_x} and \code{start_y}) and the values in the row represent
transition probability to a number of destination cells. The probabilities
are used as weights to calculate an average of both the destination x and y
coordinates (\code{end_x}, and \code{end_y}).  The ending coordinates of the arrow
represented the expected average destination of all birds that start
in the cell.
}
\examples{
bf <- BirdFlowModels::amewoo
mv <- calc_movement_vectors(bf, 7)

}
\seealso{
\code{\link[=plot_movement_vectors]{plot_movement_vectors()}} and \code{\link[=animate_movement_vectors]{animate_movement_vectors()}} call
this function and visualize the results.
}
