% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_based_validation.R
\name{calculate_interval_metrics}
\alias{calculate_interval_metrics}
\title{Calculate interval metrics}
\usage{
calculate_interval_metrics(birdflow_intervals, bf)
}
\arguments{
\item{birdflow_intervals}{A BirdFlowIntervals object}

\item{bf}{BF model}
}
\value{
mean metrics across transition pairs
}
\description{
Calculate interval based metrics, including distance metrics
for all transition pairs.
}
