% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_interval_sampling_strategy.R
\name{calculate_interval_sampling_strategy}
\alias{calculate_interval_sampling_strategy}
\title{Calculate Interval Sampling Strategy}
\usage{
calculate_interval_sampling_strategy(
  routes,
  n,
  min_day_interval,
  max_day_interval,
  min_km_interval,
  max_km_interval
)
}
\arguments{
\item{routes}{A data frame similar to the data feature in \code{Routes} -- with columns \code{route_id}, \code{date}, \code{lon} and \code{lat}.}

\item{n}{The total maximum number of intervals to sample. Notice: The actual output of intervals might be less than n, because of data deficiency. But never larger than n.}

\item{min_day_interval}{The minimum days required in an interval.}

\item{max_day_interval}{The maximum days required in an interval.}

\item{min_km_interval}{The minimum distance required for an interval}

\item{max_km_interval}{The maximum distance required for an interval}
}
\value{
A data frame with the columns:
\itemize{
\item \code{route_id}: The route ID.
\item \code{time_points}: The number of time points in the route.
\item \code{interval_pairs}: The total number of possible interval pairs for the route.
\item \code{intervals_to_sample}: The number of intervals to sample for the route.
}
}
\description{
Internal function used by \code{as_BirdFlowRoutes()} to determine
how many intervals to sample from each route based on
the total number of intervals requested.
Ensures an even distribution across routes when possible.
}
\examples{
\dontrun{
# Example usage
routes <- data.frame(route_id = c("A", "A", "B", "B", "B"),
                      lon=c(100,101,102,103,104),
                      lat=c(40,42,44,46,48),
                     date = as.Date("2024-01-01") + (0:4)*10)
sampling_strategy <- calculate_interval_sampling_strategy(routes, n = 10,
min_day_interval=20, min_km_interval=100)
}
}
\keyword{internal}
