% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareGeom.R
\name{compareGeom}
\alias{compareGeom}
\alias{compareGeom-BirdFlow}
\alias{compareGeom,BirdFlow,BirdFlow-method}
\alias{compareGeom,SpatRaster,BirdFlow-method}
\alias{compareGeom,BirdFlow,SpatRaster-method}
\title{BirdFlow compareGeom methods}
\usage{
\S4method{compareGeom}{BirdFlow,BirdFlow}(x, y, ...)

\S4method{compareGeom}{SpatRaster,BirdFlow}(x, y, ...)

\S4method{compareGeom}{BirdFlow,SpatRaster}(x, y, ...)
}
\arguments{
\item{x}{A BirdFlow or SpatRaster object}

\item{y}{A BirdFlow or SpatRaster object}

\item{...}{
  Arguments passed on to \code{\link[terra:compareGeom]{terra::compareGeom}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
These are methods for \code{\link[terra:compareGeom]{terra::compareGeom()}} that work when one or both
of arguments are BirdFlow objects.
}
