% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{dimensions}
\alias{dimensions}
\alias{nrow.BirdFlow}
\alias{nrow}
\alias{ncol}
\alias{dim}
\alias{ext}
\alias{ncol.BirdFlow}
\alias{dim.BirdFlow}
\alias{n_timesteps}
\alias{n_transitions}
\alias{n_active}
\title{Dimensions of a BirdFlow object}
\usage{
\method{nrow}{BirdFlow}(x)

\method{ncol}{BirdFlow}(x)

\method{dim}{BirdFlow}(x)

n_timesteps(x)

n_transitions(x)

n_active(x)
}
\arguments{
\item{x}{A BirdFlow object}
}
\value{
\code{nrow()} returns the number of rows in the raster extent of the BirdFlow model

\code{ncol()} returns the number of columns in the raster extent of a BirdFlow model

\code{dim()} returns the number of rows and columns in the raster extent of a BirdFlow model

\code{n_timesteps()} returns the number of timesteps in a BirdFlow model

\code{n_transitions()} returns the number of transitions in a BirdFlow model

\code{n_active()} returns the number of active cells (locations)
in a BirdFlow model.
}
\description{
Functions to return BirdFlow model dimensions
}
