% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_performance.R
\name{distribution_performance}
\alias{distribution_performance}
\title{Evaluate a BirdFlow models ability to recreate training distributions}
\usage{
distribution_performance(x, metrics = NULL, ...)
}
\arguments{
\item{x}{A BirdFlow object}

\item{metrics}{If NULL calculate all metrics.  Otherwise set to a subset of
the metric names to calculate only those metrics.}

\item{...}{
  Arguments passed on to \code{\link[=lookup_timestep_sequence]{lookup_timestep_sequence}}
  \describe{
    \item{\code{start}}{The starting points in time specified as
timesteps, character dates, or date objects; or  may be set to "all" or a
season name to be interpreted by \code{\link[=lookup_season_timesteps]{lookup_season_timesteps()}}.}
    \item{\code{end}}{If start is a timestep or date \code{end} should be a timestep or date
indicating the ending point in time.}
    \item{\code{direction}}{Either "forward" or "backward".

If \code{start} and \code{end}
represent dates and \code{direction} is used an error will
be thrown if \code{direction} isn't consistent with direction indicated by the
dates.

If \code{start} and \code{end} are not dates, \code{direction} defaults to "forward" and
\code{start} and \code{end} should either both be timesteps (numeric); or
\code{end} should be omitted and start should be "all" or a season name.

If \code{start} is a timestep or date, \code{end} is omitted, and \code{n} is an integer
than \code{direction} defaults to "forward".}
    \item{\code{season_buffer}}{Only used if \code{start} is a season. \code{season_buffer} is
passed to \code{\link[=lookup_season_timesteps]{lookup_season_timesteps()}} and defaults to 1; it is the number
of timesteps to extend the season by at each end.}
  }}
}
\value{
\describe{
\item{mean_step_cor}{Indicating on average how well the model projects a
single timestep, \code{mean_step_cor} is the mean correlation, across all
timesteps, between the
training (eBird Status and Trends) distribution projected forward one
timestep, and the training distribution for that projected timestep.
}
\item{min_step_cor}{Indicates the quality of the worst single step
projection. The minimum correlation (across all timesteps) between
a single step projection of the training distribution and the training
distribution for the projected timestep.}

\item{mean_distr_cor}{Indicates on average how well the marginal preserves
the training distributions. The mean correlation between the training
distributions and distributions calculated from the marginals.}
\item{min_distr_cor}{Indicates how well the poorest marginal preserves the
training distribution. The minimum observed correlation between a marginal and
training distribution.}
\item{st_traverse_cor, md_traverse_cor}{Indicates how well the model
projects a distribution through multiple timesteps. They are the correlation
between last distributionin a series projected iteratively forward from the
fist distribution; and the ebird Status and Trends (training) distribution
for the last timestep. \code{st_traverse_cor} starts with the first timestep
Status and Trends (st) distribution, while \code{md_traverse_cor} starts with a
the marginal distribution for the same timestep;  both versions compare the
projected distribution to the eBird Status and Trends distribution for same
timestep.}

}
}
\description{
Calculate the correlation between model derived distributions and the eBird
Status and Trend distributions used to train the BirdFlow model.
}
\details{
BirdFlow models are trained on eBird Status and Trends distributions
("Training distributions"). "Marginal distribution" describes a
distribution calculated from row or column sums of a marginal, joint
probability matrix (also part of the model). "Projected distribution" is
used to describe a distribution that is multiplied with a transition matrix
(derived from the marginal distribution) to project forward one timestep
or multiplied repeatedly with a series of transition matrices to project
forward multiple timesteps

Correlations are calculated for only the non-dynamically masked cells.

The \code{...} argument can be used to define a subset of time to evaluate over
in which case all metrics will be calculated only on that subset. The
default is to use all timesteps.
}
\examples{
bf <- BirdFlowModels::amewoo
distribution_performance(bf)

# Just for prebreeding_migration
distribution_performance(bf, start = "prebreeding_migration")

}
