% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_state.R
\name{expand_state}
\alias{expand_state}
\title{Convert a collapsed model state into an expanded state}
\usage{
expand_state(x, obj)
}
\arguments{
\item{x}{Either a vector with one value per active cell in the model or a
matrix with a row for each active cell and a column for each state, or
possible a higher dimension object.  In all cases the first dimension is
for states in the model.}

\item{obj}{A BirdFlow model or the \code{geom} component of one.}
}
\value{
An expanded version of \code{x} with one additional dimension in which
the first two dimensions represent rows and columns in space and
replace the first dimension in the input.
}
\description{
This function converts a collapsed state in which the first dimension
represents all the active cells in the model into an expanded state in which
the first two dimensions represent the locations of the cell in physical
space (row, col) and are thus suitable for plotting or conversion into
spatial objects.
}
\details{
In its collapsed form a state vector represents the state of a model at one
point in time as a probability distribution. Each value can be interpreted as
the relative abundance or probability of finding a bird at the corresponding
location. A special case is when the vector has mostly 0's and a single 1 in
which case it represents a single position of a bird or group of birds.
Additional dimensions can represent multiple timesteps and/or multiple
individual birds, model runs etc.

The location information in the collapsed state is not easily accessible as
position in the vector is ordered based on row-major ordered unmasked cells
in the extent and R uses column-major order.
}
