% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_threshold.R
\name{find_threshold}
\alias{find_threshold}
\title{function to determine the threshold in values that preserves a given
proportion of the value.}
\usage{
find_threshold(x, p, method = "weight")
}
\arguments{
\item{x}{A vector of values}

\item{p}{A proportion of x that we wish retain}

\item{method}{Either \code{"weight"} to retain a proportion of the total weight
or \code{"values"} to retain a proportion of non-zero values.}
}
\value{
A threshold in values of x such that dropping everything smaller than
the threshold retains at least proportion \code{p} of the total weight or
number of values in \code{x}
}
\description{
function to determine the threshold in values that preserves a given
proportion of the value.
}
\keyword{internal}
