% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dates.R
\name{get_dates}
\alias{get_dates}
\title{Retrieve dates component of a BirdFlow model}
\usage{
get_dates(bf)
}
\arguments{
\item{bf}{A BirdFlow object to retrieve dates from.}
}
\value{
A data frame with:
\item{interval}{The interval or timestep associated with each date.
It will range from 1 to \code{\link[=n_timesteps]{n_timesteps()}}.  With full models this is
equivalent to the week of the year but with truncated models may
not be.}
\item{date}{The date associated with the interval's midpoint.}
\item{midpoint, start, end}{The midpoint, start, and end of each interval
as a proportion of the total year that has elapsed.}
\item{doy}{The day of year associated with the midpoint of each interval.}
}
\description{
\code{get_dates()} retrieves the dates component of a BirdFlow model.  Most of
the resulting data is derived from \link[ebirdst:ebirdst_weeks]{ebirdst::ebirdst_weeks} and added
to the model object by   \code{\link[=preprocess_species]{preprocess_species()}}.  \code{\link[=truncate_birdflow]{truncate_birdflow()}}
creates models with a subset of the intervals and renames them
\verb{1:[n_timesteps()]} of the truncated model so if it has been used the
data may differ from \code{\link[ebirdst:ebirdst_weeks]{ebirdst::ebirdst_weeks()}}.
}
