% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dates.R
\name{get_dates}
\alias{get_dates}
\title{Function to calculate the midpoint in each interval (probaly week) in the
data.}
\usage{
get_dates(year, n = 52)
}
\arguments{
\item{year}{the year for which dates are calculated}

\item{n}{the year is broken into this many even intervals.}
}
\value{
data frame with
\item{doy}{The day of year associated with the midpoint of each interval.}
\item{interval}{The interval ID; it will range from 1 to n.}
\item{date}{The date associated with the interval's midpoint}
}
}
\description{
Derived from st_week_dates() in the BirdFlowForescast shiny application. This
is generalized to work with various interval lengths, but for now intervals
are weeks and n = 52.
}
\details{
By necessity one week ends up longer - 52 weeks = 264 days.

This function returns a \link[base:data.frame]{base::data.frame} that describes
the midpoint of each interval.
}
\examples{
get_dates(2019)
}
