% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dynamic_mask.R
\name{get_dynamic_mask}
\alias{get_dynamic_mask}
\title{Get the dynamic mask from a BirdFlow object}
\usage{
get_dynamic_mask(x, which = "all")
}
\arguments{
\item{x}{a BirdFlow object}

\item{which}{indicates which timesteps to return. Can be one or more
integers indicating timesteps; character dates in the format
year-month-day e.g. \code{"2019-02-25"}; \code{\link[base:Dates]{Date}} objects;
or \code{"all"} which will return distributions for all timesteps.}
}
\value{
a vector or matrix with elements or rows corresponding to active
cells in the model indicating whether a cell is included in the model for the
given timestep.
}
\description{
If a BirdFlow object has a dynamic mask it controls which subset of the
active cells in the model are active for a given timestep.  This is derived
from eBird Status and Trends (S&T) distributions; cells that are zero in the
S&T data for each timestep are excluded from the model for that timestep,
and tracked with the dynamic mask.  The dynamic mask has the same dimensions
as the distributions (see \code{\link[=get_distr]{get_distr()}}) with either a vector of \code{\link[=n_active]{n_active()}}
for a single timestep or a matrix with \code{\link[=n_active]{n_active()}} rows if more than one
timestep. The values are logical with TRUE corresponding to the cells that
are part of the model (for the given timestep).
}
\examples{
library(terra) # For plotting of spatRaster
bf <- BirdFlowModels::amewoo |> add_dynamic_mask()
dm <- get_dynamic_mask(bf, c(1,  10, 21, 31, 42, 52))
r <- rasterize_distr(dm, bf)
plot(r, legend = FALSE, mar = c(1.3, 1.3, 1.3, .3)  )

}
