% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loss.R
\name{get_loss}
\alias{get_loss}
\title{Get loss values for each step in the model fitting process}
\usage{
get_loss(bf)
}
\arguments{
\item{bf}{A BirdFlow object.}
}
\value{
A data frame with columns:
\item{dist}{The distance loss}
\item{ent}{The entropy loss}
\item{obs}{The observation loss}
\item{total}{The total weighted loss}
}
\description{
\code{get_loss()} returns a data frame of loss values. Each row corresponds to
a step in the fitting process.
}
\examples{
bf <- BirdFlowModels::amewoo
get_loss(bf)
}
\seealso{
\code{\link[=plot_loss]{plot_loss()}}
}
