% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mask.R
\name{get_mask}
\alias{get_mask}
\title{Extract mask from BirdFlow model}
\usage{
get_mask(bf, format = "SpatRaster")
}
\arguments{
\item{bf}{A BirdFlow model}

\item{format}{One of \code{'SpatRaster'} for a \link[terra:SpatRaster-class]{terra::SpatRaster} object,
\code{'numeric'} for a matrix or array, or\code{'dataframe'} for raster data
suitable for plotting with \code{\link[ggplot2:geom_tile]{ggplot2::geom_raster()}}}
}
\value{
The return type of \code{get_mask()}depends on the \code{format} argument:
\itemize{
\item \code{"SpatRaster"} (the default) returns a \link[terra:SpatRaster-class]{terra::SpatRaster} object.
\item \code{"numeric"} returns the mask as a matrix.
\item \code{"dataframe"} will return a data frame suitable for plotting with
\link[ggplot2:geom_tile]{ggplot2::geom_raster} with columns:
\itemize{
\item \code{row}, \code{col} the row and column indices of each cell.
\item \code{x}, \code{y} the x and y coordinates of the cell center.
\item \code{i} the location index (in \code{bf}) of the cell.
\item \code{mask}   \code{TRUE} for cells included in the \emph{model}, \code{FALSE} for excluded
cells.
}
}
}
\description{
\code{get_mask()} extracts the static mask from a BirdFlow model.  The
static mask is a logical raster indicating which cells are included in the
model (at any timestep).  These are also the cells (in row major order)
that correspond with distribution values, and location indices.
}
\examples{
\donttest{
bf <- BirdFlowModels::amewoo
m <- get_mask(bf)
}

\dontrun{
library(terra)
plot(m)
}

}
