% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_naturalearth.R
\name{get_naturalearth}
\alias{get_naturalearth}
\alias{get_states}
\alias{get_coastline}
\alias{get_countries}
\title{retrieve, crop, and transform Natural Earth data}
\usage{
get_naturalearth(
  bf,
  type,
  scale = "medium",
  buffer = 10,
  keep_attributes = FALSE,
  country,
  ...
)

get_states(bf, country, scale = "medium", buffer = 15, keep_attributes = FALSE)

get_coastline(bf, scale = "medium", buffer = 15, keep_attributes = FALSE)

get_countries(bf, scale = "medium", buffer = 15, keep_attributes = FALSE)
}
\arguments{
\item{bf}{BirdFlow object}

\item{type}{The type of data to retrieve. One of "coastline", "country", or
"states" for data included in \pkg{rnaturalearth}; or any value accepted by
\code{\link[rnaturalearth:ne_download]{rnaturalearth::ne_download()}}.}

\item{scale}{The scale of data to return. Ignored if type is "states",
otherwise passed to one of \link[rnaturalearth:ne_download]{ne_download()}.
\link[rnaturalearth:ne_coastline]{ne_coastline()},
\link[rnaturalearth:ne_countries]{ne_countries()}, or
\link[rnaturalearth:ne_states]{ne_states()}.
Valid values are 110, 50, 10, small', 'medium', and 'large'.}

\item{buffer}{A buffer in degrees (latitude and longitude) to add to the
extent of \code{bf} prior to cropping the Natural Earth data.}

\item{keep_attributes}{If \code{FALSE}, the default attribute columns are dropped
to facilitate clean plotting. sf by default creates one facet per column in
the data.}

\item{country}{if retrieving states with \code{get_states()} or
\code{get_naturalearth(type = "states")}) this is used to select a country. If
omitted, states from all countries are returned.}

\item{...}{Other arguments to be passed to \code{\link[rnaturalearth:ne_download]{rnaturalearth::ne_download()}}.
Quite possibly you will want to use \code{category = "physical"}.}
}
\value{
\link[sf:sf]{sf} object with the same coordinate reference system
(CRS) as \code{bf} and (with default buffer) a somewhat larger extent.
}
\description{
These are convenience wrappers to \pkg{rnaturalearth} functions. They
retrieve, crop, format, and project
\href{https://www.naturalearthdata.com/}{Natural Earth}
data to facilitate plotting with
BirdFlow objects. The output is the desired data in the same CRS as the
BirdFlow object with a somewhat larger extent.
}
\details{
This function eliminates the hassle of clipping Natural Earth data in its
CRS to bounds from the BirdFlow object and CRS, and then transforming the
data to the BirdFlow CRS, which is the necessary workflow to prevent
artifacts caused when polygons or lines are shifted across the bounds of
the CRS.

\code{get_states()} requires \pkg{rnaturalearthhires}. Install with:\cr
\code{ install.packages("devtools") # if you don't have it already
 devtools::install_github("ropensci/rnaturalearthhires") }
}
\examples{
 bf <- BirdFlowModels::amewoo
 coast <- get_coastline(bf)

 \dontrun{
 library(terra)
 library(sf)
 plot(rast(bf, 1))
 plot(coast, add = TRUE)   }
}
