% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.R
\name{has}
\alias{has}
\alias{has_marginals}
\alias{has_transitions}
\alias{has_distr}
\alias{has_dynamic_mask}
\title{Does a BirdFlow object have certain compoenents}
\usage{
has_marginals(x)

has_transitions(x)

has_distr(x)

has_dynamic_mask(x)
}
\arguments{
\item{x}{A BirdFlow model}
}
\value{
Logical indicating the BirdFlow model has the relevant element
}
\description{
These functions return information about the contents of a
BirdFlow object.
}
