% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.R
\name{has}
\alias{has}
\alias{has_marginals}
\alias{has_transitions}
\alias{has_distr}
\title{Does a BirdFlow object have  marginals, distributions or transitions

These are private functions to query the contents of a BirdFlow object and
allow for more concise and clear code.}
\usage{
has_marginals(x)

has_transitions(x)

has_distr(x)
}
\arguments{
\item{x}{a BirdFlow model}
}
\value{
logical indicating the BirdFlow model has the relevant element
}
\description{
Does a BirdFlow object have  marginals, distributions or transitions

These are private functions to query the contents of a BirdFlow object and
allow for more concise and clear code.
}
