% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_birdflow.R
\name{import_birdflow}
\alias{import_birdflow}
\title{Import BirdFlow models from HDF5 files}
\usage{
import_birdflow(hdf5, ..., version)
}
\arguments{
\item{hdf5}{Path to an HDF5 file containing a fitted BirdFlow model.}

\item{...}{Arguments to be passed to a version specific internal function.
Currently deprecated was used in early BirdFlow models.}

\item{version}{(optional) force reading of BirdFlow models as a particular
version. Normally, this will be determined from metadata in the HDF5
file.}
}
\value{
A BirdFlow object.
}
\description{
This function imports a BirdFlow model from an HDF5
(Hierarchical Data Format version 5) file. The workflow is to
first use \code{\link[=preprocess_species]{preprocess_species()}} to download and format the data for
model fitting, producing an HDF5 file; then use BirdFlow Python software to
fit the model and add marginals and hyperparameters to the HDF5 file; and,
finally import the model with \code{import_birdflow()}.
}
\details{
After importing you may want to call \code{\link[=sparsify]{sparsify()}} to reduce the
object size.  If you plan on doing many calls to \code{\link[=route]{route()}} or \code{\link[=predict]{predict()}}
you should call \code{\link[=build_transitions]{build_transitions()}} as well.
}
\section{HDF5 file version}{


The HDF5 BirdFlow model files have an internal version number that is
incremented on major changes to the HDF5 file structure. The file
version is retained in the imported object.
\itemize{
\item Version 1 predates \link[=preprocess_species]{BirdFlowR preprocessing} and only
contains marginals and a few other python objects, importing it requires
having an associated TIFF file that has the extent and distribution data.
The version number (1) is kept in \code{"/version"} within the HDF5 file.
\item Version 2 is the first version that includes preprocessing in BirdFlowR.
It does not include dynamic masking.  The version number (2) is stored
in  \code{"/metadata/birdflow_version"}.  Some version 2 files include
hyper parameters in the HDF5 but these are not read into R.
\item Version 3 marks the transition to dynamic masking, the HDF5 created during
preprocessing gained \code{"/geom/dynamic_mask"} and \code{"/distances"}  (great
circle distance matrix).  As in version 2 the version is
stored in \code{"/metadata/birdflow_version"} but unlike previous versions the
R package version that did the preprocessing is saved in character
format to \code{"/metdata/birdflowr_version"}. When version 3 files are imported
into R with model fits the hyperparameters are saved to
\verb{$metadata$hyperparameters} (a list).
}
}

