% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_prototype.R
\name{import_prototype}
\alias{import_prototype}
\title{Function to import BirdFlow model from the Shiny App}
\usage{
import_prototype(base_dir, species)
}
\arguments{
\item{base_dir}{The base directory of the shiny app}

\item{species}{The ebird species code for the model}
}
\value{
a BirdFlow object
}
\description{
This function imports the BirdfFlow model data from the prototype Shiny
application. It was a stopgap.  Importing models will ussually be done
with \link{import_birdflow}.
}
\details{
Forward transition matrices are transposed on read so that
all transitions can be implemented as M \%*\% d.
}
