% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_conversions.R
\name{index_conversions}
\alias{index_conversions}
\alias{x_to_col}
\alias{y_to_row}
\alias{row_to_y}
\alias{col_to_x}
\alias{i_to_rowcol}
\alias{i_to_row}
\alias{i_to_col}
\alias{i_to_x}
\alias{i_to_y}
\alias{rc_to_i}
\alias{xy_to_i}
\title{Functions to convert among spacial references}
\usage{
x_to_col(x, obj)

y_to_row(y, obj)

row_to_y(row, obj)

col_to_x(col, obj)

i_to_rowcol(i, obj)

i_to_row(i, obj)

i_to_col(i, obj)

i_to_x(i, obj)

i_to_y(i, obj)

rc_to_i(row, col, obj)

xy_to_i(x, y, obj)
}
\arguments{
\item{x, y}{The x and y coordinates associated with the BirdFlow model or data.
This will typically be an easting and northing in meters. A two column
matrix containing x and y columns can also be passed to\code{x} in which case \code{y}
should be omitted.}

\item{obj}{Either a full BirdFlow model or the geom component of one.}

\item{row, col}{The row and column index of a cell in the BirdFlow model and
associated raster data. Optionally, a two column
matrix containing row and column indices in columns 1 and 2 respectively
can be passed to \code{row} in which case \code{col} should be omitted.}

\item{i}{The index along a state vector that contains the data for
unmasked cells.}
}
\value{
\describe{

\item{\code{x_to_col(x, obj)} and \code{y_to_row(y, obj)}}{return the column or row
index that each \code{x} or \code{y} coordinate falls within.}

\item{\code{row_to_y(row, obj)} and \code{col_to_x(col, obj)}}{return the y or x
coordinate of the center each row or column. }

\item{\code{i_to_rowcol(i, obj)}}{returns the row and column index in the full
matrix corresponding to the index, \code{i} of the vector state.}

\item{\code{i_to_row(i, obj)} and \code{i_to_col(i, obj)}}{return the row or column
index corresponding to the vector state index, \code{i}.}

\item{\code{i_to_x(i, obj)} and \code{i_to_y(i, obj)}}{return the x or y coordinate
from the vector state index, \code{i}.}

\item{\code{rc_to_i(row, col, obj)} and \code{xy_to_i(x, y, obj)}}{return the state
space index corresponding to x and y coordinates or row and column indices.}
}
}
\description{
These functions allow conversions among ways of referencing
values in BirdFlow models and output. Specifically they convert among the
spacial coordinates, row and col index in the full raster matrix, and index
along the state vector.
}
\seealso{
\itemize{
\item \link{expand_state} converts a state vector into it's matrix equivalent or
a state matrix (representing multiple states) into an array equivalent.
\item \link{rasterize_state} converts a state vector into a \link[terra:terra-package]{terra::terra}
}
}
