% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_between.R
\name{is_between}
\alias{is_between}
\title{Determine if points are between BirdFlow cells}
\usage{
is_between(bf, points = NULL, radius = NULL, n_directions = 1)
}
\arguments{
\item{bf}{A BirdFlow model}

\item{points}{The points to evaluate betweenness on. If NULL the cell
centers of all the raster cells within the BirdFlow model that are between
active cells in the model will be used. This is calculated by comparing the
cell centers to a buffered convex hull around the active cell centers.}

\item{radius}{A point is considered between two locations if it is within
\code{radius} meters (along a great circle) of the great circle line between the
locations. \code{radius} defaults to half the cell size (\code{mean(res(bf))/2}).}

\item{n_direction}{The number of (equally spaced) directional bins to
classify bearings into.  Currently only \code{1} is supported.}
}
\value{
An array with dimensions representing the
"from" location, the "to" location, and the \code{points}. Cells are \code{TRUE}
if the associated point is between the associated from and to locations.
}
\description{
This internal function is used to create a betweenness array with dimensions
\href{n_active()}{n_active(bf)}, \href{n_active()}{n_active(bf)}, and \code{length(points)}.
The first two dimensions represent pairs of locations within the
BirdFlow model (active cells).
The third dimension represents the \code{points}. Each value in the array is
\code{TRUE} if the point is between the associated model cells. Specifically,
if the point is within \code{radius} meters (along a great circle)
of the great circle line connecting the cell centers.
}
\details{
If \code{points} and \code{radius} are \code{NULL} they default to the cell radius and the
center of all cells within the BirdFlow extent that fall between
any active cells.
}
\keyword{internal}
