% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{load_model}
\alias{load_model}
\title{Load BirdFlow models from a collection}
\usage{
load_model(
  model,
  update = TRUE,
  collection_url = birdflow_options("collection_url")
)
}
\arguments{
\item{model}{The model name to load}

\item{update}{If \code{TRUE} (the default) then both the index and cached model
file are checked against the server's version to make sure they are
up-to-date and downloaded again if they are not.
If \code{FALSE} then neither the index nor the model will be checked.
Set to \code{FALSE} after downloading the model(s) you need if you want to make
sure the model does not change during your analysis (even if updated on
the server); or if working offline.}

\item{collection_url}{The url of a collection. Should be the path to
the base directory (not an index.html file).}
}
\value{
The designated BirdFlow model is returned.
}
\description{
Load a named model from the current (likely the default) model collection.
If the model doesn't exist in the local cache or isn't up-to-date the the
cache will be updated prior to loading.
}
\examples{
\dontrun{

 index <- load_collection_index()
 bf <- load_model(index$model[1])

}

}
\seealso{
\code{\link[=load_collection_index]{load_collection_index()}}
}
