% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_date.R
\name{lookup_date}
\alias{lookup_date}
\title{Retrieve date associated with timesteps, transitions, or marginals}
\usage{
lookup_date(x, bf, timestep = NULL)
}
\arguments{
\item{x}{A vector of one of the following formats:
\enumerate{
\item Integer between 1 and \code{n_timesteps(bf)} representing timestep.
\item Character with "T" followed by digits that indicate timesteps,
(this format is used internally to label timestep dimensions of objects)
\item Marginal or Transition names. These start with either "T_" or "M_", and
then have two timesteps respresented by digits and separated by a dash,
E.g. "T_01-02".
}}

\item{bf}{A BirdFlow object}

\item{timestep}{Deprecated alternative to \code{x}.  Previous versions of
\code{lookup_dates()} only supported timestep input and used \code{timestep} as
the first argument.}
}
\value{
A Date object
}
\description{
Retrieve date associated with timesteps, transitions, or marginals
}
\examples{
bf <- BirdFlowModels::amewoo
lookup_date(1:5, bf)
}
\seealso{
\code{\link[=get_dates]{get_dates()}}, \code{\link[=lookup_timestep]{lookup_timestep()}}, \code{\link[=lookup_timestep_sequence]{lookup_timestep_sequence()}}
}
