% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_date.R
\name{lookup_date}
\alias{lookup_date}
\title{Retrieve date associated with a timestep}
\usage{
lookup_date(timestep, bf)
}
\arguments{
\item{timestep}{Integer between 1 and \code{n_timesteps(bf)}}

\item{bf}{A BirdFlow object}
}
\value{
A Date object
}
\description{
Retrieve date associated with a timestep
}
\examples{
bf <- BirdFlowModels::amewoo
lookup_date(1:5, bf)
}
\seealso{
\code{\link[=get_dates]{get_dates()}}, \code{\link[=lookup_timestep]{lookup_timestep()}}, \code{\link[=lookup_timestep_sequence]{lookup_timestep_sequence()}}
}
