% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_timestep_sequence.R
\name{lookup_season_timesteps}
\alias{lookup_season_timesteps}
\title{Lookup breeding, non-breeding, or migration season timesteps}
\usage{
lookup_season_timesteps(x, season, buffer = 1)
}
\arguments{
\item{x}{a BirdFlow object}

\item{season}{the season to lookup timesteps for one of the four seasons
returned by \code{\link[=species_info]{species_info()}} or one of the alternative  names listed  in
details.}

\item{buffer}{the number of extra timesteps to add to the beginning and end
of the season.}
}
\value{
a series of integers indicating which timesteps correspond with the
(possibly buffered) season.
}
\description{
Retrieve the timesteps associated with a season for the species modeled by
a BirdFlow object, possibly with a buffer (in timesteps) added on.
}
\details{
\code{season} should be one of the the four seasons or their alternate names
listed below.\tabular{ll}{
   season \tab alternate names \cr
   \code{prebreeding_migration} \tab \code{pre}, \code{prebreeding}, \code{spring} \cr
   \code{breeding} \tab \code{breed}, \code{summer} \cr
   \code{postbreeding_migration} \tab \code{post}, \code{postbreeding}, \code{fall} \cr
   \code{nonbreeding} \tab \code{non}, \code{winter} \cr
}
}
\examples{
bf <- BirdFlowModels::rewbla
lookup_season_timesteps(bf, "breeding", buffer = 0)

}
