% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_species_metadata.R
\name{lookup_species_metadata}
\alias{lookup_species_metadata}
\title{Lookup eBird species metadata}
\usage{
lookup_species_metadata(
  species,
  skip_checks = FALSE,
  min_season_quality = 3,
  quiet = FALSE
)
}
\arguments{
\item{species}{An eBird species code, common name, or scientific name. It
will be processed by \code{\link[ebirdst:get_species]{ebirdst::get_species()}}.}

\item{skip_checks}{Set to \code{TRUE} to skip the checks for eBird model quality
and for migrant status. Defaults to \code{TRUE} - do the tests.}

\item{min_season_quality}{If \code{skip_checks = FALSE} and the model
quality for any of the four seasons drops below this threshold an error
is thrown.}

\item{quiet}{Set to \code{TRUE} to supress messages.}
}
\value{
A list of species information derived from ebirdst.
See \code{\link[=species_info]{species_info()}} for a description of the items.
}
\description{
\code{lookup_species_metadata()} uses \pkg{ebirdst} to generate
a list  identical to the \code{species} component of a BirdFlow model.
It is an internal function used by \code{\link[=preprocess_species]{preprocess_species()}} and \code{\link[=Routes]{Routes()}}.
See \code{\link[=species_info]{species_info()}} for a description of the list items.
}
\seealso{
\itemize{
\item \code{\link[=species]{species()}} and \code{\link[=species_info]{species_info()}} for getting species information
from a BirdFlow model.
\item \code{\link[ebirdst:get_species]{ebirdst::get_species()}} to resolve a species name or code.
\item \link[ebirdst:ebirdst_runs]{ebirdst::ebirdst_runs} for the source of the information.
}
}
\keyword{internal}
