% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_timestep.R
\name{lookup_timestep}
\alias{lookup_timestep}
\title{lookup timestep}
\usage{
lookup_timestep(x, bf)
}
\arguments{
\item{x}{a character object representing date as year-month-day e.g.
"2023-03-29", date object, or a numeric timestep.}

\item{bf}{a BirdFlow object}
}
\value{
a vector of timesteps corresponding to elements in \code{x}
}
\description{
This function returns the timestep or timesteps associated with \code{x} in a
particular BirdFlow model, where \code{x} dates or timesteps in various formats.
}
\details{
So far all BirdFlow objects have had timesteps corresponding to weeks of
the year and matching the S&T timesteps. However, it is likely that
we will add the ability to make BirdFlow objects that only model part of the
year. If we do this the timestep values will not necessarily match weeks.

If \code{x} is numeric it is assumed to already be a timestep. This is useful
when using this function internally within \code{\link[=route]{route()}} and \code{\link[=predict]{predict()}}.
}
\examples{
bf <- BirdFlowModels::amewoo
lookup_timestep(c("2001-3-23", "2022-12-05"), bf)

}
