% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_timestep.R
\name{lookup_timestep}
\alias{lookup_timestep}
\title{lookup timestep}
\usage{
lookup_timestep(x, bf)
}
\arguments{
\item{x}{character representing date as year-month-day eg "2023-03-29",
a date object, or a numeric timestep.}
}
\value{
a vector of timesteps corresponding to elements in \code{x}
}
\description{
This function looks up the timestep or timesteps associated with \code{x}.
}
\examples{
bf <- BirdFlowModels::amewoo
lookup_timestep(c("2001-3-23", "2022-12-05"), bf)

}
