% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_transitions.R
\name{lookup_transitions}
\alias{lookup_transitions}
\title{Lookup a series of transitions connecting two dates or timesteps}
\usage{
lookup_transitions(x, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=lookup_timestep_sequence]{lookup_timestep_sequence}}
  \describe{
    \item{\code{season}}{a season name, season alias, or "all".  See
\code{\link[=lookup_season_timesteps]{lookup_season_timesteps()}} for options.}
    \item{\code{start}}{The starting point in time specified as a
timestep, character date, or date object.}
    \item{\code{end}}{The ending point in time as a date or timestep.}
    \item{\code{direction}}{Either "forward" or "backward" defaults to
\code{"forward"} if not processing dates.  If using date input \code{direction} is
optional and is only used to verify the direction implicit in the dates.}
    \item{\code{season_buffer}}{Only used with \code{season} input. \code{season_buffer} is
passed to \code{\link[=lookup_season_timesteps]{lookup_season_timesteps()}} and defaults to 1; it is the number
of timesteps to extend the season by at each end.}
    \item{\code{n_steps}}{Alternative to \code{end}  The end will be \code{n_steps} away from
\code{start} in \code{direction}; and the resulting sequence will have \code{n_step}
transitions and \code{n_steps + 1}  timesteps.}
  }}
}
\value{
A character vector with the named transitions required to get between
\code{start} and  \code{end}
}
\description{
The private function\code{lookup_transitions()} returns an ordered vector of transition names that
connect start to end. If \code{start} and \code{end} are dates than their order
determines whether the transitions flow forward or backward in time.  If they
are timesteps than the \code{direction} argument should be used to indicate
whether to project "forward" or "backward" in time possibly passing the year
boundary.
}
\details{
Transitions are named "T_[from]-[to]" where [from] and [to]
are timesteps padded with zeros. Direction is important; "T_03-04"
represents a transition backward in time.
}
\keyword{internal}
