% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_transitions.R
\name{lookup_transitions}
\alias{lookup_transitions}
\title{Lookup a series of transitions connecting two dates or timesteps}
\usage{
lookup_transitions(x, start, end, direction, season_buffer)
}
\arguments{
\item{x}{A BirdFlow object}

\item{start, end}{The starting and ending points in time specified as
timesteps, character dates, date objects, or if  \code{end} is omitted
start may be set to "all" or a season name to be interpreted by
\code{\link[=lookup_season_timesteps]{lookup_season_timesteps()}}.}

\item{direction}{Either "forward" or "backward".

If \code{start} and \code{end}
represent dates and \code{direction} is used an error will
be thrown if \code{direction} isn't consistent with direction indicated by the
dates.

If \code{start} and \code{end} are not dates, \code{direction} defaults to "forward" and
\code{start} and \code{end} should either both be timesteps (numeric); or
\code{end} should be omitted and start should be "all" or a season name.}

\item{season_buffer}{Only used if \code{start} is a season. \code{season_buffer} is
passed to \code{\link[=lookup_season_timesteps]{lookup_season_timesteps()}} and defaults to 1; it is the number
of timesteps to extend the season by at each end.}
}
\value{
A character vector with the named transitions required to get between
\code{start} and  \code{end}
}
\description{
\code{lookup_transitions()} returns an ordered vector of transition names that
connect start to end. If \code{start} and \code{end} are dates than their order
determines whether the transitions flow forward or backward in time.  If they
are timesteps than the \code{direction} argument should be used to indicate
whether to project "forward" or "backward" in time possibly passing the year
boundary.
}
\details{
Transitions are named "T_[from]-[to]" where [from] and [to]
are timesteps padded with zeros. Direction is important; "T_03-04"
represents a transition backward in time.
}
\keyword{internal}
