% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_transitions.R
\name{lookup_transitions}
\alias{lookup_transitions}
\title{Lookup a series of transitions connecting two dates}
\usage{
lookup_transitions(start, end, obj)
}
\arguments{
\item{start, end}{The starting and ending points in time. In one of the
following formats: character, a date in the form
year-month-day e.g. "2022-11-25" for Nov. 25, 2022); numeric, a
timestep; or  Date, a \code{\link[base::Dates]{Date}} object}

\item{obj}{A BirdFlow object or the dates component of one.}
}
\value{
A character vector with the named transitions required to get between
\code{start} and  \code{end}
}
\description{
This function returns an ordered vector of transition names that connect
start to end.  Currently time is linear and transitions can not cross forward
from December into January.  If \code{start} is later in the year than \code{end}
(regardless of what year) than the transitions will flow backwards in time.
}
