% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dates.R
\name{make_dates}
\alias{make_dates}
\title{Internal function to make the dates component of a BirdFlow model}
\usage{
make_dates(version_year = NULL)
}
\arguments{
\item{version_year}{leave NULL for typical usage.  Set to
a eBird version year to override.  Used by \code{switch_date_format()}.
version_year < 2021 will yield old date format.}
}
\value{
Dates table appropriate for the current version of \pkg{ebirdst}
}
\description{
Called from \link{preprocess_species} and not intended for other use.
}
\keyword{internal}
