% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mask.R
\name{make_mask}
\alias{make_mask}
\title{function to determine the extent of a SpatRaster}
\usage{
make_mask(x, count = FALSE)
}
\arguments{
\item{x}{a SpatRaster, typically it's multilayered and containins the
distribution of a species over time.}

\item{count}{= FALSE}
}
\value{
a single layer SpatRaster with the same CRS, alignment, and
resolution as x; cropped to the extent of the data in \code{x}.  With
\code{count = FALSE} the result is a logical that is \code{TRUE} if there are non-zero
cells in any layer in \code{x} and \code{FALSE} otherwise.  With \code{count = TRUE} the
result is an integer with a count of the number of timesteps in which the
cell appears (0 to 52 given weekly steps).
}
\description{
This function returns a logical single layer SpatRaster
which is TRUE if \code{x} has non-zero values in any layer.
NA does not count as non-zero.  The result is cropped to the extent of the
data in \code{x}. It's used in Birdflow to preprocess data.
}
\details{
\code{get_data_mask} is currently a public function but may be made private in
future versions of the package.
}
\keyword{internal}
