% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_timestep.R
\name{pad_timestep}
\alias{pad_timestep}
\title{pad timesteps from BirdFlow models}
\usage{
pad_timestep(x, bf)
}
\arguments{
\item{x}{A vector of timestep integers}

\item{bf}{A BirdFlow model}
}
\value{
a string with padded versions of x;  \code{1} becomes \code{"01"}.
}
\description{
This function is mostly for internal use but exported for advanced users.
It's primary purpose is to pad timesteps with zeros for looking up
transition names. It is called from \code{\link[=lookup_transitions]{lookup_transitions()}} which is in
turn used by \code{\link[=get_transition]{get_transition()}}.  Padding is generally two digit but the
level of padding is stored in the BirdFlow object so that we'll
be able to switch to three digit timesteps easily if, for example,
someday we decide to have daily timesteps.
}
\examples{
bf <- BirdFlowModels::amewoo
pad_timestep(1:5, bf)
}
\keyword{internal}
