% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distr.R
\name{plot_distr}
\alias{plot_distr}
\title{Plot distributions with ggplot2}
\usage{
plot_distr(
  distr,
  bf,
  subset = NULL,
  density_colors = NULL,
  coast_linewidth = 0.25,
  coast_color = grey(0.5),
  inactive_color = rgb(0.9, 0.9, 0.9, 0.5),
  limits = NULL,
  title = species(bf)
)
}
\arguments{
\item{distr}{a vector or matrix with distribution values corresponding to
active cells in \code{bf}.}

\item{bf}{A BirdFlow object}

\item{subset}{Indicates a subset of distr should be plotted. Use either
column numbers, column names, or a logical vector.}

\item{density_colors}{A color gradient that will be used to plot the density
values.  Leave or set to \code{NULL} for the default of
\code{ebirdst::abundance_palette(10, season = "weekly")}.}

\item{coast_linewidth}{The linewidth to use when plotting the coast. Default
is 0.25.  If NULL the coast will not be plotted.}

\item{coast_color}{The color to use for plotting the coastline. If NULL the
coast will not be plotted.}

\item{inactive_color}{The color to use for inactive cells in the landscape.}

\item{limits}{The range of density values to map \code{density_colors} to. The
default is the range of the values in distr (prior to applying \code{subset}).
If you want to standardize the range across multiple models of a single
species you might want to set to \code{c(0, max)} where \code{max} is the maximum
observed value across all models.}

\item{title}{The title for the plot. It defaults to the common name of the
species.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.  Use \code{print()} to render it.
}
\description{
Return a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} based on one or more distributions.
}
\examples{
plot_distr(get_distr(bf, 1), bf)
}
\seealso{
\itemize{
\item \code{\link[=plot_route]{plot_route()}} for plotting routes generated from a BirdFlow object.
\item \href{rasterize_distr()}{rasterize_distr(distr, bf, format = "dataframe")}
for converting a distribution to a dataframe with raster data for use with
\code{\link[ggplot2:geom_tile]{ggplot2::geom_raster()}}
}
}
