% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distr.R
\name{plot_distr}
\alias{plot_distr}
\title{Plot distributions with ggplot2}
\usage{
plot_distr(
  distr,
  bf,
  subset = NULL,
  show_mask = TRUE,
  show_dynamic_mask = FALSE,
  limits = NULL,
  dynamic_scale = FALSE,
  coast_linewidth = 0.25,
  coast_color = gray(0.5),
  gradient_colors = NULL,
  active_cell_color = rgb(1, 1, 1, 0.3),
  inactive_cell_color = rgb(0, 0, 0, 0.2),
  title = species(bf),
  value_label = "Density"
)
}
\arguments{
\item{distr}{a vector or matrix with distribution values corresponding to
active cells in \code{bf}.}

\item{bf}{A BirdFlow object}

\item{subset}{Defines an optional subset of \code{distr} that should be plotted.
Use either column numbers, column names, or a logical vector.}

\item{show_mask}{If \code{TRUE} (the default) the static mask that indicates which
cells are active in the model at any timestep will be shown.}

\item{show_dynamic_mask}{Defaults to \code{FALSE}. Set to \code{TRUE} to visualize the
dynamic mask. This is achieved by overwriting cells that are dynamically
masked with NA. For \code{show_dynamic_mask = TRUE} to work the column names in
\code{distr} should all be in \code{colnames(get_distr(bf))}.  This is
true for distributions returned by \href{predict.BirdFlowR}{\code{predict()}}
and \code{\link[=get_distr]{get_distr()}}.}

\item{limits}{The range of density values to map \code{gradient_colors} to. The
default is the range of the values in \code{distr} after applying \code{subset}. If
you want to standardize the range across multiple models of a single
species you might want to set to \code{c(0, max)} where \code{max} is the maximum
observed value across all models. Alternatively if the range is highly
variable among the columns in \code{distr} as when density spreads out from a
single point in the results of \href{predict.BirdFlowR}{\code{predict(bf)}} you may
want to set this smaller than the full range in which case the values will
be trunctated to the limits (see examples).}

\item{dynamic_scale}{Set to \code{TRUE} to have the range of the data in each
distribution mapped to the full color gradient. This allows visualizing the
full range of values within each timestep optimally at the cost of
consistency in the color scale among the facets - or animation frames
if using \code{\link[=animate_distr]{animate_distr()}}.}

\item{coast_linewidth}{The line width to use when plotting the coast. Default
is \code{0.25}.  If \code{NULL} the coast will not be plotted.}

\item{coast_color}{The color to use for plotting the coastline. If \code{NULL} the
coast will not be plotted.}

\item{gradient_colors}{A color gradient that will be used to plot the density
values. Leave or set to \code{NULL} for the default of
\href{ebirdst::abundance_palette()}{ebirdst::abundance_palette(10, season = "weekly")}.}

\item{active_cell_color}{The background color for active cells in the
landscape. Only used if \code{show_mask} is \code{TRUE}. These cells will only be
visible if there are \code{NA} values in \code{distr} or if \code{show_dynamic_mask}
is \code{TRUE}.}

\item{inactive_cell_color}{The color to use for inactive cells in the landscape.
These are cells that are always masked. Only relevant when
\code{show_mask = TRUE}.}

\item{title}{The title for the plot. It defaults to the common name of the
species (\code{species(bf)}).}

\item{value_label}{The label used for the values in the distribution.
Defaults to "Density"}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.  Use \code{print()} to render it.
}
\description{
Return a \href{ggplot2::ggplot}{ggplot} object with plots of one or more
distributions.
}
\details{
\subsection{Masks}{

BirdFlow objects have both a mask and a dynamic
mask. For any given timestep the model only covers cells that aren't excluded
by either of these masks.
\itemize{
\item \strong{mask} The static mask, AKA "mask", indicates which cells
in the raster extent are included in the model at any timestep these are
also called active cells (e.g. \code{\link[=n_active]{n_active()}}). The number of elements in a
distribution always match the number of unmasked cells in the (static) mask.
Cells are masked if they always have zero abundance in the Bird S&T data
for every week of the year.
\item \strong{dynamic mask} The dynamic mask indicates which of the cells not excluded
by the static mask are modeled within each timestep. It,like distributions,
only has values for active cells. Thus the dimensions of objects returned
by \href{get_dynamic_mask()}{get_dynamic_mask(bf)} and get
\href{get_distr()}{get_distr(bf)} will be identical.
The purposoe of the dynamic mask is to improve efficiency of fitting,
storing, and using BirdFlow models by eliminating unlikely locations
in the model.  The dynamic mask includes cells that have a zero in the
eBird S&T abundance for the associated week.
}

To display both masks set \code{show_dynamic_mask} to \code{TRUE} and leave
\code{show_mask} at it's default (\code{TRUE}).  Showing the dynamic mask relies on
matching the columns names of the distribution to timesteps in the model so
requires that the column names match the column names in \code{get_distr()}.
}

\subsection{Subtitles}{

Plot subtitles will be derived from the column names in \code{distr}. It may be
useful to define single distributions as a one column matrix so that the
label information is there. For example if \code{d} is a distribution in a
matrix with multiple columns use \code{drop=FALSE} while subsetting,
\code{plot_distr(d[, 1, drop = FALSE], bf)}; and to add a label to a vector
distribution use \verb{d <- matrix(d, ncol = 1);  colnames(d) <- "new label"}
}
}
\examples{
bf <- BirdFlowModels::amewoo
p <- plot_distr(get_distr(bf, c(1,11, 21)), bf, show_dynamic_mask = TRUE)
}
\seealso{
\itemize{
\item \code{\link[=animate_distr]{animate_distr()}} for animating distributions.
\item \code{\link[=plot_routes]{plot_routes()}} and  \code{\link[=animate_routes]{animate_routes()}} for visualizing routes.
\item \code{\link[=as_distr]{as_distr()}}, \code{\link[=get_distr]{get_distr()}}, \href{predict.BirdFlow}{\code{predict()}}
for functions that produce distributions.
}
}
