% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flux.R
\name{plot_flux}
\alias{plot_flux}
\title{Plot bird flux}
\usage{
plot_flux(
  flux,
  bf,
  subset = NULL,
  limits = NULL,
  dynamic_scale = FALSE,
  coast_linewidth = 0.25,
  coast_color = gray(0.5),
  gradient_colors = NULL,
  title = species(bf),
  value_label = "Flux"
)
}
\arguments{
\item{flux}{A flux object as created by
\link[=calc_flux]{calc_flux(format = "dataframe")}}

\item{bf}{A BirdFlow object}

\item{subset}{A subset of the transitions in \code{flux} to plot, can be
a logical vector of the same length as the number of transitions in \code{flux};
a numeric index of transitions in \code{flux}, or a subset of the transition names
in \code{flux}.}

\item{limits}{Two numbers representing the range in flux values to
display. Values outside of this range will be truncated to the range. With
the default of \code{NULL} the entire range is plotted.}

\item{dynamic_scale}{If \code{TRUE} then the range of the data in each
transition is mapped to the color palette. This makes it easier to see
the variation within a single transition but results in an inconsistent
scale among transitions.}

\item{coast_linewidth}{The line width used to plot the coast. Set to \code{NULL}
to skip plotting the coastline.}

\item{coast_color}{The color used to plot the coastline, or \code{NULL} to skip
plotting the coastline.}

\item{gradient_colors}{The colors palette used to represent the flux
intensity.}

\item{title}{The plot title}

\item{value_label}{The label for the flux values.}
}
\value{
\code{plot_flux} returns a \strong{ggplot2} object.  It can be displayed with
\code{print()}.
}
\description{
Plot bird flux
}
\examples{

\dontrun{
bf <- BirdFlowModels::amewoo
flux <- calc_flux(bf)

plot_flux(flux, bf)

animate_flux(flux, bf)
}

}
