% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loss.R
\name{plot_loss}
\alias{plot_loss}
\title{Plot changes in component and total loss during model fitting}
\usage{
plot_loss(bf, transform = "log10")
}
\arguments{
\item{bf}{A fitted Bird Flow model}

\item{transform}{Passed to \code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}} to set the y-axis
transformation. Reasonable values for this function include
"identity", "log", "log10", "log2", and "sqrt".}
}
\value{
a \strong{ggplot2} plot object.
}
\description{
Model fitting  - in \url{BirdFlowPy} - attempts to minimize the total weighted
loss.  This plot shows four lines:
\itemize{
\item \strong{Total loss} is the weighted sum of the three loss components. The
weighting may cause it to be lower than some of the components.
\item \strong{Observation loss} captures how well the model predicts the Status and
Trend distributions it was trained on.  Its weight is always set to 1 and its
relative weight is changed by adjusting the other to weights which are
usually much less than 1.
\item \strong{Distance loss} is lower when the routes encoded in the model are
shorter.
\item \strong{Entropy loss} is lower when the entropy in the model is higher.
}
}
\examples{
 bf <- BirdFlowModels::amewoo
 plot_loss(bf)
}
