% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.BirdFlow}
\alias{predict.BirdFlow}
\alias{predict}
\title{Predict bird distributions}
\usage{
\method{predict}{BirdFlow}(object, distr, ...)
}
\arguments{
\item{object}{A BirdFlow model object}

\item{distr}{a starting distribution}

\item{...}{
  Arguments passed on to \code{\link[=lookup_timestep_sequence]{lookup_timestep_sequence}}
  \describe{
    \item{\code{season}}{a season name, season alias, or "all".  See
\code{\link[=lookup_season_timesteps]{lookup_season_timesteps()}} for options.}
    \item{\code{start}}{The starting point in time specified as a
timestep, character date, or date object.}
    \item{\code{end}}{The ending point in time as a date or timestep.}
    \item{\code{direction}}{Either "forward" or "backward" defaults to
\code{"forward"} if not processing dates.  If using date input \code{direction} is
optional and is only used to verify the direction implicit in the dates.}
    \item{\code{season_buffer}}{Only used with \code{season} input. \code{season_buffer} is
passed to \code{\link[=lookup_season_timesteps]{lookup_season_timesteps()}} and defaults to 1; it is the number
of timesteps to extend the season by at each end.}
    \item{\code{n_steps}}{Alternative to \code{end}  The end will be \code{n_steps} away from
\code{start} in \code{direction}; and the resulting sequence will have \code{n_step}
transitions and \code{n_steps + 1}  timesteps.}
  }}
}
\value{
If multiple starting distributions are input in a matrix the result
will be an array with dimensions: location, distribution, and time. With one
input distribution the result will be a matrix with dimensions: location
and time.
}
\description{
\code{predict()} projects bird distributions into the future or past. Given an
initial distribution and time period specified via \code{...}, \code{predict()}
generates probability distributions for each timestep.
}
\seealso{
\itemize{
\item \code{\link[=lookup_timestep_sequence]{lookup_timestep_sequence()}} processes the time inputs
(\code{start}, \code{end}, \code{direction}, and \code{season_buffer})
\item \code{\link[=route]{route()}} and \code{\link[=route_migration]{route_migration()}} are similar to \code{predict()} but
generate routes instead of distributions.
}
}
