% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_species.R
\name{preprocess_species}
\alias{preprocess_species}
\title{download and format eBird Status and Trends data for BirdFlow model fitting}
\usage{
preprocess_species(
  species,
  out_dir,
  res = 80,
  p = 1,
  tiff = FALSE,
  overwrite = FALSE,
  crs = birdflow_crs
)
}
\arguments{
\item{species}{a species in any format accepted by \code{\link[ebirdst:get_species]{ebirdst::get_species()}}}

\item{out_dir}{output directory, files will be written here}

\item{res}{the target resolution of the BirdFlow model in kilometers.}

\item{p}{the proportion of the species distribution to be retained
(within each timestep), must be greater than 0 and less than or equal to 1.
In practice it should be above 0.9 and will probably be above 0.99. This
controls how much (if any) of the tail of the distribution density is
trimmed from the model.  1 indicates no trimming.}

\item{tiff}{if TRUE geoTIFF files will be exported}

\item{overwrite}{if TRUE any pre-existing output files will be overwritten,
otherwise the presence of pre-existing files is an error.}

\item{crs}{coordinate reference system to use.  Defaults to \link{birdflow_crs}.}
}
\value{
this function invisibly returns a list of the objects that are
written to the hdf5 file.
}
\description{
This uses the \pkg{ebirdst} package to download crop and resample abundance
data for the given species. It then writes the abundance data along with
a mask, spatial reference information, and species metadata to an HDF5 file.
}
