% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.BirdFlow}
\alias{print.BirdFlow}
\title{Print BirdFlow models}
\usage{
\method{print}{BirdFlow}(x, ...)
}
\arguments{
\item{x}{A BirdFlow object.}

\item{...}{arguments passed from other methods}
}
\value{
\code{x} returned invisibly and unchanged.
}
\description{
This is a print method for BirdFlow objects.
}
