% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RouteDataFunction.R
\name{print.BirdFlowIntervals}
\alias{print.BirdFlowIntervals}
\title{Print a BirdFlowIntervals Object}
\usage{
print.BirdFlowIntervals(birdflow_intervals)
}
\arguments{
\item{birdflow_intervals}{A \code{BirdFlowIntervals} object to print.}
}
\value{
Invisibly returns the input \code{birdflow_intervals} object.
}
\description{
Custom print method for \code{BirdFlowIntervals} objects, summarizing interval data
and metadata, including temporal and spatial ranges.
}
\examples{
# Create a BirdFlowIntervals object
birdflow_intervals <- BirdFlowIntervals(interval_df, species = NULL, geom = NULL, dates = NULL)
print(birdflow_intervals)
}
