% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RouteDataFunction.R
\name{print.BirdFlowIntervals}
\alias{print.BirdFlowIntervals}
\title{Print a BirdFlowIntervals Object}
\usage{
\method{print}{BirdFlowIntervals}(x, ...)
}
\arguments{
\item{x}{A \code{BirdFlowIntervals} object to print.}
}
\value{
Invisibly returns the input \code{birdflow_intervals} object.
}
\description{
Custom print method for \code{BirdFlowIntervals} objects, summarizing interval data
and metadata, including temporal and spatial ranges.
}
\examples{
# Create a BirdFlowIntervals object
interval_df <- data.frame(
  interval_id = 1:3,
  route_id = c("route1", "route1", "route2"),
  lon1 = c(-90, -89, -88),
  lon2 = c(-89, -88, -87),
  lat1 = c(40, 41, 42),
  lat2 = c(41, 42, 43),
  x1 = c(1000, 1100, 1200),
  x2 = c(1100, 1200, 1300),
  y1 = c(500, 600, 700),
  y2 = c(600, 700, 800),
  i1 = c(1, 2, 3),
  i2 = c(2, 3, 4),
  date1 = as.Date(c("2024-01-01", "2024-01-02", "2024-01-03")),
  date2 = as.Date(c("2024-01-02", "2024-01-03", "2024-01-04")),
  timestep1 = as.integer(c(1, 2, 3)),
  timestep2 = as.integer(c(2, 3, 4)),
  route_type = c("tracking", "tracking", "banding")
)
bf <- BirdFlowModels::amewoo
birdflow_intervals <- BirdFlowIntervals(interval_df, species = bf$species, 
metadata = bf$metadata, geom = bf$geom, dates = get_dates(bf))
}
