% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RouteDataFunction.R
\name{print.BirdFlowRoutes}
\alias{print.BirdFlowRoutes}
\title{Print a BirdFlowRoutes Object}
\usage{
\method{print}{BirdFlowRoutes}(x, ...)
}
\arguments{
\item{x}{A \code{BirdFlowRoutes} object to print.}

\item{...}{Additional arguments passed to \code{print.data.frame}.}
}
\value{
Invisibly returns the input \code{birdflow_routes} object.
}
\description{
Custom print method for \code{BirdFlowRoutes} objects, summarizing their contents, metadata,
and BirdFlow-specific attributes.
}
\examples{
# Create a BirdFlowRoutes object
route_df <- data.frame(
  route_id = 1:3,
  date = as.Date(c("2024-01-01", "2024-01-02", "2024-01-03")),
  lon = c(-90, -89, -88),
  lat = c(40, 41, 42),
  route_type = c("tracking", "banding", "unknown")
)
routes <- Routes(route_df)
bf <- BirdFlowModels::amewoo
birdflow_routes <- as_BirdFlowRoutes(routes, bf)
print(birdflow_routes)
}
