% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.BirdFlowRoutes.R
\name{print.BirdFlowRoutes}
\alias{print.BirdFlowRoutes}
\title{Print BirdFlow routes}
\usage{
\method{print}{BirdFlowRoutes}(x, ...)
}
\arguments{
\item{x}{A BirdFlowRoutes object.}

\item{...}{arguments passed from other methods}
}
\value{
\code{x} is returned invisibly and unchanged.
}
\description{
This is a print method for BirdFlowRouteq
objects.
}
