% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RouteDataFunction.R
\name{print.Routes}
\alias{print.Routes}
\title{Print Routes and BirdFlowRoutes objects}
\usage{
\method{print}{Routes}(x, ...)
}
\arguments{
\item{x}{A \code{Routes} or \code{BirdFlowRoutes} object to print.}

\item{...}{other arguments not used by this method.}
}
\value{
Invisibly returns the input object.
}
\description{
Print a summary of a \code{Routes} and \code{BirdFlowRoutes} objects
}
\examples{
# Create a Routes object
route_df <- data.frame(
  route_id = 1:3,
  date = as.Date(c("2024-01-01", "2024-01-02", "2024-01-03")),
  lon = c(-90, -89, -88),
  lat = c(40, 41, 42),
  route_type = c("tracking", "banding", "unknown")
)
routes <- Routes(route_df, species = list(common_name = "American Woodcock"))

print(routes)

# BirdFlowRoutes
bf <- BirdFlowModels::amewoo
bf_routes <- as_BirdFlowRoutes(routes, bf)

}
