% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_distr.R
\name{rasterize}
\alias{rasterize}
\alias{rasterize_distr}
\alias{rast,BirdFlow-method}
\title{Functions to convert a BirdFlow object or distribution into a SpatRaster}
\usage{
rasterize_distr(distr, bf)

\S4method{rast}{BirdFlow}(x, which = "all")
}
\arguments{
\item{distr}{a distribution in its vector form or a
matrix in which each column represents a different distribution.}

\item{bf}{A BirdFlow object.}

\item{x}{A BirdFlow object}

\item{which}{indicates which distributions to return. Can be one or more
integers indicating timesteps; character dates in the format
year-month-day e.g. \code{"2019-02-25"}; \code{\link[base:Dates]{Date}} objects; or
\code{"all"} which will return distributions for all timesteps.}
}
\value{
A \link[terra:SpatRaster-class]{terra::SpatRaster} object.
}
\description{
\code{rasterize_distr()} creates a \link[terra:SpatRaster-class]{SpatRaster} similar to those
created by \code{\link[terra:rast]{terra::rast()}} from one or more distributions in their compact
vector form. \code{rast()} converts a BirdFlow object directly to a
\link[terra:SpatRaster-class]{SpatRaster}.
}
