% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Routes-internal.R
\name{reset_index}
\alias{reset_index}
\title{Reset Route Indices}
\usage{
reset_index(routes)
}
\arguments{
\item{routes}{A \code{Routes} or data frame object.}
}
\value{
A data frame with updated route IDs.
}
\description{
Resets the route IDs in a \code{Routes} object to a new sequential numbering.
}
\keyword{internal}
