% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{route}
\alias{route}
\title{Generate BirdFlow routes}
\usage{
route(x, x_coord, y_coord, n, row, col, ...)
}
\arguments{
\item{x}{A BirdFlow object}

\item{x_coord, y_coord}{One or more sets of coordinates identifying starting
positions.}

\item{n}{Optional, if provided each starting position will be duplicated this
many times. \code{n} can be a single integer or a vector with one integer per
starting position.}

\item{row, col}{One or more row and column indices to begin routing from.
These are an alternative to \code{x_coord} and \code{y_coord} and both sets of
parameters should not be used at the same time.}

\item{...}{
  Arguments passed on to \code{\link[=lookup_timestep_sequence]{lookup_timestep_sequence}}
  \describe{
    \item{\code{start,end}}{The starting and ending points in time specified as
\code{timesteps}, dates as characters or date objects, or if  \code{end} is omitted
start may be set to "all" or a season name to be interpreted by
\code{\link[=lookup_season_timesteps]{lookup_season_timesteps()}}.}
    \item{\code{direction}}{Either "forward" or "backward".  Determines direction if
\code{start} and \code{end} are timesteps (numeric), or \code{start} is "all" or a season.
If \code{start} and \code{end} represent dates and \code{direction} is used an error will
be thrown if \code{direction} isn't consistent with direction implied by the
dates.  If \code{start} and \code{end} do not represent dates then \code{direction} will
default to "forward".}
    \item{\code{buffer}}{Only used if \code{start} is a season. A number of
timesteps to extend the season by. It is passed to
\code{\link[=lookup_season_timesteps]{lookup_season_timesteps()}} and defaults to 1.}
  }}
}
\value{
This will likely change. Currently returns a list with:
\item{points}{A data.frame with coordinates, date, and route id}
\item{lines}{a \link[sf:sf]{sf} object containing one line per route.}
}
\description{
route() projects bird positions over time based on the probabilities
embedded in a BirdFlow model. The output is linear, stochastic routes.
}
