% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{route}
\alias{route}
\title{Generate BirdFlow routes}
\usage{
route(x, x_coord, y_coord, n_each, row, col, ...)
}
\arguments{
\item{x}{A BirdFlow object}

\item{x_coord, y_coord}{One or more sets of coordinates identifying starting
positions.}

\item{n_each}{Optional, if provided each starting position will be
duplicated this many times. \code{n_each} can be a single integer or a vector
with one integer per starting position.}

\item{row, col}{One or more row and column indices to begin routing from.
These are an alternative to \code{x_coord} and \code{y_coord} and both sets of
parameters should not be used at the same time.}

\item{...}{
  Arguments passed on to \code{\link[=lookup_timestep_sequence]{lookup_timestep_sequence}}
  \describe{
    \item{\code{season}}{a season name, season alias, or "all".  See
\code{\link[=lookup_season_timesteps]{lookup_season_timesteps()}} for options.}
    \item{\code{start}}{The starting point in time specified as a
timestep, character date, or date object.}
    \item{\code{end}}{The ending point in time as a date or timestep.}
    \item{\code{direction}}{Either "forward" or "backward" defaults to
\code{"forward"} if not processing dates.  If using date input \code{direction} is
optional and is only used to verify the direction implicit in the dates.}
    \item{\code{season_buffer}}{Only used with \code{season} input. \code{season_buffer} is
passed to \code{\link[=lookup_season_timesteps]{lookup_season_timesteps()}} and defaults to 1; it is the number
of timesteps to extend the season by at each end.}
    \item{\code{n_steps}}{Alternative to \code{end} for specifying when a sequence should
end \code{n_steps} indicates how many timesteps the sequence will progress in
\code{direction} from \code{start} The sequence will have \code{n_steps + 1} timesteps.}
  }}
}
\value{
This will likely change. Currently returns a list with:
\item{points}{A dataframe with columns:
\describe{
\item{\code{x}, \code{y}}{coordinates of point along route}
\item{\code{date}}{date associated with that point}
\item{\code{timestep}}{timestep associated with point}
\item{\code{route}}{unique id for that route or individual}
\item{\code{i}}{location index for the point (see \code{\link[=i_to_xy]{i_to_xy()}})}
\item{\code{stay_id}}{within each route a sequential id for locations}
\item{\code{stay_len}}{how many timesteps was the Bird at that point during
the stay (minumum of 1)}
}
}
\item{lines}{a \link[sf:sf]{sf} object containing one line per route.}
}
\description{
\code{route()} projects bird positions over time based on the probabilities
embedded in a BirdFlow model. The output is linear, stochastic routes.
}
\seealso{
\code{\link[=route_migration]{route_migration()}}
}
