% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{route}
\alias{route}
\title{Generate BirdFlow routes}
\usage{
route(bf, n = 1, x_coord = NULL, y_coord = NULL, from_marginals = FALSE, ...)
}
\arguments{
\item{bf}{A BirdFlow object.}

\item{n}{If sampling starting positions (\code{x_coord}, and \code{y_coord} are NULL).
Generate this many samples.  Otherwise the \code{x_coord} and \code{y_coord} positions
will each be duplicated \code{n} times.}

\item{x_coord, y_coord}{Optional, if \code{NULL} starting points will be drawn from
the species distribution at the initial timestep.}

\item{from_marginals}{Use \code{FALSE} (the default) to use
distributions derived directly from eBird Status and Trends when sampling
starting locations. Set to \code{TRUE} to
sample from distributions derived from the fitted model parameters stored in
the marginals. Passed to \code{\link[=get_distr]{get_distr()}}.}

\item{...}{
  Arguments passed on to \code{\link[=lookup_timestep_sequence]{lookup_timestep_sequence}}
  \describe{
    \item{\code{season}}{a season name, season alias, or "all".  See
\code{\link[=lookup_season_timesteps]{lookup_season_timesteps()}} for options.}
    \item{\code{start}}{The starting point in time specified as a
timestep, character date, or date object.}
    \item{\code{end}}{The ending point in time as a date or timestep.}
    \item{\code{direction}}{Either "forward" or "backward" defaults to
\code{"forward"} if not processing dates.  If using date input \code{direction} is
optional and is only used to verify the direction implicit in the dates.}
    \item{\code{season_buffer}}{Only used with \code{season} input. \code{season_buffer} is
passed to \code{\link[=lookup_season_timesteps]{lookup_season_timesteps()}} and defaults to 1; it is the number
of timesteps to extend the season by at each end.}
    \item{\code{n_steps}}{Alternative to \code{end}  The end will be \code{n_steps} away from
\code{start} in \code{direction}; and the resulting sequence will have \code{n_step}
transitions and \code{n_steps + 1}  timesteps.}
  }}
}
\value{
A BirdFlowRoutes object with columns:
\item{\code{x}, \code{y}}{Coordinates of point along route.}
\item{\code{date}}{Date associated with that point.}
\item{\code{timestep}}{Timestep associated with point.}
\item{\code{route}}{Unique ID for that route or individual.}
\item{\code{i}}{Location index for the point (see \code{\link[=i_to_xy]{i_to_xy()}}).}
\item{\code{stay_id}}{Within each route a sequential id for locations.}
\item{\code{stay_len}}{How many timesteps was the Bird at that point during
the stay (minimum of 1).}
It also has \strong{experimental} attributes:
\describe{
\item{\code{geom}, \code{species}, \code{dates}}{The \code{geom}, \code{species}, and \code{dates}
components of the BirdFlow object the routes are derived from.}
\item{\code{metadata}}{The \code{metadata} component of the parent BirdFlow object,
with one additional item \code{route_type = "synthetic"}.}
}
}
\description{
\code{route()} projects bird positions over time based on the probabilities
embedded in a BirdFlow model. The output is linear, stochastic routes.
}
\examples{
bf <- BirdFlowModels::amewoo
rts <- route(bf, 10, season = "prebreeding")

\dontrun{
plot_routes(rts)
}

}
\seealso{
\code{\link[=plot_routes]{plot_routes()}}, \code{\link[=animate_routes]{animate_routes()}}
}
