% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{route}
\alias{route}
\title{Generate BirdFlow routes}
\usage{
route(x, x_coord, y_coord, n, row, col, start, end, direction)
}
\arguments{
\item{x}{A BirdFlow model object}

\item{x_coord, y_coord}{One or more sets of coordinates identifying starting
positions.}

\item{n}{Optional, if provided each starting position will be duplicated this
many times. \code{n} can be a single integer or a vector with one integer per starting
position.}

\item{row, col}{One or more row and column indices to begin routing from. These
are an alternative to \code{x_coord} and \code{y_coord} and both sets of parameters
should not be used at the same time.}

\item{start, end}{These define the time period to route over. They can be
\link[base:date]{Date} objects, integer timesteps, or a string with
"year-month-day" e.g. "2022-11-28".}

\item{direction}{either "forward" or "backwards", only used if \code{start} and
\code{end} represent timesteps.}
}
\value{
This will likely change. Currently returns a list with:
\item{points}{A data.frame with coordinates, date, and route id}
\item{lines}{a \link[sf:sf]{sf} object containing one line per route.}
}
\description{
route() projects bird positions over time based on the probabilities
embedded in a BirdFlow model. The output is linear, stochastic routes.
}
