% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_migration.R
\name{route_migration}
\alias{route_migration}
\title{Generate migration routes from a BirdFlow model}
\usage{
route_migration(bf, n, migration = "prebreeding", season_buffer = 1)
}
\arguments{
\item{bf}{\code{BirdFlow} model}

\item{n}{the number of routes to generate}

\item{migration}{"prebreeding", "pre", or "spring" for the prebreeding
migration; or "postbreeding", "post", or "fall" for the postbreeding
migration.}

\item{season_buffer}{a buffer in timesteps (likely weeks) to add to the
beginning and end of the season. The default of 1 means we start a week
before the metadata suggests the season starts and continue one week past
the end.}
}
\value{
This will likely change. Currently returns a list with:
\item{points}{A dataframe with columns:
\describe{
\item{\code{x}, \code{y}}{coordinates of point along route}
\item{\code{date}}{date associated with that point}
\item{\code{timestep}}{timestep associated with point}
\item{\code{route}}{unique id for that route or individual}
\item{\code{i}}{location index for the point (see \code{\link[=i_to_xy]{i_to_xy()}})}
\item{\code{stay_id}}{within each route a sequential id for locations}
\item{\code{stay_len}}{how many timesteps was the Bird at that point during
the stay (minumum of 1)}
}
}
\item{lines}{a \link[sf:sf]{sf} object containing one line per route.}
}
\description{
Create stochastic migration routes for a species by sampling appropriate
starting locations and then \link[=route]{routing} for the
duration of the migration window.
}
\details{
The two migration periods are defined based on the species information in the
BirdFlow model:
\describe{
\item{\code{prebreeding}}{ migration starts at \code{species_info("nonbreeding_end")}
and ends at \code{species_info("breeding_start")}}
\item{\code{postbreeding}}{ migration starts at \code{species_info("breeding_end")}
and ends at \code{species_info("nonbreeding_start")}}
}
}
\examples{
  bf <-  BirdFlowModels::amewoo
  rts <- route_migration(bf, 2)
  plot(rts$lines)
  head(rts$points)

}
\seealso{
\itemize{
\item \code{\link[=route]{route()}} allows setting the starting locations explicitly.
\item \link[=predict.BirdFlow]{predict(BirdFlow)} projects future or past
\item \code{\link[=lookup_season_timesteps]{lookup_season_timesteps()}} does what its name suggests.
}
}
