% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_migration.R
\name{route_migration}
\alias{route_migration}
\title{generate migration routes from a BirdFlow model}
\usage{
route_migration(bf, n, migration = "prebreeding")
}
\arguments{
\item{bf}{\code{BirdFlow} model}

\item{n}{the number of routes to generate}

\item{migration}{"prebreeding", "pre", or "spring" for the prebreeding
migration; or "postbreeding", "post", or "fall" for the postbreeding
migration.}
}
\value{
This will likely change. Currently returns a list with:
\item{points}{A data.frame with coordinates, date, and route id}
\item{lines}{a \link[sf:sf]{sf} object containing one line per route.}
}
\description{
create stochastic migration routes for a species by sampling approriate
starting locations and then routing (with \code{\link[=route]{route()}}) for the
duration of the migration window.
}
\details{
The two migration periods are defined based on the species information in the
BirdFlow model:\tabular{ll}{
   \code{prebreeding} \tab migration starts at \code{species_info("nonbreeding_end")} and ends at \code{species_info("breeding_start")} \cr
   \code{postbreeding} \tab migration starts at \code{species_info("breeding_end")} and ends at \code{species_info("nonbreeding_start")} \cr
}
}
