% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_migration.R
\name{route_migration}
\alias{route_migration}
\title{Deprecated function to generate migration routes from a BirdFlow model}
\usage{
route_migration(bf, n, migration = "prebreeding", season_buffer = 1)
}
\arguments{
\item{bf}{\code{BirdFlow} model}

\item{n}{the number of routes to generate}

\item{migration}{"prebreeding", "pre", or "spring" for the prebreeding
migration; or "postbreeding", "post", or "fall" for the postbreeding
migration.}

\item{season_buffer}{a buffer in timesteps (likely weeks) to add to the
beginning and end of the season. The default of 1 means we start a week
before the metadata suggests the season starts and continue one week past
the end.}
}
\value{
A BirdFlowRoutes object with columns:
\item{\code{x}, \code{y}}{coordinates of point along route}
\item{\code{date}}{date associated with that point}
\item{\code{timestep}}{timestep associated with point}
\item{\code{route}}{unique id for that route or individual}
\item{\code{i}}{location index for the point (see \code{\link[=i_to_xy]{i_to_xy()}})}
\item{\code{stay_id}}{within each route a sequential id for locations}
\item{\code{stay_len}}{how many timesteps was the Bird at that point during
the stay (minumum of 1)}
It also has \strong{experimental} attributes:
\describe{
\item{\code{geom}, \code{species}, \code{dates}}{The \code{geom}, \code{species}, and \code{dates} components
of the BirdFlow object the routes are derived from.}
\item{\code{metadata}}{The \code{metadata} component of the parent BirdFlow object, with
one additional item \code{route_type = "synthetic"}}
}
}
\description{
This function is now deprecated and will eventually be deleted. Please
transition to using  \code{route()} which can now both generate starting locations
by sampling the distributions in \code{bf} and use a season name to specify the
time period to route over.  The only adjustment that needs to be made is to
use the \code{season} argument to \code{route()} in place of the \code{migration} argument
to \code{route_migration()}.
}
\seealso{
\itemize{
\item \code{\link[=route]{route()}} should be used instead of this function.
}
}
\keyword{internal}
