% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Routes.R
\name{Routes}
\alias{Routes}
\title{\code{Routes()} is used to convert data frames containing bird movement
data into a formal \code{Routes} object with the same data.}
\usage{
Routes(data, species = NULL, source = NULL)
}
\arguments{
\item{data}{A data frame with data on bird movement. Likely
tracking, banding, or Motus data on real birds; or possibly
synthetic versions of the same.It must have the following columns:
\describe{
\item{\code{date}}{Date or Date Time object of class \code{Date}, \code{POSIXlt},
or \code{POSIXct}}
\item{\code{lat},\code{lon}}{The latitude and longitude of the location in
WGS84 (EPSG::4326)}
\item{\code{route_type}}{The type of route - one of
\code{"tracking"}, \code{"banding"}, \code{"motus"}, \code{"unknown"}, or \code{"synthetic"}
Types can be mixed in the column.
}
}
Other columns are permitted and will be retained in \code{Routes} object
but dropped if they are converted to \code{BirdFlowRoutes}.}

\item{species}{Either: a character scalar suitable for use with
\code{\link[ebirdst:get_species]{ebirdst::get_species()}}; or a list with
species metadata which must include \code{common_name} and can optionally
also include \code{scientific_name} and  \code{species_code} and any other standard
BirdFlow species metadata. See \code{\link[=species_info]{species_info()}} for a description
of the full list. Note list input is not checked against eBird species
codes and names. Scalar input is preferred unless the species does not
conform to eBird's taxonomy.}

\item{source}{Optional text describing the source of the data.
\code{source()} must be of class \code{character} can have one or more elements.}
}
\value{
An object of class \code{Routes} which has the following components
\item{data}{A data frame with the input \code{data}}
\item{species}{A list with, at a minimum items
\code{common_name},  \code{scientific_name}, and \code{species_code} and depending on
the \code{species} argument potentially having all the items returned by
\code{\link[=lookup_species_metadata]{lookup_species_metadata()}}}
\item{source}{Same as the input \code{source}}
}
\description{
\code{Routes()} is used to convert data frames containing bird movement
data into a formal \code{Routes} object with the same data.
}
