% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_distr.R
\name{sample_distr}
\alias{sample_distr}
\title{Sample locations from a distribution}
\usage{
sample_distr(x, n = 1)
}
\arguments{
\item{x}{a vector representing a single distribution; a matrix respresenting
one distribution per column, or (unlikely) an  array in which the first
dimension represents model states. The values in x are treated as relative
probability of the species being in each position.}

\item{n}{Only used if \code{x} is a vector representing a single model state, in
which case that model state will be sampled \code{n} times to generate a
matrix representing \code{n} sampled locations from that distribution.}
}
\value{
An object ussually with the same dimensions as \code{x} (but see \code{n})
single location (containing a 1) while the remaining locations
have 0's.
}
\description{
This function treats \code{x} as one or more probability distributions and converts
each to a single location by assigning a 1 to one value and
converting the rest to 0.
}
\examples{
# Generate a hypothetical state space vector
# representing the probability of finding a bird in each location.
# This is based on a normal distribution with higher probabilities in the
# center of the vector.
s <- dnorm(seq(-2, 2, by = .25))
s <- s / sum(s)

# sample it once
loc <- sample_state(s)

# sample it 200 times
locations <- sample_state(s, 200)

\dontrun{
# Plot the probability of each location
plot(1:length(s), s, xlab = "location", ylab = "prob")
# Plot the number of times each location was selected
plot(1:length(s), apply(locations, 1, sum),
     xlab = "location", ylab = "times sampled")
}

}
