% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink_birdflow.R
\name{shrink_birdflow}
\alias{shrink_birdflow}
\title{Shrink BirdFlow extent}
\usage{
shrink_birdflow(x)
}
\arguments{
\item{x}{A single BirdFlow object, or one or more paths to BirdFlow objects
stored as either hdf5 or rds files.}
}
\value{
If \code{x} is an extended BirdFlow model, see \code{\link[=extend_birdflow]{extend_birdflow()}}\code{than}shrink_birdflow()` returns the same model with its original extent.

If \code{x} is the path to one or more BirdFlow models than those files are
shrunk to their original extent and a logical vector of the
same length is returned with TRUE for success.
}
\description{
Shrink BirdFlow extent
}
\examples{
bf <- BirdFlowModels::amewoo

# Define extended extent for example
e <-  ext(bf)
buffer <- 3 * res(bf)
e[1] <- e[1] - buffer[1]
e[2] <- e[2] + buffer[1]
e[3] <- e[3] - buffer[2]
e[4] <- e[4] + buffer[2]

bf2 <- extend_birdflow(bf, e)

bf3 <- shrink_birdflow(bf2)

# Compare extents
data.frame(item = names(as.vector(ext(bf))),
            initial = as.vector(ext(bf)),
            extended = as.vector(ext(bf2)),
            shrunk = as.vector(ext(bf3)))

\dontrun{
# Plot
library(terra)
plot_distr(get_distr(bf, 1), bf)
plot_distr(get_distr(bf2, 1), bf2)
plot_distr(get_distr(bf3, 1), bf3)
  }

}
