% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap_to_birdflow.R
\name{snap_to_birdflow}
\alias{snap_to_birdflow}
\title{Assign cells and timesteps to coordinates and dates}
\usage{
snap_to_birdflow(
  d,
  bf,
  x_col = "lon",
  y_col = "lat",
  date_col = "date",
  id_cols = "id",
  crs = "EPSG:4326",
  aggregate = NULL
)
}
\arguments{
\item{d}{A data frame with bird movement data it must have the columns
specified by \code{x_col}, \code{y_col}, \code{date_col}, and \code{id_cols}.  The default
values for the first three are  \code{"lon"}, \code{"lat"}, and\code{"date"}.}

\item{bf}{A BirdFlow model}

\item{x_col}{The name of the column that holds x or longitude coordinates.
Default is \code{"lon"}.}

\item{y_col}{The name of the y or latitude column, default is \code{"lat"}.}

\item{date_col}{Name of the date column, defaults to \code{"date"}}

\item{id_cols}{The name of identification columns. One or more columns that
collectively define a unique track. If \code{aggregate} is not \code{NULL} these along
with \code{timestep} (calculated from the \code{date_col}) are used to define the
groups.  If aggregate is \code{NULL} these columns are not used but are retained
in the output.}

\item{crs}{The coordinate reference system used by \code{x_col} and
\code{y_col}. The default \code{"EPSG:4326"} corresponds to
\href{https://epsg.io/4326}{WGS 1984}}

\item{aggregate}{Leave the default \code{NULL} for no aggregation - each row in
\code{d} is processed separately and represented in the returned object.
If \code{aggregate} is set to one of the values below then the locations (rows)
in \code{d} that fall in the same week and track (see \code{id_cols}) will be
aggregated together such that there is only one location per week for each
track.

The possible values for \code{aggregate} are:
\itemize{
\item \code{mean} The mean of the "x", "y", and "date" columns are used for the week.
\item \code{median} The median of the "x", "y", and "date" columns are  used for the
week.
\item \code{midweek} The observation that is closest to the middle of the week is
used to represent the week. With ties the observation that occurs first
is used.
\item \code{random} One observation is randomly selected for each week.
}

Pending ideas, not yet implemented:
\itemize{
\item \code{gmedian}
\href{https://cran.r-project.org/web/packages/Gmedian/index.html}{geometric median}
\item \code{central} The point closest to the centroid of all the points is used to
represent the week.
}}
}
\value{
A data frame with columns
\item{id_cols}{The columns identified with \code{id_cols} will be retained
and be leftmost.}
\item{date}{This column will have information from \code{date_col} but not retain
its name or original formatting. If aggregate is NULL the input dates
will be retained, if not the date will vary with the aggregation method but
will represent the dates that went into the summary and not the mid-week
date associated with \code{timestep}. If the input \code{date_col} is a date-time
class then the output date column will be as well.
For example with \code{aggregate = "mean"} the date will be the average date of
the points in the week.}
\item{timestep}{The model timestep associated with \code{date}.}
\item{x, y}{The x and y coordinates of the point. These will always be in
\code{crs(bf)} and will represent the original or aggregated location and
not the cell center.}
\item{i}{The location index of the cell associated with the \code{x} and \code{y} in
\code{bf}.  See \code{\link[=i_to_xy]{i_to_xy()}}.}
\item{n}{The number of rows in \code{d} that contributed to each output row.
If aggregate is NULL every value will be 1. }
\item{error}{TRUE if there was an error.}
\item{message}{NA or the error message. The possible messages are:
\itemize{
\item "err_date" - The date could not be parsed with \code{\link[lubridate:as_date]{lubridate::as_date()}}
\item "err_truncated" - \code{bf} is \href{truncate_birdflow}{truncated} and the date
falls outside of portion of the year the model covers.
\item \code{"err_coords"} - The coordinates could not be transformed into \code{crs(bf)}
and thus likely are corrupt in some way.
\item \code{"err_not_active"} - the location does not fall within an active cell
as defined by the static mask.
\item \code{"err_dynamic"} - the location does not fall within the dynamic mask on
the associated date.
\item \code{"err_sparse"} - the location falls within the dynamic mask but that
location and date combination has been eliminated by
\href{sparsify}{sparsification}.
}

The function will always return the error message that appears first on
this list, even though in some cases multiple errors can be triggered.

With aggregation the first three errors prevent a row from being used and
it will be dropped prior to aggregation with a warning.
}
}
\description{
This function aligns bird movement data that has both location and date
information with a BirdFlow model's cells and timesteps.
Optionally,it will aggregate multiple points from a track within the same
week into a single point and date prior to determining the location ID.
}
\details{
If \code{aggregate} is NULL than than each row in the output will correspond with
the same row in the input.  With aggregation rows can be lost in two ways
(1) if the coordinates or timestep can't be resolved (first three errors
on list) then the row is dropped prior to aggregation.
(2) via aggregation all the rows within a week for a track will be collapsed
to one. These aggregated locations can still trigger the other three
errors on the list.

A location id will only be assigned if the location is a valid location for
the model on the associated date.
}
\examples{
\donttest{
bf <- BirdFlowModels::rewbla |> add_dynamic_mask()
obs <- BirdFlowModels::rewbla_observations
a <- snap_to_birdflow(obs, bf, id_cols = "bird_id")
}
}
