% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RouteDataFunction.R
\name{sort_by_id_and_dates}
\alias{sort_by_id_and_dates}
\title{Sort Routes by ID and Date}
\usage{
sort_by_id_and_dates(routes)
}
\arguments{
\item{routes}{A \code{Routes} or data frame object.}
}
\value{
A sorted data frame.
}
\description{
Sorts a \code{Routes} or data frame object by route ID and date.
}
