% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_birdflow.R
\name{truncate_birdflow}
\alias{truncate_birdflow}
\title{Truncate the timespan of a BirdFlow model}
\usage{
truncate_birdflow(bf, ...)
}
\arguments{
\item{bf}{A BirdFlow object.}

\item{...}{
  Arguments passed on to \code{\link[=lookup_timestep_sequence]{lookup_timestep_sequence}}
  \describe{
    \item{\code{season}}{a season name, season alias, or "all".  See
\code{\link[=lookup_season_timesteps]{lookup_season_timesteps()}} for options.}
    \item{\code{start}}{The starting point in time specified as a
timestep, character date, or date object.}
    \item{\code{end}}{The ending point in time as a date or timestep.}
    \item{\code{direction}}{Either "forward" or "backward" defaults to
\code{"forward"} if not processing dates.  If using date input \code{direction} is
optional and is only used to verify the direction implicit in the dates.}
    \item{\code{season_buffer}}{Only used with \code{season} input. \code{season_buffer} is
passed to \code{\link[=lookup_season_timesteps]{lookup_season_timesteps()}} and defaults to 1; it is the number
of timesteps to extend the season by at each end.}
    \item{\code{n_steps}}{Alternative to \code{end}  The end will be \code{n_steps} away from
\code{start} in \code{direction}; and the resulting sequence will have \code{n_step}
transitions and \code{n_steps + 1}  timesteps.}
  }}
}
\value{
A BirdFlow model that only contains information about transitions
for a subset of the year as specified by \code{...}.
}
\description{
\code{truncate_birdflow()} Eliminates marginals and/or transitions from a BirdFlow
model and adjusts other aspects of the model so that it only covers part of a
year.  The intent is to reduce object size and processing time when only part
of the year is of interest.
}
\details{
The model timesteps will always be numbered from 1 to \code{n_timesteps()} and so
likely will not be consistent with weeks of the year with the truncated
model.

Currently it's possible to truncate both a fitted model and a model produced
by \code{\link[=preprocess_species]{preprocess_species()}} but I have yet to create a way to export a
preprocessed model after truncating it (independently of
\code{\link[=preprocess_species]{preprocess_species()}}) or to truncate during preprocessing. So currently
the utility is limited to reducing the size of fitted models.

It's possible to truncate a model over the year boundary but routes generated
from such a model will not plot correctly so it's not recommended;
see \href{https://github.com/birdflow-science/BirdFlowR/issues/120}{issue #120}.
}
