% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_BirdFlow.R
\name{validate_BirdFlow}
\alias{validate_BirdFlow}
\title{Function to validate a BirdFlow object}
\usage{
validate_BirdFlow(x, error = TRUE, allow_incomplete = FALSE)
}
\arguments{
\item{x}{a BirdFlow object}

\item{error}{if TRUE throw an error if there are problems if FALSE return
any problems as a data.frame.}

\item{allow_incomplete}{if TRUE only throw an error if the problem type is
not \code{"incomplete"}; this allows checking output of
\code{\link[=preprocess_species]{preprocess_species()}}.}
}
\value{
If \code{error = FALSE} the function returns
a data.frame describing any errors with columns:
\describe{
\item{problem}{a character description of any problems}
\item{type}{the problem type, either "error" or "incomplete"}
}
Otherwise, if there are no problems a similar data.frame with no rows is
returned invisibly.
}
\description{
Throw error if a BirdFlow object is malformed or incomplete.
}
\details{
\code{\link[=preprocess_species]{preprocess_species()}} creates a BirdFlow object that lacks both marginals
and transitions and thus can't be used to make projections.
\code{validate_BirdFlow()} tags the absence of these with the type "incomplete".
Any other missing or malformed components are tagged "error".

Since marginals can be used to calculate both distributions and
transition matrices, a BirdFlow object can be complete if it has marginals;
or has both transitions and distributions. Having redundancy in these three
is not considered an error.
}
